% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAfforestationMask.R
\name{calcAfforestationMask}
\alias{calcAfforestationMask}
\title{calcAfforestationMask}
\usage{
calcAfforestationMask(subtype)
}
\arguments{
\item{subtype}{afforestation mask sub type}
}
\value{
magpie object in cellular resolution
}
\description{
Afforestation mask for where afforestation possible
}
\examples{
\dontrun{ calcOutput("AfforestationMask", subtype="noboreal", aggregate = FALSE) }

}
\author{
David Chen
}
