% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaActuallyIrrigated.R
\name{calcAreaActuallyIrrigated}
\alias{calcAreaActuallyIrrigated}
\title{calcAreaActuallyIrrigated}
\usage{
calcAreaActuallyIrrigated(aggregationlevel = "iso", selectyears = "y1995")
}
\arguments{
\item{aggregationlevel}{default is iso}

\item{selectyears}{select years}
}
\value{
magpie object with results on cellular or iso country level
}
\description{
retrieves irrigated crop area from croparea intialization
}
\examples{

\dontrun{
calcOutput("AreaActuallyIrrigated")
}


}
\seealso{
\code{\link{calcLanduseInitialisation}}
}
\author{
Felicitas Beier
}
