% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaPotIrrig.R
\name{calcAreaPotIrrig}
\alias{calcAreaPotIrrig}
\title{calcAreaPotIrrig}
\usage{
calcAreaPotIrrig(
  selectyears,
  comAg,
  cropAggregation,
  cropmix,
  iniyear,
  landScen
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{comAg}{If TRUE: committed irrigated areas are subtracted,
if FALSE: total potential croparea is used}

\item{cropAggregation}{TRUE (aggregate over crop types), FALSE (returns area per crop type)}

\item{cropmix}{Cropmix for which potential areas are calculated
(options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_rainf" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{iniyear}{Initialization year for current cropland area}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, or one of the scenarios available in calcConservationPriorities,
   e.g., 30by30, BH, BH_IFL, PBL_HalfEarth,
   or NA for no protection).
For case of no land protection select "NA" in second part of argument
or do not specify second part of the argument.}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates land that is potentially available
             for irrigated agriculture
}
\examples{
\dontrun{
calcOutput("AreaPotIrrig", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
