% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAtmosphericDepositionRates.R
\name{calcAtmosphericDepositionRates}
\alias{calcAtmosphericDepositionRates}
\title{calcAtmosphericDepositionRates}
\usage{
calcAtmosphericDepositionRates()
}
\value{
magpie object in cellular resolution
}
\description{
Function that calculates Atmospheric deposition rates
}
\examples{
\dontrun{ calcOutput("AtmosphericDepositionRates", aggregate = FALSE) }

}
\author{
Marcos Alves
}
