% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlIrrig.R
\name{calcAvlIrrig}
\alias{calcAvlIrrig}
\title{calcAvl_irrig}
\usage{
calcAvlIrrig()
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts Area equipped for irrigation for the period 1997-2002,
data revision: 18, citation: Siebert et al.; 2007 ; Global Map of Irrigation Areas version 4.0.1
}
\examples{
\dontrun{ calcOutput("AvlIrrig", aggregate = FALSE) }

}
\author{
David Chen
}
