% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlLandSi.R
\name{calcAvlLandSi}
\alias{calcAvlLandSi}
\title{calcAvlLandSi}
\usage{
calcAvlLandSi()
}
\value{
magpie object in cellular resolution
}
\description{
Extracts si0 and nsi0 areas based on Ramankutty dataset
}
\examples{
\dontrun{ calcOutput("AvlLandSi", aggregate = FALSE) }

}
\author{
David Chen
}
