% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBasinViolations.R
\name{calcBasinViolations}
\alias{calcBasinViolations}
\title{calcBasinViolations}
\usage{
calcBasinViolations(
  lpjml,
  selectyears,
  climatetype,
  EFRmethod,
  accessibilityrule,
  rankmethod,
  yieldcalib,
  allocationrule,
  thresholdtype,
  gainthreshold,
  irrigationsystem,
  iniyear,
  avlland_scen,
  cropmix,
  com_ag,
  multicropping,
  scenario
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{selectyears}{Years for which irrigatable area is calculated}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year (Qx).
(e.g. Q75: 0.25, Q50: 0.5)}

\item{rankmethod}{Method of calculating the rank:
"meancellrank": mean over cellrank of proxy crops,
"meancroprank": rank over mean of proxy crops (normalized),
"meanpricedcroprank": rank over mean of proxy crops (normalized using price),
"watervalue": rank over value of irrigation water;
and fullpotentail TRUE/FALSE separated by ":"
(TRUE: Full irrigation potential (cell receives full irrigation requirements in total area).
FALSE: reduced potential of cell receives at later stage in allocation algorithm)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst", "equality")}

\item{thresholdtype}{Thresholdtype of yield improvement:
TRUE: monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{gainthreshold}{Threshold of yield improvement potential required
(same unit as thresholdtype)}

\item{irrigationsystem}{Irrigation system used
("surface", "sprinkler", "drip", "initialization")}

\item{iniyear}{Initialization year of irrigation system}

\item{avlland_scen}{Land availability scenario (currCropland, currIrrig, potIrrig)
combination of land availability scenario and initialization year separated by ":".
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{com_ag}{If TRUE: currently already irrigated areas in
         initialization year are reserved for irrigation,
if FALSE: no irrigation areas reserved (irrigation potential)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{scenario}{Combination of EFP ("on", "off") and
non-ag. water use scenario ("ssp2", "ISIMIP", ...)
separated by "."}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates basins that violate environmental flow
             requirements
}
\examples{
\dontrun{
calcOutput("BasinViolations", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Benjamin Leon Bodirsky
}
