% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBlueWaterConsumptionMain.R
\name{calcBlueWaterConsumptionMain}
\alias{calcBlueWaterConsumptionMain}
\title{calcBlueWaterConsumptionMain}
\usage{
calcBlueWaterConsumptionMain(selectyears, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Climate model (e.g., "MRI-ESM2-0:ssp370")
or historical baseline (e.g., "GSWP3-W5E5:historical")}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates consumptive blue water consumption
             for the main growing period based on
             LPJmL transpiration, evaporation and interception for different
             irrigation systems (sprinkler, surface, drip)
}
\examples{
\dontrun{
calcOutput("BlueWaterConsumptionMain", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
