% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBphTCRE.R
\name{calcBphTCRE}
\alias{calcBphTCRE}
\title{calcBphTCRE}
\usage{
calcBphTCRE()
}
\value{
magpie object in cellular resolution
}
\description{
Transient Climate Response to accumulated doubling of CO2. File based on CMIP5 +1perc CO2 per year experiment. To be used in the translation to carbon equivalents of BphEffect
# @param
}
\examples{
\dontrun{ calcOutput("BphTCRE", aggregate = FALSE) }

}
\author{
Michael Windisch
}
