% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCalibratedArea.R
\name{calcCalibratedArea}
\alias{calcCalibratedArea}
\title{calcCalibratedArea}
\usage{
calcCalibratedArea()
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts calibrated area
}
\examples{
\dontrun{ calcOutput("CalibratedArea", aggregate = FALSE) }

}
\author{
David Chen
}
