% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClimateClass.R
\name{calcClimateClass}
\alias{calcClimateClass}
\title{calcClimateClass}
\usage{
calcClimateClass()
}
\value{
Clustered MAgPIE object on requested resolution
}
\description{
fraction of a cell belonging to a given climate classification based on Koeppen Geiger Classification. http://koeppen-geiger.vu-wien.ac.at/.
}
\examples{
\dontrun{ calcOutput("ClimateClass", aggregate = FALSE) }

}
\author{
Abhijeet Mishra
}
