% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterBase.R
\name{calcClusterBase}
\alias{calcClusterBase}
\title{calcClusterBase}
\usage{
calcClusterBase(clusterdata = "yield_airrig")
}
\arguments{
\item{clusterdata}{similarity data to be used to determine clusters: yield_airrig (current default) or yield_increment}
}
\value{
A matrix containing the data
}
\description{
Reads a series of MAgPIE files and combines them to a matrix
which is then used for calculating a clustering.
}
\seealso{
\code{\link{calcCluster}}
}
\author{
Jan Philipp Dietrich, Felicitas Beier
}
