% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterHierarchical.R
\name{calcClusterHierarchical}
\alias{calcClusterHierarchical}
\title{MAgPIE Hierarchical clustering}
\usage{
calcClusterHierarchical(regionscode, ncluster, mode = "h", weight = NULL)
}
\arguments{
\item{regionscode}{regionscode of the regional mapping to be used. Must agree
with the regionscode of the mapping mentioned in the madrat config! Can be
retrieved via \code{regionscode()}.}

\item{ncluster}{The desired total number of clusters.}

\item{mode}{Clustering type. At the moment you can choose between complete
linkage clustering (h), single linkage clustering (s) and Ward clustering
(w).}

\item{weight}{named vector with weighting factors for each region for the cluster distribution
,e.g. weight=c(AFR=3,EUR=0.5). weight > 1 will grant more cluster to a region and
weight < 1 less cluster than by default.}
}
\value{
A mapping between regions and clusters
}
\description{
Performs MAgPIE hierarchical clustering and calculates corresponding spam
relation matrix
}
\details{
As the creation of a clustering tree is very time consuming the function
checks first in the input folder if the corresponding data already exists
and if not it stores the tree information in the input folder for later use
in the next execution of this function.
}
\seealso{
\code{\link{calcCluster}}, \code{\link{calcClusterKMeans}}
}
\author{
Jan Philipp Dietrich
}
