% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClusterTreeHierarchical.R
\name{calcClusterTreeHierarchical}
\alias{calcClusterTreeHierarchical}
\title{MAgPIE Hierarchical clustering tree}
\usage{
calcClusterTreeHierarchical(regionscode, mode = "h", weight = NULL)
}
\arguments{
\item{regionscode}{regionscode of the regional mapping to be used. Must agree with the regionscode of the mapping
mentioned in the madrat config! Can be retrieved via \code{regionscode()}.}

\item{mode}{Clustering type. At the moment you can choose between complete
linkage clustering (h), single linkage clustering (s) and Ward clustering
(w).}

\item{weight}{named vector with weighting factors for each region for the cluster distribution
,e.g. weight=c(AFR=3,EUR=0.5). weight > 1 will grant more cluster to a region and
weight < 1 less cluster than by default.}
}
\value{
A spam relation matrix
}
\description{
calculates hierarchical clustering tree
}
\seealso{
\code{\link{cluster_per_region}}, \code{\link{mag_kmeans}},
\code{\link{clusterspam}}
}
\author{
Jan Philipp Dietrich
}
