% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCO2Atmosphere.R
\name{calcCO2Atmosphere}
\alias{calcCO2Atmosphere}
\title{calcRrLayer}
\usage{
calcCO2Atmosphere(rcp = "rcp8p5", level = "cellular")
}
\arguments{
\item{rcp}{specify the RCP}

\item{level}{specify the spatial output level}
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts range-rarity as used for biodiversity loss
}
\examples{
\dontrun{ calcOutput("RrLayer", aggregate = FALSE) }

}
\author{
Michael Windisch
}
