% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropareaAdjusted.R
\name{calcCropareaAdjusted}
\alias{calcCropareaAdjusted}
\title{calcCropareaAdjusted}
\usage{
calcCropareaAdjusted(sectoral, physical, cellular, cells, irrigation)
}
\arguments{
\item{sectoral}{"area_harvested" returns croparea aggregated to FAO products,
"ProductionItem" unaggregated ProdSTAT items,
"FoodBalanceItem" Food Balance Sheet categories,
"kcr" MAgPIE items, and "lpj" LPJmL items}

\item{physical}{if TRUE the sum over all crops agrees with the cropland area per country}

\item{cellular}{if TRUE: calculates cellular MAgPIE crop area for all magpie croptypes.
Crop area from LUH2 crop types (c3ann, c4ann, c3per, c4per, cnfx)
are mapped to MAgpIE crop types using mappingLUH2cropsToMAgPIEcrops.csv.
Harvested areas of FAO weight area within a specific LUH crop type
to divide into MAgPIE crop types.}

\item{cells}{Switch between "magpiecell" (59199) and "lpjcell" (67420)}

\item{irrigation}{If true: cellular areas are returned separated
into irrigated and rainfed (see setup in calcLUH2v2)}
}
\value{
magpie object in cellular resolution
}
\description{
This function adjusts the cell names that mrcommons::calcCroparea
             returns, so that they have the correct format for the mrwater library
             Note: This function is only necessary until the 67k cell names
             are fully integrated in mrcommons (specifically calcCroparea and calcLUH2v2)
}
\examples{
\dontrun{
calcOutput("CropareaAdjusted", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
