% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropAreaInitialization.R
\name{calcCropAreaInitialization}
\alias{calcCropAreaInitialization}
\title{calcCropAreaInitialization}
\usage{
calcCropAreaInitialization(aggregate = FALSE, irrigation = TRUE)
}
\arguments{
\item{aggregate}{This parameter determines if the crops should be differentiated between irrigation methods}

\item{irrigation}{This parameter determines if the crops should be differentiated between irrigation methods}
}
\value{
magpie object in cellular resolution
}
\description{
Function to generate the initialization values for crop-specific physical area with half-degree resolution
}
\examples{
\dontrun{ calcOutput("calcCropAreaInitialization", aggregate = FALSE) }
}
\author{
Edna Molina Bacca
}
