% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropAreaShare.R
\name{calcCropAreaShare}
\alias{calcCropAreaShare}
\title{calcCropAreaShare}
\usage{
calcCropAreaShare(iniyear, cropmix)
}
\arguments{
\item{iniyear}{Croparea initialization year}

\item{cropmix}{Cropmix for which croparea share is calculated
(options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the crop area share for a chosen cropmix
}
\examples{
\dontrun{
calcOutput("CropAreaShare", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Benjamin L. Bodirsky
}
