% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCshareReleased.R
\name{calcCshareReleased}
\alias{calcCshareReleased}
\title{calcCshareReleased}
\usage{
calcCshareReleased()
}
\value{
magpie object in cellular resolution
}
\description{
"Share of soil carbon that is released on cropland during the harvest. Calculated by Benjamin Bodirsky"
}
\examples{
\dontrun{calcOutput("CshareReleased", aggregate = FALSE) }

}
\author{
David Chen
}
