% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDischargeInaccessibleAdd.R
\name{calcDischargeInaccessibleAdd}
\alias{calcDischargeInaccessibleAdd}
\title{calcDischargeInaccessibleAdd}
\usage{
calcDischargeInaccessibleAdd(
  selectyears = selectyears,
  lpjml = lpjml,
  climatetype = climatetype,
  accessibilityrule = accessibilityrule,
  efrMethod = efrMethod
)
}
\arguments{
\item{selectyears}{Years to be returned
(Note: does not affect years of harmonization or smoothing)}

\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios
or historical baseline "GSWP3-W5E5:historical"}

\item{accessibilityrule}{Method used: Quantile method (Q) or Coefficient of Variation (CV)
combined with scalar value defining the strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year
(Qx, e.g. Q75: 0.25, Q50: 0.5)
or base value for exponential curve separated by : (CV:2)}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(Smakhtin:good, VMF:fair)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water that is inaccessible to humans
             but not part of EFRs.
             Reason: Inacessible discharge is highly variable discharge that can also
             serve as high flow requirements (HFR) for EFRs.
             This has to be accounted in the Discharge Allocation Algorithm
             for the determination of potential irrigation.
}
\examples{
\dontrun{
calcOutput("DischargeInaccessibleAdd", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
