% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmulatorPastCorrectionFactor3.R
\name{calcEmuPastCorrectFactor3}
\alias{calcEmuPastCorrectFactor3}
\title{calcEmuPastCorrectFactor}
\usage{
calcEmuPastCorrectFactor3(
  model = "f41f19be67",
  GCMModel = "HadGEM2",
  rcp = "rcp85"
)
}
\arguments{
\item{model}{grass harvest machine learning model ID}

\item{GCMModel}{Global climate model name}

\item{rcp}{RCP}
}
\value{
MAgPIE objects with pasture correction factor on a cellular level.
}
\description{
Loads the ANN trained grass emulator with historical data and calculates the pasture correction factor
}
\examples{

\dontrun{
calcOutput("EmuPastCorrectFactor", model = "f41f19be67", GCMModel = "HadGEM2", rcp = "rcp85")
}

}
\author{
Marcos Alves
}
