% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnvmtlFlowRequirements.R
\name{calcEnvmtlFlowRequirements}
\alias{calcEnvmtlFlowRequirements}
\title{calcEnvmtlFlowRequirements}
\usage{
calcEnvmtlFlowRequirements(lpjml, selectyears, climatetype, EFRmethod)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{selectyears}{Years to be returned
(Note: does not affect years of harmonization or smoothing)}

\item{climatetype}{Switch between different climate scenarios
or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs
(Smakhtin:good, VMF:fair)}
}
\value{
magpie object with EFRs, LFRs and HFRs in cellular resolution
}
\description{
This function calculates environmental flow requirements (EFR)
             based on EFR share calculated from LPJmL monthly discharge
}
\examples{
\dontrun{
calcOutput("EnvmtlFlowRequirements", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
