% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnvmtlFlowRequirementsShare.R
\name{calcEnvmtlFlowRequirementsShare}
\alias{calcEnvmtlFlowRequirementsShare}
\title{calcEnvmtlFlowRequirementsShare}
\usage{
calcEnvmtlFlowRequirementsShare(lpjml, climatetype, efrMethod)
}
\arguments{
\item{lpjml}{LPJmL version used for monthly discharge}

\item{climatetype}{Switch between different climate scenarios
or historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}
}
\value{
magpie object in cellular resolution representing share of discharge
        that is reserved for environmental flows
}
\description{
This function calculates environmental flow requirements (EFR)
             (as share of discharge) based on LPJmL monthly discharge
}
\examples{
\dontrun{
calcOutput("EnvmtlFlowRequirementsShare", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
