% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEvapotranspiration.R
\name{calcEvapotranspiration}
\alias{calcEvapotranspiration}
\title{calcEvapotranspiration}
\usage{
calcEvapotranspiration(selectyears, runtype, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{runtype}{Select crop and management type for which ET should be calculated
options: `cft:ir` (crop irrigated in irrigated growing period of respective crop),
         `cft:noir` (crop not irrigated in irrigated growing period of respective crop),
         `grass:ir` (grass irrigated in irrigated growing period of crop),
         `grass:noir` (grass not irrigated in irrigated growing period of crop)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates evapotranspiration (ET) from LPJmL inputs
             (transpiration, evaporation, interception)
}
\examples{
\dontrun{
calcOutput("Evapotranspiration", aggregate = FALSE)
}
}
\author{
Felicitas Beier, Jens Heinke
}
