% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFullIrrigationRequirement.R
\name{calcFullIrrigationRequirement}
\alias{calcFullIrrigationRequirement}
\title{calcFullIrrigationRequirement}
\usage{
calcFullIrrigationRequirement(
  lpjml,
  climatetype,
  selectyears,
  iniyear,
  comAg,
  irrigationsystem,
  landScen,
  cropmix,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned}

\item{iniyear}{Croparea initialization year}

\item{comAg}{If TRUE: committed irrigated areas are subtracted,
if FALSE: total potential croparea is used}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, or one of the scenarios available in
   calcConservationPriorities, e.g., 30by20, BH, BH_IFL, PBL_HalfEarth,
   or NA for no protection).
For case of no land protection select "NA" in second part of argument
or do not specify second part of the argument}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multiple cropped areas calculated from total harvested areas
                and total physical areas per cell from readLandInG
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multiple cropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multiple cropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the water requirements for full
             irrigation per cell per crop given potentially available land
}
\examples{
\dontrun{
calcOutput("FullIrrigationRequirement", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
