% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigAreaActuallyCommitted.R
\name{calcIrrigAreaActuallyCommitted}
\alias{calcIrrigAreaActuallyCommitted}
\title{calcIrrigAreaActuallyCommitted}
\usage{
calcIrrigAreaActuallyCommitted(
  lpjml,
  climatetype,
  selectyears,
  iniyear,
  efrMethod,
  multicropping,
  transDist
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios
or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned
(Note: does not affect years of harmonization or smoothing)}

\item{iniyear}{Initialization year of irrigation system}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(Smakhtin:good, VMF:fair)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; TRUE:exogenous; FALSE)}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area reserved for irrigation based on area irrigated
             in initialization and available water resources
}
\examples{
\dontrun{
calcOutput("IrrigAreaActuallyCommitted", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
