% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigAreaCommitted.R
\name{calcIrrigAreaCommitted}
\alias{calcIrrigAreaCommitted}
\title{calcIrrigAreaCommitted}
\usage{
calcIrrigAreaCommitted(selectyears, iniyear, aggregateCrops = FALSE)
}
\arguments{
\item{selectyears}{select years}

\item{iniyear}{initialization year}

\item{aggregateCrops}{Logical. If FALSE (default), result is returned per crop.
If TRUE, values are aggregated over the crop dimension.}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area reserved for irrigation based on area irrigated
             in initialization year and depreciation parameter (set to 0.1)
}
\examples{
\dontrun{
calcOutput("IrrigAreaCommitted", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
