% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigatableArea.R
\name{calcIrrigatableArea}
\alias{calcIrrigatableArea}
\title{calcIrrigatableArea}
\usage{
calcIrrigatableArea(
  selectyears,
  climatetype,
  allocationrule,
  thresholdtype,
  gainthreshold,
  irrigationsystem,
  avlland_scen,
  proxycrop
)
}
\arguments{
\item{selectyears}{years for which irrigatable area is calculated}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm: TRUE (default): monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{gainthreshold}{Threshold of yield improvement potential required for water allocation in upstreamfirst algorithm (in tons per ha)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{avlland_scen}{Land availability scenario: current or potential; optional additionally: protection scenario in case of potential (when left empty: no protection) and initialization year of cropland area
combination of land availability scenario and initialization year separated by ":". land availability scenario: currIrrig (only currently irrigated cropland available for irrigated agriculture), currCropland (only current cropland areas available for irrigated agriculture), potIrrig (suitable land is available for irrigated agriculture, potentially land restrictions activated through protect_scen argument)
protection scenario separated by "_" (only relevant when potIrrig selected): WDPA, BH, FF, CPD, LW, HalfEarth. Areas where no irrigation water withdrawals are allowed due to biodiversity protection.}

\item{proxycrop}{proxycrop(s) selected for crop mix specific calculations: average over proxycrop(s) yield gain. NULL returns all crops individually}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area that can potentially be irrigated given available water and land
}
\examples{
\dontrun{ calcOutput("IrrigatableArea", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
