% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigation.R
\name{calcIrrigation}
\alias{calcIrrigation}
\title{calcIrrigation}
\usage{
calcIrrigation(years = seq(1995, 2095, by = 5))
}
\arguments{
\item{years}{years to be returned}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts irrigation water (airrig: water applied additionally to rainfall) from LPJmL to MAgPIE
}
\examples{
\dontrun{ calcOutput("Irrigation", aggregate = FALSE) }

}
\author{
Felicitas Beier, Abhijeet Mishra
}
