% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigationDataMAgPIE.R
\name{calcIrrigationDataMAgPIE}
\alias{calcIrrigationDataMAgPIE}
\title{calcIrrigationDataMAgPIE}
\usage{
calcIrrigationDataMAgPIE(output, lpjml, climatetype, selectyears, iniyear)
}
\arguments{
\item{output}{"PotIrrigAreas": potentially irrigated areas
"WaterAvlMAgPIE": potential irrigation water withdrawals}

\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years for which irrigatable area is calculated}

\item{iniyear}{Initialization year for initial croparea}
}
\value{
magpie object in cellular resolution
}
\description{
returns irrigation data required for MAgPIE
             for different scenarios
             given available water and land:
             Potentially Irrigated Areas (PIA) and
             Potential Irrigation Water Withdrawals (PIWW)
}
\examples{
\dontrun{
calcOutput("IrrigationDataMAgPIE", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
