% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigationPotentials.R
\name{calcIrrigationPotentials}
\alias{calcIrrigationPotentials}
\title{calcIrrigationPotentials}
\usage{
calcIrrigationPotentials(
  output,
  transDist,
  selectyears,
  iniyear,
  lpjml,
  climatetype
)
}
\arguments{
\item{output}{output to be returned:
"IrrigArea": irrigated area (in Mha/yr) or
"wat_ag_ww": available irrigation water for withdrawals (in km^3/yr) or
"wat_ag_wc": available consumptive irrigation water (in km^3/yr)}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}

\item{selectyears}{years for which potentially irrigated area is calculated}

\item{iniyear}{initialization year}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object
}
\description{
calculates irrigation potentials in terms of potentially irrigated
             areas and potential irrigation water for MAgPIE
}
\examples{
\dontrun{
calcOutput("IrrigationPotentials", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
