% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigCellranking.R
\name{calcIrrigCellranking}
\alias{calcIrrigCellranking}
\title{calcIrrigCellranking}
\usage{
calcIrrigCellranking(
  lpjml,
  climatetype,
  cellrankyear,
  iniyear,
  comAg,
  irrigationsystem,
  landScen,
  method,
  cropmix,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical" for yields}

\item{cellrankyear}{Year(s) for which cell rank is calculated}

\item{iniyear}{Initialization year for price}

\item{comAg}{If TRUE: committed irrigated areas are subtracted,
if FALSE: total potential croparea is used}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, or one of the scenarios available in calcConservationPriorities,
   e.g., 30by20, BH, BH_IFL, PBL_HalfEarth,
   or NA for no protection).
For case of no land protection select "NA" in second part of argument
or do not specify second part of the argument}

\item{method}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for relative area return, or
USD_m3 (USD per cubic meter) for relative volumetric return;
USD for absolute return (total profit);
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices;
and boolean indicating fullpotential (TRUE) or reduced potential (FALSE)}

\item{cropmix}{Selected cropmix for which yield improvement potential
is calculated (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)
NULL returns all crops individually}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
Also needs specification of refYields, separated by ":".
Options: FALSE (for single cropping analyses) or
         "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates a cellranking
             for the river basin discharge allocation
             based on yield improvement potential through irrigation
}
\examples{
\dontrun{
calcOutput("IrrigCellranking", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
