% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigCropYieldGain.R
\name{calcIrrigCropYieldGain}
\alias{calcIrrigCropYieldGain}
\title{calcIrrigCropYieldGain}
\usage{
calcIrrigCropYieldGain(
  lpjml,
  climatetype,
  priceAgg,
  iniyear,
  selectyears,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{priceAgg}{Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices}

\item{iniyear}{initialization year for food price and cropmix area}

\item{selectyears}{Years to be returned by the function}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
         Also needs specification of refYields, separated by ":".
         Options: FALSE (for single cropping analyses) or
                  "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the yield gains per crop through irrigation
             (including negatives)
}
\examples{
\dontrun{
calcOutput("IrrigCropYieldGain", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
