% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigSystemShr.R
\name{calcIrrigSystemShr}
\alias{calcIrrigSystemShr}
\title{calcIrrigSystemShr}
\usage{
calcIrrigSystemShr(iniyear)
}
\arguments{
\item{iniyear}{Year to be used for irrigation system share initialization}
}
\value{
magpie object in cellular resolution
}
\description{
This function returns the share of the
             irrigation area under a specific irrigation system type
             (surface, sprinkler, drip)
             per crop type
}
\examples{
\dontrun{
calcOutput("IrrigSystemShr", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
