% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigWatValue.R
\name{calcIrrigWatValue}
\alias{calcIrrigWatValue}
\title{calcIrrigWatValue}
\usage{
calcIrrigWatValue(
  lpjml,
  selectyears,
  climatetype,
  iniyear,
  cropmix,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version to be used}

\item{selectyears}{Years to be returned}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{iniyear}{Initialization year for price in price-weighted normalization of meanpricedcroprank and yield improvement potential prices}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{yieldcalib}{FAO (LPJmL yields calibrated with current FAO yield) or
calibrated (LPJmL yield potentials harmonized to baseline and calibrated for proxycrops) or
smoothed (smoothed LPJmL yield potentials, not harmonized, not calibrated) or
smoothed_calib}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the value of irrigation water
             (added value of water to the production process)
             based on yield gain, crop prices and irrigation water requirements
}
\examples{
\dontrun{
calcOutput("IrrigWatValue", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
