% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISIMIPYields.R
\name{calcISIMIPYields}
\alias{calcISIMIPYields}
\title{calcISIMIPYields}
\usage{
calcISIMIPYields(
  subtype = "ISIMIP3b:yields.EPIC-IIASA_ukesm1-0-ll_ssp585_default",
  time = "spline",
  dof = 4
)
}
\arguments{
\item{subtype}{subtype of yield based on readISIMIPoutputs, for crop yields}

\item{time}{spline or average, if spline specify dof, if average specify averaging range}

\item{dof}{degrees of freedom for spline
@return magpie object in cellular resolution}
}
\description{
reads and cleans up ISIMIP3b crop yield data
}
\examples{
\dontrun{ calcOutput("ISIMIPYields", aggregate = FALSE) }

}
\author{
David Meng-Chuen Chen
}
