% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKoeppen_geiger.R
\name{calcKoeppen_geiger}
\alias{calcKoeppen_geiger}
\title{calcKoeppen_geiger}
\usage{
calcKoeppen_geiger()
}
\value{
magpie object in cellular resolution
}
\description{
fraction of a cell belonging to a given climate classification,reference: Rubel, F., and M. Kottek, 2010: Observed and projected climate shifts 1901-2100 depicted by world maps of the Koeppen-Geiger climate classification. Meteorol. Z., 19, 135-141. DOI: 10.1127/0941-2948/2010/0430."
note: Missing values are replaced with the classification information which appears most often in the cells closest to the one with the missing information.
}
\examples{
\dontrun{ calcOutput("Koeppen_geiger", aggregate = FALSE) }

}
\author{
David Chen
}
