% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLakeFlows.R
\name{calcLakeFlows}
\alias{calcLakeFlows}
\title{calcLakeFlows}
\usage{
calcLakeFlows(lpjml, climatetype, subtype)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios
or historical baseline "GSWP3-W5E5:historical"}

\item{subtype}{Switch between water inputs to water bodies ("input_lake")
and evaporation from water bodies ("evap_lake")}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates evaporation from water bodies
             or inputs to water bodies
             based on LPJmL inputs
}
\examples{
\dontrun{
calcOutput("LakeFlows", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
