% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandAvl.R
\name{calcLandAvl}
\alias{calcLandAvl}
\title{calcLandAvl}
\usage{
calcLandAvl(
  selectyears = 1995,
  cells = "lpjcell",
  output = "irrigatable_area",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = "CRU_4",
  ref_year = "y2015",
  allocationrule = "optimization",
  allocationshare = NULL,
  gainthreshold = 1,
  irrigationsystem = "initialization",
  irrigini = "Jaegermeyr_lpjcell",
  iniyear = 1995,
  landtype = "potentialcropland",
  protectscen = "WDPA",
  proxycrops = "maiz"
)
}
\arguments{
\item{selectyears}{years to be returned by this function}

\item{cells}{cells to be returned by the function (lpjcell or magpiecell)}

\item{output}{Type of area to be returned: irrigatable_area (default, considers both water and land availability), irrig_area_ww or irrig_area_wc (area that can be irrigated given water available for withdrawal or consumption), avl_land, protect_area (restricts water withdrawal in protected areas)}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{allocationshare}{Share of water to be allocated to cell (only needs to be selected in case of allocationrule=="equality")}

\item{gainthreshold}{Threshold of yield improvement potential required for water allocation in upstreamfirst algorithm (in tons per ha)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{irrigini}{When "initialization" selected for irrigation system: choose initialization data set for irrigation system initialization ("Jaegermeyr_lpjcell", "LPJmL_lpjcell")}

\item{iniyear}{initialization year used for irrigated area}

\item{landtype}{current cropland area (currentcropland) or potential cropland area (potentialcropland) or nonprotected (restriction of withdrawals in protected areas)}

\item{protectscen}{protection scenario for protected areas ("WDPA", "HalfEarth", )}

\item{proxycrops}{Proxycrops for water requirements}

\item{depreciation}{parameter defining yearly depreciation rate at which previously irrigated cropland becomes "unreserved" for irrigation}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area that is available for (irrigated) cropland under different assumptions
}
\examples{
\dontrun{ calcOutput("LandAvl", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
