% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivestockDistribution.R
\name{calcLivestockDistribution}
\alias{calcLivestockDistribution}
\title{calcLivestockDistribution}
\usage{
calcLivestockDistribution()
}
\value{
MAgPIE objects with livestock numbers on a cellular level.
}
\description{
Disaggregate Livestock estimates based on the GLW3 dataset.
}
\examples{

\dontrun{
calcOutput("LivestockDistribution")
}

}
\author{
Marcos Alves
}
