% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLuh2SideLayers.R
\name{calcLuh2SideLayers}
\alias{calcLuh2SideLayers}
\title{calcLuh2SideLayers}
\usage{
calcLuh2SideLayers()
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts biodiversity data for LUH2 land cover types
}
\examples{
\dontrun{ calcOutput("Luh2SideLayers", aggregate = FALSE) }

}
\author{
Michael Windisch
}
