% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMissingWater.R
\name{calcMissingWater}
\alias{calcMissingWater}
\title{calcMissingWater}
\usage{
calcMissingWater(
  output,
  lpjml,
  climatetype,
  transDistGW = 100,
  multicropping = "TRUE:actual:irrig_crop",
  selectyears,
  iniyear
)
}
\arguments{
\item{output}{sector to be reported:
non-agricultural sector water use ("nonAg"),
(committed) agricultural water use ("comAg"),
or both ("total")}

\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{transDistGW}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability
to determine missing water that is tagged as non-renewable groundwater}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from LandInG
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year}
}
\value{
cellular magpie object
}
\description{
Calculates difference between requested water
             and water that can be fulfilled in algorithm
             given available surface water resources
             for all scenarios.
             This information is used to derive
             non-renewable groundwater use.
}
\examples{
\dontrun{
calcOutput("MissingWater", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
