% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonAgWaterDemand.R
\name{calcNonAgWaterDemand}
\alias{calcNonAgWaterDemand}
\title{calcNonAgWaterDemand}
\usage{
calcNonAgWaterDemand(
  selectyears = "all",
  source = "WATCH_ISIMIP_WATERGAP",
  waterusetype = "withdrawal",
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  seasonality = "grper",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  harmonize_baseline = "CRU_4",
  ref_year = "y2015"
)
}
\arguments{
\item{selectyears}{years to be returned}

\item{source}{data source to be used (e.g. WATERGAP2020)}

\item{waterusetype}{withdrawal (default) or consumption}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{seasonality}{grper (default): non-agricultural water demand in growing period per year; total: non-agricultural water demand throughout the year}

\item{climatetype}{switch between different climate scenarios (default: "CRU_4") for calcGrowingPeriod}

\item{harmonize_baseline}{FALSE (default), if a baseline is specified here input data is harmonized to that baseline (from ref_year onwards) Note: only applies for calcGrowingPeriod}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year for calcGrowingPeriod}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts non-agricultural water demand
}
\examples{
\dontrun{ calcOutput("NonAgWaterDemand", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
