% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonrenGroundwatUse.R
\name{calcNonrenGroundwatUse}
\alias{calcNonrenGroundwatUse}
\title{calcNonrenGroundwatUse}
\usage{
calcNonrenGroundwatUse(
  lpjml,
  climatetype,
  transDistGW = 100,
  multicropping = "TRUE:actual:irrig_crop",
  selectyears,
  iniyear,
  efrMethod
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{transDistGW}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability
to determine missing water that is tagged as non-renewable groundwater}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}
}
\value{
cellular magpie object
}
\description{
Calculates current non-renewable (fossil) groundwater use
             as negative difference between
             water availability from renewable sources and
             human water use (consumption and withdrawal).
             The water use in the initialization year is fixed in the future as
             non-renewable groundwater.
}
\examples{
\dontrun{
calcOutput("NonrenGroundwatUse", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
