% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNpiNdcAdAolcPol.R
\name{calcNpiNdcAdAolcPol}
\alias{calcNpiNdcAdAolcPol}
\title{calcNpiNdcAdAolcPol}
\usage{
calcNpiNdcAdAolcPol()
}
\value{
magpie object in cellular resolution
}
\description{
Function creates dummy NPI/NDC policies
}
\examples{
\dontrun{ calcOutput("NpiNdcAdAolcPol", aggregate = FALSE) }
}
\author{
Patrick v. Jeetze, Michael Windisch
}
