% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNpiNdcAffPol.R
\name{calcNpiNdcAffPol}
\alias{calcNpiNdcAffPol}
\title{calcNpiNdcAffPol}
\usage{
calcNpiNdcAffPol()
}
\value{
magpie object in cellular resolution
}
\description{
Function creates dummy NPI/NDC policies
}
\examples{
\dontrun{ calcOutput("NpiNdcAffPol", aggregate = FALSE) }
}
\author{
Patrick v. Jeetze, Michael Windisch
}
