% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPeatland.R
\name{calcPeatland}
\alias{calcPeatland}
\title{calcPeatland}
\usage{
calcPeatland(subtype = "degraded")
}
\arguments{
\item{subtype}{degraded (default) or intact}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates degraded and intact peatland area at cell level.
The function takes degraded and intact peatland area from the Global Peatland Database (GPD) at the national level and downscales the peatland area to grid cell level using gridded potential peatland area.
The GPD has been provided by Alexandra Barthelmes. The potential peatland area has been provided by Leifeld_2018 (DOI 10.1038/s41467-018-03406-6).
}
\examples{
\dontrun{ calcOutput("Peatland", aggregate = FALSE) }

}
\author{
Florian Humpenoeder
}
