% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPhotosynthesisTemperature.R
\name{calcPhotosynthesisTemperature}
\alias{calcPhotosynthesisTemperature}
\title{calcPhotosynthesisTemperature}
\usage{
calcPhotosynthesisTemperature(threshold = 0.8)
}
\arguments{
\item{threshold}{Photosynthesis efficiency threshold (between 0 and 1)}
}
\value{
magpie object
}
\description{
This function calculates crop-specific temperature limits for
             the multicropping mask based on the photosynthesis optimum
             and the LPJmL parameters temp_co2 and temp_photos
}
\examples{
\dontrun{
calcOutput("PhotosynthesisTemperature", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
