% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialIrrigWatRequirements.R
\name{calcPotentialIrrigWatRequirements}
\alias{calcPotentialIrrigWatRequirements}
\title{calcPotentialIrrigWatRequirements}
\usage{
calcPotentialIrrigWatRequirements(
  selectyears,
  iniyear,
  lpjml,
  climatetype,
  multicropping,
  transDist
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year (for weight by cropland)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; FALSE:NULL)}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates actual irrigation water requirements
             per cell given a certain irrigation system
}
\examples{
\dontrun{
calcOutput("PotentialIrrigWatRequirements", aggregate = FALSE)
}

}
\seealso{
\code{\link{calcActualIrrigWatRequirements}}, \code{\link{calcIrrigWatRequirements}}
}
\author{
Felicitas Beier
}
