% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProtectArea.R
\name{calcProtectArea}
\alias{calcProtectArea}
\title{calcProtectArea}
\usage{
calcProtectArea()
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts conservation protected area
}
\examples{
\dontrun{ calcOutput("ProtectArea", aggregate = FALSE) }

}
\author{
David Chen
}
