% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverDischargeNatAndHuman.R
\name{calcRiverDischargeNatAndHuman}
\alias{calcRiverDischargeNatAndHuman}
\title{calcRiverDischargeNatAndHuman}
\usage{
calcRiverDischargeNatAndHuman(
  lpjml,
  climatetype,
  selectyears,
  iniyear,
  efrMethod,
  comAg,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{iniyear}{Initialization year of irrigation system}

\item{efrMethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{comAg}{if TRUE: the currently already irrigated areas in
         initialization year are reserved for irrigation,
if FALSE: no irrigation areas reserved (irrigation potential)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; TRUE:exogenous; FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates cellular discharge after considering
             known human consumption (non-agricultural and committed agricultural)
             along the river calculated and accounted for in previous river
             routings (see calcRiverNaturalFlows and calcRiverHumanUses)
}
\examples{
\dontrun{
calcOutput("RiverDischargeNatAndHuman", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
