% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverHumanUses_ifandfor.R
\name{calcRiverHumanUses_ifandfor}
\alias{calcRiverHumanUses_ifandfor}
\title{calcRiverHumanUses_ifandfor}
\usage{
calcRiverHumanUses_ifandfor(
  selectyears,
  humanuse,
  subtype,
  climatetype,
  EFRmethod
)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{humanuse}{Human use type to which river routing shall be applied (non_agriculture or committed_agriculture). Note: non_agriculture must be run prior to committed_agriculture}

\item{subtype}{Subtype to be returned: discharge or required_wat_min or frac_fulfilled}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (Smakhtin:good, VMF:fair)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates natural discharge for the river routing derived from inputs from LPJmL
}
\examples{
\dontrun{ calcOutput("RiverHumanUses_ifandfor", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
