% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverHumanUses_ifandfor_backup.R
\name{calcRiverHumanUses_ifandfor_backup}
\alias{calcRiverHumanUses_ifandfor_backup}
\title{calcRiverHumanUses_ifandfor_backup}
\usage{
calcRiverHumanUses_ifandfor_backup(
  selectyears = "all",
  humanuse = "non_agriculture",
  subtype = "discharge",
  version = "LPJmL4",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = "CRU_4",
  ref_year = "y2015"
)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{humanuse}{Human use type to which river routing shall be applied (non_agriculture or committed_agriculture). Note: non_agriculture must be run prior to committed_agriculture}

\item{subtype}{Subtype to be returned: discharge or required_wat_min or frac_fulfilled}

\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates natural discharge for the river routing derived from inputs from LPJmL
}
\examples{
\dontrun{ calcOutput("RiverHumanUses_ifandfor_backup", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
