% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverHumanUses_magpie.R
\name{calcRiverHumanUses_magpie}
\alias{calcRiverHumanUses_magpie}
\title{calcRiverHumanUses_magpie}
\usage{
calcRiverHumanUses_magpie(
  selectyears = "all",
  humanuse = "non_agriculture",
  subtype = "discharge",
  iniyear = 1995,
  climatetype = "HadGEM2_ES:rcp2p6:co2"
)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{humanuse}{Human use type to which river routing shall be applied (non_agriculture or committed_agriculture). Note: non_agriculture must be run prior to committed_agriculture}

\item{subtype}{Subtype to be returned: discharge or required_wat_min or frac_fulfilled}

\item{iniyear}{Initialization year of irrigation system}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates natural discharge for the river routing derived from inputs from LPJmL
}
\examples{
\dontrun{ calcOutput("RiverHumanUses_magpie", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
