% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverNaturalFlows22.R
\name{calcRiverNaturalFlows22}
\alias{calcRiverNaturalFlows22}
\title{calcRiverNaturalFlows_magpie_magpie22}
\usage{
calcRiverNaturalFlows22(selectyears, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop. Note: Default version arguments need to be updated when new versions are used!}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates natural discharge and associated lake evaporation for the river routing derived from inputs from LPJmL
}
\examples{
\dontrun{ calcOutput("RiverNaturalFlows_magpie22", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
