% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverNaturalFlows_old.R
\name{calcRiverNaturalFlows_old}
\alias{calcRiverNaturalFlows_old}
\title{calcRiverNaturalFlows_old}
\usage{
calcRiverNaturalFlows_old(selectyears, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates natural discharge for the river routing derived from inputs from LPJmL
}
\examples{
\dontrun{ calcOutput("RiverNaturalFlows_old", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
