% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRrLayer.R
\name{calcRrLayer}
\alias{calcRrLayer}
\title{calcRrLayer}
\usage{
calcRrLayer()
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts range-rarity as used for biodiversity loss
}
\examples{
\dontrun{ calcOutput("RrLayer", aggregate = FALSE) }

}
\author{
Michael Windisch
}
