% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRunoffMonthly.R
\name{calcRunoffMonthly}
\alias{calcRunoffMonthly}
\title{calcRunoffMonthly}
\usage{
calcRunoffMonthly(lpjml, climatetype)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates monthly runoff from inputs provided by LPJmL
             It is used in calcAvlWater and will be deprecated when the mrwater
             pipeline is fully integrated in MAgPIE
}
\examples{
\dontrun{
calcOutput("RunoffMonthly", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
