% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSeaLevelRise.R
\name{calcSeaLevelRise}
\alias{calcSeaLevelRise}
\title{calcSeaLevelRise}
\usage{
calcSeaLevelRise()
}
\value{
gridded magpie object, with percentage of land lost per grid
}
\description{
Calculates gridded percentage of land lost from 2015, relative to 2015.
Currently linearly interpolated (for lack of better SLR function) from 2100 back, see readGCF_SeaLevelRise
}
\author{
David Chen
}
