% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShrHumanUsesFulfilled.R
\name{calcShrHumanUsesFulfilled}
\alias{calcShrHumanUsesFulfilled}
\title{calcShrHumanUsesFulfilled}
\usage{
calcShrHumanUsesFulfilled(
  lpjml,
  climatetype,
  transDist,
  multicropping,
  selectyears,
  iniyear,
  efrMethod
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}
}
\value{
cellular magpie object
}
\description{
calculates of share of current non-agricultural and irrigation
             that can be fulfilled
             given renewable water availability of the algorithm
}
\examples{
\dontrun{
calcOutput("ShrHumanUsesFulfilled", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
