% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSOCLossShare.R
\name{calcSOCLossShare}
\alias{calcSOCLossShare}
\title{calcSOCLossShare}
\usage{
calcSOCLossShare()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Calculates soil organic carbon loss share on cellular level
}
\examples{

\dontrun{
calcOutput("SOCLossShare", aggregate=FALSE)
}
}
\author{
Kristine Karstens
}
