% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSoilClassification.R
\name{calcSoilCharacteristics}
\alias{calcSoilCharacteristics}
\title{calcSoilCharacteristics}
\usage{
calcSoilCharacteristics()
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie objects with results on cellular level.
}
\description{
Calculate Soil Characteristics based on a HWDS soil classification map
}
\examples{

\dontrun{
  readSource("SoilClassification", subtype="HWSD.soil", convert="onlycorrect")
}

}
\seealso{
\code{\link{readSoilClassification}},
}
\author{
Marcos Alves
}
