% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSOMinitialsiationPools.R
\name{calcSOMinitialsiationPools}
\alias{calcSOMinitialsiationPools}
\title{calcSOMinitialsiationPools}
\usage{
calcSOMinitialsiationPools()
}
\value{
List of magpie object with results on country or cellular level, weight on cellular level, unit and description.
}
\description{
calculates Soil Organic Matter Pool, accounting for the management history as initialisation to magpie
}
\seealso{
\code{\link[mrcommons]{calcSOM}}
}
\author{
Benjamin Leon Bodirsky, Kristine Karstens
}
