% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopsoilCarbon.R
\name{calcTopsoilCarbon}
\alias{calcTopsoilCarbon}
\title{calcTopsoilCarbon}
\usage{
calcTopsoilCarbon(
  version = "LPJmL4",
  climatetype = "CRU_4",
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  harmonize_baseline = FALSE,
  ref_year = "y2015"
)
}
\arguments{
\item{version}{Switch between LPJmL4 and LPJmL4 (or LPJmL4+5)}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{average, spline or raw (default)}

\item{averaging_range}{just specify for time=="average": number of time steps to average}

\item{dof}{just specify for time=="spline": degrees of freedom}

\item{harmonize_baseline}{FALSE (default) nothing happens, if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts topsoil carbon densities from LPJ to MAgPIE
}
\examples{
\dontrun{ calcOutput("TopsoilCarbon", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
