% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransport_Distance.R
\name{calcTransport_Distance}
\alias{calcTransport_Distance}
\title{calcTransport_Distance}
\usage{
calcTransport_Distance(x)
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts travel time to major cities in minutes
}
\examples{
\dontrun{ calcOutput("Transport_Distance", aggregate = FALSE) }

}
\author{
David Chen
}
