% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterAllocation.R
\name{calcWaterAllocation}
\alias{calcWaterAllocation}
\title{calcWaterAllocation}
\usage{
calcWaterAllocation(
  selectyears = "all",
  output = "consumption",
  finalcells = "magpiecell",
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  allocationrule = "optimization",
  allocationshare = NULL,
  thresholdtype = TRUE,
  gainthreshold = 10,
  irrigationsystem = "initialization",
  iniyear = 1995,
  avlland_scen,
  proxycrop
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{output}{Water availability output to be returned: withdrawal or consumption}

\item{finalcells}{Number of cells to be returned by the function (lpjcell: 67420, magpiecell: 59199)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop. Note: Default version arguments need to be updated when new versions are used!}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{allocationshare}{Share of water to be allocated to cell (only needs to be selected in case of allocationrule=="equality")}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm: TRUE (default): monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{gainthreshold}{Threshold of yield improvement potential required for water allocation in upstreamfirst algorithm (in tons per ha)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{iniyear}{Initialization year of irrigation system}

\item{avlland_scen}{Land protection scenario}

\item{proxycrop}{historical crop mix pattern ("historical") or list of proxycrop(s)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water availability for MAgPIE retrieved from LPJmL using a river routing and allocation algorithm for distribution of discharge within the river basin
}
\examples{
\dontrun{ calcOutput("WaterAllocation", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
