% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterDemand.R
\name{calcWaterDemand}
\alias{calcWaterDemand}
\title{calcWaterDemand}
\usage{
calcWaterDemand(
  years = seq(1995, 2095, by = 5),
  seasonality = "grper",
  source = "WATERGAP2013"
)
}
\arguments{
\item{years}{years to be returned}

\item{source}{data source to be used}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts non-agricultural water demand
}
\examples{
\dontrun{ calcOutput("WaterDemand", aggregate = FALSE) }

}
\author{
Felicitas Beier
}
