% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterPotUse.R
\name{calcWaterPotUse}
\alias{calcWaterPotUse}
\title{calcWaterPotUse}
\usage{
calcWaterPotUse(
  selectyears,
  climatetype,
  EFRmethod,
  variabilitythreshold,
  rankmethod,
  allocationrule,
  thresholdtype,
  gainthreshold,
  irrigationsystem,
  iniyear,
  avlland_scen,
  proxycrop
)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{variabilitythreshold}{Scalar value defining the strictness of accessibility restriction: discharge that is exceeded x percent of the time on average throughout a year (Qx). Default: 0.5 (Q50) (e.g. Q75: 0.25, Q50: 0.5)}

\item{rankmethod}{method of calculating the rank: "meancellrank" (default): mean over cellrank of proxy crops, "meancroprank": rank over mean of proxy crops (normalized), "meanpricedcroprank": rank over mean of proxy crops (normalized using price), "watervalue": rank over value of irrigation water; and fullpotentail TRUE/FALSE separated by ":" (TRUE: Full irrigation potential (cell receives full irrigation requirements in total area). FALSE: reduced potential of cell receives at later stage in allocation algorithm)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm: TRUE (default): monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{gainthreshold}{Threshold of yield improvement potential required for water allocation in upstreamfirst algorithm (in tons per ha)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{iniyear}{Initialization year of irrigation system}

\item{avlland_scen}{Land availability scenario: current or potential; optional additionally: protection scenario in case of potential (when left empty: no protection) and initialization year of cropland area
combination of land availability scenario and initialization year separated by ":". land availability scenario: currIrrig (only currently irrigated cropland available for irrigated agriculture), currCropland (only current cropland areas available for irrigated agriculture), potIrrig (suitable land is available for irrigated agriculture, potentially land restrictions activated through protect_scen argument)
protection scenario separated by "_" (only relevant when potIrrig selected): WDPA, BH, FF, CPD, LW, HalfEarth. Areas where no irrigation water withdrawals are allowed due to biodiversity protection.}

\item{proxycrop}{historical crop mix pattern ("historical") or list of proxycrop(s)}
}
\value{
magpie object in cellular resolution
}
\description{
This function returns the potential water quantity available for different uses
}
\examples{
\dontrun{ calcOutput("WaterPotUse", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
