% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterUseActuallyCommittedAg.R
\name{calcWaterUseActuallyCommittedAg}
\alias{calcWaterUseActuallyCommittedAg}
\title{calcWaterUseActuallyCommittedAg}
\usage{
calcWaterUseActuallyCommittedAg(
  iteration = "committed_agriculture",
  lpjml,
  climatetype,
  selectyears,
  iniyear,
  multicropping,
  efrMethod,
  fossilGW,
  transDist
)
}
\arguments{
\item{iteration}{Default: "committed_agriculture",
Special case: "committed_agriculture_fullPotential".
Special case should only be used for calculation of
full multicropping potential committed agricultural area
for case of Current Irrigation.}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned}

\item{iniyear}{Year of initialization for cropland area}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from LandInG
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{efrMethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{fossilGW}{If TRUE: non-renewable groundwater can be used.
If FALSE: non-renewable groundwater cannot be used.}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates committed agricultural water uses that
             are used in the river routing algorithm for distributing
             available water across the basin
}
\examples{
\dontrun{
calcOutput("WaterUseActuallyCommittedAg", aggregate = FALSE)
}
}
\author{
Felicitas Beier
}
