% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterUseNonAg.R
\name{calcWaterUseNonAg}
\alias{calcWaterUseNonAg}
\title{calcWaterUseNonAg}
\usage{
calcWaterUseNonAg(
  selectyears = seq(1995, 2100, by = 5),
  cells = "lpjcell",
  datasource = "WATCH_ISIMIP_WATERGAP",
  usetype = "all",
  seasonality = "grper",
  harmonType = "average",
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb"),
  climatetype = "GSWP3-W5E5:historical"
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{cells}{Number of cells to be returned
(select "magpiecell" for 59199 cells or "lpjcell" for 67420 cells)}

\item{datasource}{Data source to be used (e.g. WATERGAP2020)}

\item{usetype}{water use types (domestic, industry, electricity)
and option to return withdrawals or consumption
separated by ":" (e.g. "all:withdrawal")
options for first argument:
"total" (returns the sum over different
water use types) or
"all" ( returns all water use types
(domestic, industry, electricity))
options for second argument: "all", "withdrawal", "consumption"}

\item{seasonality}{grper (default): non-agricultural water demand in growing period per year;
total: non-agricultural water demand throughout the year}

\item{harmonType}{Type of time smoothing:
average (average over 8-year time span around baseline year) or
spline (time smoothing using spline method with 4 degrees of freedom) or
NULL (no smoothing)}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios for calcGrowingPeriod}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts non-agricultural water demand
}
\examples{
\dontrun{
calcOutput("WaterUseNonAg", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
