% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWatReqMAgPIE.R
\name{calcWatReqMAgPIE}
\alias{calcWatReqMAgPIE}
\title{calcWatReqMAgPIE}
\usage{
calcWatReqMAgPIE(selectyears, iniyear, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year (for weight by cropland)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function returns irrigation water withdrawal requirements
             per crop, irrigation system, season and cell
             to be used in MAgPIE
}
\examples{
\dontrun{
calcOutput("WatReqMAgPIE", aggregate = FALSE)
}

}
\seealso{
\code{\link{calcIrrigSystemShr}}, \code{\link{calcIrrigWatRequirements}}
}
\author{
Felicitas Beier
}
