% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYearlyRunoff.R
\name{calcYearlyRunoff}
\alias{calcYearlyRunoff}
\title{calcYearlyRunoff}
\usage{
calcYearlyRunoff(selectyears, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop. Note: Default version arguments need to be updated when new versions are used!}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates yearly runoff from runoff on land and water provided by LPJmL
}
\examples{
\dontrun{ calcOutput("YearlyRunoff", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
