% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldgainWatUse.R
\name{calcYieldgainWatUse}
\alias{calcYieldgainWatUse}
\title{calcYieldgainWatUse}
\usage{
calcYieldgainWatUse(
  lpjml,
  climatetype,
  selectyears,
  iniyear,
  landScen,
  cropmix,
  yieldcalib,
  multicropping,
  rangeGT
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned}

\item{iniyear}{Year of initialization for cropland area}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_FF, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{rangeGT}{Range of gainthreshold for calculation of potentially
irrigated areas}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water use on selected land areas
             without considering water constraints
}
\examples{
\dontrun{
calcOutput("YieldgainWatUse", aggregate = FALSE)
}
}
\author{
Felicitas Beier
}
