% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYields.R
\name{calcYields}
\alias{calcYields}
\title{calcYields}
\usage{
calcYields()
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts yields from LPJ to MAgPIE
}
\examples{
\dontrun{ calcOutput("Yields", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
