% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsPotential.R
\name{calcYieldsPotential}
\alias{calcYieldsPotential}
\title{calcYieldsPotential}
\usage{
calcYieldsPotential(lpjml, climatetype, selectyears)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical" for yields}

\item{selectyears}{Years to be returned by the function}
}
\value{
magpie object in cellular resolution
}
\description{
This function returns irrigated and rainfed yield potential from LPJmL for magpie crops.
}
\examples{
\dontrun{ calcOutput("YieldsPotential", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
