% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsReplaceISIMIP.R
\name{calcYieldsReplaceISIMIP}
\alias{calcYieldsReplaceISIMIP}
\title{calcYields}
\usage{
calcYieldsReplaceISIMIP(
  version = "LPJmL5",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  dof = 4,
  harmonize_baseline = "CRU_4",
  ref_year = "y2015",
  isimip_subtype = "ISIMIP3b:yields.EPIC-IIASA_ukesm1-0-ll_ssp585_default"
)
}
\arguments{
\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{spline or average, if spline specify dof, if average specify averaging range}

\item{dof}{degrees of freedom for spline}

\item{harmonize_baseline}{FALSE (default) nothing happens, if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}

\item{isimip_subtype}{isimip 3b yields subtype}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts yields from LPJmL to MAgPIE
}
\examples{
\dontrun{ calcOutput("YieldsReplaceISIMIP", aggregate = FALSE) }

}
\author{
David Chen
}
