% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDams.R
\name{convertDams}
\alias{convertDams}
\title{convertDams}
\usage{
convertDams(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Convert Dams content
}
\examples{

\dontrun{
  readSource("Dams", convert=TRUE)
}

}
\seealso{
\code{\link{readDams}}
}
\author{
Kristine Karstens
}
