% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGPD.R
\name{convertGPD}
\alias{convertGPD}
\title{convertGPD}
\usage{
convertGPD(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on iso level, weight, unit and description.
}
\description{
convert GPD
}
\examples{

\dontrun{
  readSource("GPD", convert=TRUE)
}
}
\author{
Florian Humpenoeder
}
