% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertISIMIPoutputs.R
\name{convertISIMIPoutputs}
\alias{convertISIMIPoutputs}
\title{convertISIMIPoutputs}
\usage{
convertISIMIPoutputs(x)
}
\arguments{
\item{x}{MAgPIE object on cellular level}
}
\value{
MAgPIE object on country level
}
\description{
convert data to ISO country level
}
\examples{
\dontrun{
a <- readSource("ISIMIPoutputs", convert=TRUE)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Felicitas Beier
}
