% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctAfforestationMask.R
\name{correctAfforestationMask}
\alias{correctAfforestationMask}
\title{correctAfforestationMask}
\usage{
correctAfforestationMask(x, subtype)
}
\arguments{
\item{x}{magpie object provided by the read function}

\item{subtype}{afforestation mask sub type}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
correct Afforestation Mask where afforestation is possible
}
\examples{

\dontrun{
  readSource("AfforestationMask",subtype="unrestricted", convert="onlycorrect")
}

}
\seealso{
\code{\link{readAfforestationMask}}
}
\author{
David Chen
}
