% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctAtmosphericDepositionRates.R
\name{correctAtmosphericDepositionRates}
\alias{correctAtmosphericDepositionRates}
\title{correctAtmosphericDepositionRates}
\usage{
correctAtmosphericDepositionRates(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie object with results on cellular level
}
\description{
Correct Atmospheric deposition rate
}
\examples{

\dontrun{
  readSource("AtmosphericDepositionRates", convert="onlycorrect")
}

}
\seealso{
\code{\link{readAtmosphericDepositionRates}}
}
\author{
Marcos Alves
}
