% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctAvlIrrig.R
\name{correctAvlIrrig}
\alias{correctAvlIrrig}
\title{correctAvlIrrig}
\usage{
correctAvlIrrig(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read available irrigation file
}
\examples{

\dontrun{
  readSource("Avlirrig", convert="onlycorrect")
}

}
\seealso{
\code{\link{readAvlIrrig}}
}
\author{
David Chen
}
