% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctAvlLandSi.R
\name{correctAvlLandSi}
\alias{correctAvlLandSi}
\title{correctAvlLandSi}
\usage{
correctAvlLandSi(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Available Land Si
}
\examples{

\dontrun{
  readSource("AvlLandSi", convert="onlycorrect")
}

}
\seealso{
\code{\link{readAvlLandSi}}
}
\author{
David Chen
}
