% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBphEffect.R
\name{correctBphEffect}
\alias{correctBphEffect}
\title{correctBphEffect}
\usage{
correctBphEffect(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Biogeophysical temperature change of afforestation (degree C). File is based on observation datasets of Bright et al. 2017 and Duveiller et al. 2018
}
\examples{

\dontrun{
  readSource("BphEffect", convert="onlycorrect")
}

}
\seealso{
\code{\link{readBphEffect}}
}
\author{
Michael Windisch
}
