% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBphTCRE.R
\name{correctBphTCRE}
\alias{correctBphTCRE}
\title{correctBphTCRE}
\usage{
correctBphTCRE(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Transient Climate Response to accumulated doubling of CO2. File based on CMIP5 +1perc CO2 per year experiment. To be used in the translation to carbon equivalents of BphEffect
}
\examples{

\dontrun{
  readSource("BphTCRE", convert="onlycorrect")
}

}
\seealso{
\code{\link{readBphTCRE}}
}
\author{
Michael Windisch
}
