% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctCalibratedArea.R
\name{correctCalibratedArea}
\alias{correctCalibratedArea}
\title{correctCalibratedArea}
\usage{
correctCalibratedArea(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\description{
Read calibrated area file
#' @return List of magpie objects with results on cellular level, weight, unit and description.
}
\examples{

\dontrun{
  readSource("CalibratedArea", convert="onlycorrect")
}
}
\seealso{
\code{\link{readCalibratedArea}}
}
\author{
David Chen
}
