% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctCO2Atmosphere.R
\name{correctCO2Atmosphere}
\alias{correctCO2Atmosphere}
\title{correctCO2Atmosphere}
\usage{
correctCO2Atmosphere(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct atmosphere co2 concentration
}
\examples{

\dontrun{
readSource("CO2Atmosphere", subtype="rcp85", convert="onlycorrect")
}

}
\seealso{
\code{\link{readCO2Atmosphere}}
}
\author{
Marcos Alves, Kristine Karstens
}
