% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctCshareReleased.R
\name{correctCshareReleased}
\alias{correctCshareReleased}
\title{correctCshareReleased}
\usage{
correctCshareReleased(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read "Share of soil carbon that is released on cropland during the harvest. Calculated by Benjamin Bodirsky"
}
\examples{

\dontrun{
  readSource("CshareReleased", convert="onlycorrect")
}

}
\seealso{
\code{\link{readCshareReleased}}
}
\author{
David Chen
}
