% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctDams.R
\name{correctDams}
\alias{correctDams}
\title{correctDams}
\usage{
correctDams(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read dam file (no source information available)
}
\examples{
\dontrun{
readSource("Dams", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readDams}}
}
\author{
Kristine Karstens
}
