% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGFAD.R
\name{correctGFAD}
\alias{correctGFAD}
\title{correctGFAD}
\usage{
correctGFAD(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct Global forest age dataset
}
\examples{

\dontrun{
  readSource("GFAD", convert="onlycorrect")
}

}
\seealso{
\code{\link{readGFAD}}
}
\author{
Abhijeet Mishra, Felicitas Beier
}
