% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGFADArea.R
\name{correctGFADArea}
\alias{correctGFADArea}
\title{correctGFADArea}
\usage{
correctGFADArea(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct Global forest age dataset
}
\examples{

\dontrun{
  readSource("GFADArea", convert="onlycorrect")
}

}
\seealso{
\code{\link{readGFADArea}}
}
\author{
Abhijeet Mishra, Felicitas Beier
}
