% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGLW3.R
\name{correctGLW3}
\alias{correctGLW3}
\title{correctGLW3}
\usage{
correctGLW3(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read GLW3 file
}
\examples{

\dontrun{
  readSource("GLW3", subtype = "DA", convert="onlycorrect")
}

}
\seealso{
\code{\link{readGLW3}}
}
\author{
Marcos Alves
}
