% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGPD.R
\name{correctGPD}
\alias{correctGPD}
\title{correctGPD}
\usage{
correctGPD(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
correct GPD
}
\examples{

\dontrun{
  readSource("GPD", convert="onlycorrect")
}
}
\author{
Florian Humpenoeder
}
