% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctHalfEarth.R
\name{correctHalfEarth}
\alias{correctHalfEarth}
\title{correctHalfEarth}
\usage{
correctHalfEarth(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
correct HalfEarth data
}
\examples{

\dontrun{
  readSource("HalfEarth", convert="onlycorrect")
}

}
\seealso{
\code{\link{readHalfEarth}}
}
\author{
Felicitas Beier
}
