% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctIrrigationSystem.R
\name{correctIrrigationSystem}
\alias{correctIrrigationSystem}
\title{correctIrrigationSystem}
\usage{
correctIrrigationSystem(x)
}
\arguments{
\item{x}{MAgPIE object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct Irrigation System data
}
\examples{
\dontrun{
  readSource("IrrigationSystem", convert="onlycorrect")
}

}
\seealso{
\code{\link{readIrrigationSystem}}
}
\author{
Felicitas Beier
}
