% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctISIMIP.R
\name{correctISIMIP}
\alias{correctISIMIP}
\title{correctISIMIP}
\usage{
correctISIMIP(x)
}
\arguments{
\item{x}{MAgPIE object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct ISIMIP data
}
\examples{
\dontrun{
  readSource("ISIMIP", convert="onlycorrect")
}

}
\seealso{
\code{\link{readISIMIP}}
}
\author{
Felicitas Beier
}
