% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctISIMIPinputs.R
\name{correctISIMIPinputs}
\alias{correctISIMIPinputs}
\title{correctISIMIPinputs}
\usage{
correctISIMIPinputs(x)
}
\arguments{
\item{x}{MAgPIE object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct ISIMIP data
}
\examples{
\dontrun{
  readSource("ISIMIPinputs", convert="onlycorrect")
}

}
\seealso{
\code{\link{readISIMIPinputs}}
}
\author{
Felicitas Beier
}
