% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctKoeppen_geiger.R
\name{correctKoeppen_geiger}
\alias{correctKoeppen_geiger}
\title{correctKoeppen_geiger}
\usage{
correctKoeppen_geiger(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\description{
Read calibrated area file
#' @return List of magpie objects with results on cellular level, weight, unit and description.
}
\examples{

\dontrun{
  readSource("Koeppen_geiger", convert="onlycorrect")
}

}
\seealso{
\code{\link{correctKoeppen_geiger}}
}
\author{
David Chen
}
