% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLeifeld2018.R
\name{correctLeifeld2018}
\alias{correctLeifeld2018}
\title{correctLeifeld2018}
\usage{
correctLeifeld2018(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
correct potential peatland area from Leifeld2018
}
\examples{

\dontrun{
  readSource("Leifeld2018", convert="onlycorrect")
}
}
\author{
Florian Humpenoeder
}
