% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLuh2SideLayers.R
\name{correctLuh2SideLayers}
\alias{correctLuh2SideLayers}
\title{correctLuh2SideLayers}
\usage{
correctLuh2SideLayers(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read LUH2 land cover type file
}
\examples{

\dontrun{
  readSource("Luh2SideLayers", convert="onlycorrect")
}

}
\seealso{
\code{\link{readLuh2SideLayers}}
}
\author{
Michael Windisch
}
