% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctNitrogenFixationRateNatural.R
\name{correctNitrogenFixationRateNatural}
\alias{correctNitrogenFixationRateNatural}
\title{correctNitrogenFixationRateNatural}
\usage{
correctNitrogenFixationRateNatural(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie object with results on cellular level
}
\description{
Correct Nitrogen fixation rate of natural vegetation
}
\examples{

\dontrun{
  readSource("NitrogenFixationRateNatural", convert="onlycorrect")
}

}
\seealso{
\code{\link{readNitrogenFixationRateNatural}}
}
\author{
Marcos Alves
}
