% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctNpiNdcAdAolcPol.R
\name{correctNpiNdcAdAolcPol}
\alias{correctNpiNdcAdAolcPol}
\title{correctNpiNdcAdAolcPol}
\usage{
correctNpiNdcAdAolcPol(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read NPI/NDC forestry file
}
\examples{

\dontrun{
  readSource("NpiNdcAdAolcPol", convert="onlycorrect")
}

}
\seealso{
\code{\link{readNpiNdcAdAolcPol}}
}
\author{
Michael Windisch
}
