% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctNpiNdcAffPol.R
\name{correctNpiNdcAffPol}
\alias{correctNpiNdcAffPol}
\title{correctNpiNdcAffPol}
\usage{
correctNpiNdcAffPol(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read dummy NPI/NDC policies file
}
\examples{

\dontrun{
  readSource("NpiNdcAffPol", convert="onlycorrect")
}

}
\seealso{
\code{\link{readNpiNdcAffPol}}
}
\author{
Michael Windisch
}
