% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctProtectArea.R
\name{correctProtectArea}
\alias{correctProtectArea}
\title{correctProtectArea}
\usage{
correctProtectArea(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\description{
Read calibrated area file
#' @return List of magpie objects with results on cellular level, weight, unit and description.
}
\examples{

\dontrun{
  readSource("ProtectArea", convert="onlycorrect")
}

}
\seealso{
\code{\link{readProtectArea}}
}
\author{
David Chen
}
