% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctRrLayer.R
\name{correctRrLayer}
\alias{correctRrLayer}
\title{correctRrLayer}
\usage{
correctRrLayer(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read range-rarity layer file
}
\examples{

\dontrun{
  readSource("RrLayer", convert="onlycorrect")
}

}
\seealso{
\code{\link{readRrLayer}}
}
\author{
Michael Windisch
}
