% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTransport_Distance.R
\name{correctTransport_Distance}
\alias{correctTransport_Distance}
\title{correctTransport_Distance}
\usage{
correctTransport_Distance(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read transport distance file (no source information available)
}
\examples{

\dontrun{
  readSource("Transport_Distance", convert="onlycorrect")
}

}
\seealso{
\code{\link{readDams}}
}
\author{
Kristine Karstens
}
