% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctWATERGAP.R
\name{correctWATERGAP}
\alias{correctWATERGAP}
\title{correctWATERGAP}
\usage{
correctWATERGAP(x)
}
\arguments{
\item{x}{MAgPIE object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct WATERGAP data
}
\examples{

\dontrun{
  readSource("WATERGAP", convert="onlycorrect")
}

}
\seealso{
\code{\link{readWATERGAP}}
}
\author{
Felicitas Beier
}
