% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadJaegermeyr2015.R
\name{downloadJaegermeyr2015}
\alias{downloadJaegermeyr2015}
\title{downloadJaegermeyr2015}
\usage{
downloadJaegermeyr2015(subtype)
}
\arguments{
\item{subtype}{Data to be downloaded:
"systemShare": irrigation system share
               as provided in SI of Jaegermeyr et al. (2015),
               based on FAO 2014, ICID 2012 and Rohwer et al. 2007);
"systemSuitability": biophysical and technical irrigation system suitability
                     by crop type (CFT) as provided in Table 2 of Jaegermeyr et al. (2015)
                     based on Sauer et al. (2010) and Fischer et al (2012).}
}
\description{
Downloads irrigation system suitability per crop type
             and country-level irrigation system shares from Jaegermeyr (2015)
}
\examples{
\dontrun{
readSource("Jaegerymeyr2015", convert = FALSE)
}
}
\author{
Felicitas Beier
}
