% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullCURRENTIRRIGATION.R
\name{fullCURRENTIRRIGATION}
\alias{fullCURRENTIRRIGATION}
\title{fullCURRENTIRRIGATION}
\usage{
fullCURRENTIRRIGATION(yieldcalib = "TRUE:TRUE:actual:irrig_crop")
}
\arguments{
\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
        Also needs specification of refYields, separated by ":".
        Options: FALSE (for single cropping analyses) or
                 "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}
}
\description{
Function that produces output for irrigation potentials
             under multiple cropping on cellular resolution.
}
\author{
Felicitas Beier
}
