% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullDETERMINANTS.R
\name{fullDETERMINANTS}
\alias{fullDETERMINANTS}
\title{fullDETERMINANTS}
\usage{
fullDETERMINANTS()
}
\description{
Function that returns potential determinants of irrigation for
             regression analysis
}
\author{
Felicitas Beier
}
