% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPREPROCESSING.R
\name{fullPREPROCESSING}
\alias{fullPREPROCESSING}
\title{fullPREPROCESSING}
\usage{
fullPREPROCESSING(
  protectLand = "HalfEarth",
  yieldcalib = "TRUE:TRUE:actual:irrig_crop",
  multicropping = "TRUE:potential:endogenous",
  cropmix = "hist_total",
  climatetype = "MRI-ESM2-0:ssp370",
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb")
)
}
\arguments{
\item{protectLand}{Land protection scenario (e.g. HalfEarth, BH_IFL, NULL)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
     Also needs specification of refYields, separated by ":".
     Options: FALSE (for single cropping analyses) or
              "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}
}
\description{
Function that produces the objects for Technical and Economic
             Irrigation Potentials within land and water boundaries
}
\author{
Felicitas Beier
}
