% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTRANSPORT.R
\name{fullTRANSPORT}
\alias{fullTRANSPORT}
\title{fullTRANSPORT}
\usage{
fullTRANSPORT(multicropping)
}
\arguments{
\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
          by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
         GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; TRUE:exogenous; FALSE)}
}
\description{
Function that produces output for analysis of
             water transport for provision to cells in surrounding
}
\author{
Felicitas Beier
}
