% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullWaterOutputs.R
\name{fullWaterOutputs}
\alias{fullWaterOutputs}
\title{fullWaterOutputs}
\usage{
fullWaterOutputs(
  EFRmethod = "VMF:fair",
  accessibilityrule = "CV:2",
  allocationrule = "optimization",
  rankmethod = "USD_ha:TRUE",
  thresholdtype = "USD_ha",
  gainthreshold = 500,
  avlland_scen = "currCropland:2010"
)
}
\arguments{
\item{EFRmethod}{EFR method used including selected strictness of EFRs
(Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Method used: Quantile method (Q) or Coefficient of Variation (CV)
combined with scalar value defining the strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year
(Qx, e.g. Q75: 0.25, Q50: 0.5)
or base value for exponential curve separated by : (CV:2)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst")}

\item{rankmethod}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return;
and boolean indicating fullpotential (TRUE, i.e. cell
receives full irrigation requirements in total area)
or reduced potential (FALSE, reduced potential of cell
receives at later stage in allocation algorithm);
separated by ":"}

\item{thresholdtype}{Unit of yield improvement potential used as threshold:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{gainthreshold}{Threshold of yield improvement potential required for
water allocation in upstreamfirst algorithm
(in same unit as thresholdtype)}

\item{avlland_scen}{Land availability scenario: current or potential;
optional additionally: protection scenario in case of potential
(when left empty: no protection)
and initialization year of cropland area
combination of land availability scenario and initialization year separated by ":".
land availability scenario: currIrrig (only currently
irrigated cropland available for irrigated agriculture),
currCropland (only current cropland areas available for irrigated agriculture),
potIrrig (suitable land is available for irrigated agriculture,
potentially land restrictions activated through protect_scen argument)
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth}
}
\description{
Function that produces the objects for water outputs that can
             be plotted with the mrwaterPlots functions.
}
\author{
Felicitas Beier
}
