% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistReturnToIrrig.R
\name{plotHistReturnToIrrig}
\alias{plotHistReturnToIrrig}
\title{plotHistReturnToIrrig}
\usage{
plotHistReturnToIrrig(
  scenario,
  iniyear,
  lpjml,
  selectyears,
  climatetype,
  efrMethod,
  cropmix,
  thresholdtype,
  multicropping,
  yieldcalib
)
}
\arguments{
\item{scenario}{EFP and non-agricultural water use scenario separated by "."
(e.g. "on.ssp2")}

\item{iniyear}{Initialization year for committed agricultural uses}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{Year of plot (Note: one single year must be selected)}

\item{climatetype}{Switch between different climate models or historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm: TRUE (default): monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}
}
\value{
map of magpie cells
}
\description{
plot map of share of current irrigation that can be fulfilled given surface water availability of the algorithm
}
\examples{
\dontrun{
plotHistReturnToIrrig()
}

}
\author{
Felicitas Beier
}
