% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistShrCurrIrrigFulfilled.R
\name{plotHistShrCurrIrrigFulfilled}
\alias{plotHistShrCurrIrrigFulfilled}
\title{plotHistShrCurrIrrigFulfilled}
\usage{
plotHistShrCurrIrrigFulfilled(
  scenario,
  iniyear,
  lpjml,
  selectyears,
  climatetype,
  EFRmethod
)
}
\arguments{
\item{scenario}{EFP and non-agricultural water use scenario separated by "." (e.g. "on.ssp2")}

\item{iniyear}{initialization year}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{years for which irrigatable area is calculated}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}
}
\value{
map of magpie cells
}
\description{
plot map of share of current irrigation that can be fulfilled given surface water availability of the algorithm
}
\examples{
\dontrun{
plotHistShrCurrIrrigFulfilled()
}

}
\author{
Felicitas Beier
}
