% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapCellranking.R
\name{plotMapCellranking}
\alias{plotMapCellranking}
\title{plotMapCellranking}
\usage{
plotMapCellranking(
  lpjml,
  climatetype,
  selectyears,
  rankmethod,
  cropmix,
  iniyear,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Year to be displayed}

\item{rankmethod}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return;
and boolean indicating fullpotential (TRUE, i.e. cell receives full irrigation
                                            requirements in total area)
or reduced potential (FALSE, reduced potential of cell receives
                             at later stage in allocation algorithm);
separated by ":"}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{iniyear}{Initialization year of irrigation system}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
map of cellranking
}
\examples{
\dontrun{
plotMapCellranking()
}

}
\author{
Felicitas Beier
}
