% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmapIrrigatableArea.R
\name{plotmapIrrigatableArea}
\alias{plotmapIrrigatableArea}
\title{plotmapIrrigatableArea}
\usage{
plotmapIrrigatableArea(
  selectyears = 1995,
  cells = "lpjcell",
  output = "irrigatable_area",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = "CRU_4",
  ref_year = "y2015",
  allocationrule = "optimization",
  allocationshare = NULL,
  gainthreshold = 1,
  irrigationsystem = "initialization",
  irrigini = "Jaegermeyr_lpjcell",
  iniyear = 1995,
  landtype = "potentialcropland",
  proxycrops = "maiz",
  EFP = "on",
  scen = "ssp2"
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{cells}{cells used for calculation ("lpjcell" or "magpiecell")}

\item{output}{Type of area to be returned: irrigatable_area (default, considers both water and land availability), irrig_area_ww or irrig_area_wc (area that can be irrigated given water available for withdrawal or consumption), avl_land}

\item{climatetype}{switch between different climate scenarios for yields}

\item{time}{time smoothing: average, spline (default) or raw}

\item{averaging_range}{just specify for time=="average": number of time steps to average}

\item{dof}{just specify for time=="spline": degrees of freedom}

\item{harmonize_baseline}{FALSE (default) no harmonization, harmonization: if a baseline is specified here data is harmonized to that baseline (from ref_year onwards)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{allocationshare}{Share of water to be allocated to cell (only needs to be selected in case of allocationrule=="equality")}

\item{gainthreshold}{Threshold of yield improvement potential required for water allocation in upstreamfirst algorithm (in tons per ha)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{irrigini}{When "initialization" selected for irrigation system: choose initialization data set for irrigation system initialization ("Jaegermeyr_lpjcell", "LPJmL_lpjcell")}

\item{iniyear}{Initialization year of irrigation system}

\item{landtype}{current cropland area (currentcropland) or potential cropland area (potentialcropland)}

\item{proxycrops}{proxycrops used for cell ranking}

\item{EFP}{switch for environmental flow policy ("on" or "off")}

\item{scen}{non-agricultural water demand scenario ("ssp1", "ssp2", "ssp3")}
}
\value{
magpie object in cellular resolution
}
\description{
This function plots the irrigatable area on a global map
}
\examples{
\dontrun{ plotmapIrrigatableArea() }
}
\author{
Felicitas Beier
}
