% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmapIrrigCellranking.R
\name{plotmapIrrigCellranking}
\alias{plotmapIrrigCellranking}
\title{plotIrrigCellranking}
\usage{
plotmapIrrigCellranking(
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = FALSE,
  ref_year = "y2015",
  cellrankyear = "y1995",
  cells = "lpjcell",
  crops = "magpie",
  method = "meancellrank",
  proxycrop = c("maiz", "rapeseed", "puls_pro"),
  iniyear = 1995
)
}
\arguments{
\item{climatetype}{switch between different climate scenarios for yields}

\item{time}{time smoothing: average, spline (default) or raw}

\item{averaging_range}{just specify for time=="average": number of time steps to average}

\item{dof}{just specify for time=="spline": degrees of freedom}

\item{harmonize_baseline}{FALSE (default) no harmonization, harmonization: if a baseline is specified here data is harmonized to that baseline (from ref_year onwards)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}

\item{cellrankyear}{year(s) for which cell rank is calculated}

\item{cells}{switch between "lpjcell" (67420) and "magpiecell" (59199)}

\item{crops}{switch between "magpie" and "lpjml" (default) crops}

\item{method}{method of calculating the rank: "meancellrank" (default): mean over cellrank of proxy crops, "meancroprank": rank over mean of proxy crops (normalized), "meanpricedcroprank": rank over mean of proxy crops (normalized using price)}

\item{proxycrop}{proxycrop(s) selected for rank calculation}

\item{iniyear}{initialization year for price in price-weighted normalization of meanpricedcroprank}
}
\value{
magpie object in cellular resolution
}
\description{
This function plots the cellranking based on yield improvement potential through irrigation on a global map
}
\examples{
\dontrun{ plotIrrigCellranking() }
}
\author{
Felicitas Beier
}
