% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapPotReturnToIrrigation.R
\name{plotMapPotReturnToIrrigation}
\alias{plotMapPotReturnToIrrigation}
\title{plotMapPotReturnToIrrigation}
\usage{
plotMapPotReturnToIrrigation(
  selectyears,
  unit,
  iniyear,
  lpjml,
  climatetype,
  yieldcalib,
  multicropping,
  cropmix,
  landScen
)
}
\arguments{
\item{selectyears}{year for which map is displayed}

\item{unit}{unit of return to irrigation
(perha for monetary yield gain per cell in USD05/ha
or perm3 for irrigation water value in USD05/m^3)}

\item{iniyear}{Initialization year for return to irrigation}

\item{lpjml}{LPJmL version required for return to irrigation}

\item{climatetype}{Climate scenario for return to irrigation}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{multicropping for return to irrigation (TRUE or FALSE)}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_FF, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}
}
\value{
map of magpie cells
}
\description{
plots map of return to irrigation in terms of monetary yield gain in USD/ha
             or water value in USD/m^3
}
\examples{
\dontrun{
plotMapPotReturnToIrrigation()
}

}
\author{
Felicitas Beier
}
