% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapShrCurrIrrigFulfilled.R
\name{plotMapShrCurrIrrigFulfilled}
\alias{plotMapShrCurrIrrigFulfilled}
\title{plotMapShrCurrIrrigFulfilled}
\usage{
plotMapShrCurrIrrigFulfilled(
  scenario,
  iniyear,
  lpjml,
  selectyears,
  climatetype,
  efrMethod
)
}
\arguments{
\item{scenario}{EFP and non-agricultural water use scenario
separated by "." (e.g. "on.ssp2")}

\item{iniyear}{Initialization year}

\item{lpjml}{LPJmL version used}

\item{selectyears}{Years to be returned}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}
}
\value{
map of magpie cells
}
\description{
plot map of share of current irrigation that can be fulfilled
             given renewable water availability of the algorithm
}
\examples{
\dontrun{
plotMapShrCurrIrrigFulfilled()
}

}
\author{
Felicitas Beier
}
