% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginalReturnToIrrigationFull.R
\name{plotMarginalReturnToIrrigationFull}
\alias{plotMarginalReturnToIrrigationFull}
\title{plotMarginalReturnToIrrigationFull}
\usage{
plotMarginalReturnToIrrigationFull(
  y_axis_range,
  x_axis,
  region = "GLO",
  scenario,
  lpjml,
  selectyears,
  iniyear,
  climatetype,
  EFRmethod,
  accessibilityrule,
  rankmethod,
  yieldcalib,
  allocationrule,
  thresholdtype,
  irrigationsystem,
  cropmix,
  potential_wat = TRUE,
  com_ag,
  multicropping
)
}
\arguments{
\item{y_axis_range}{range of y-axis (monetary irrigation gain) to be depicted on the curve}

\item{x_axis}{x_axis type to be displayed: irrigated area "IrrigArea" or available water volume "wat_ag_ww" "wat_ag_wc" "wat_tot_ww" "wat_tot_wc"}

\item{region}{regional resolution (can be country iso-code, region name and respective mapping "EUR:H12", "GLO" for global)}

\item{scenario}{non-agricultural water use scenario to be displayed in plot}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{years for which irrigatable area is calculated}

\item{iniyear}{initialization year}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Scalar value defining the strictness of accessibility restriction: discharge that is exceeded x percent of the time on average throughout a year (Qx). Default: 0.5 (Q50) (e.g. Q75: 0.25, Q50: 0.5)}

\item{rankmethod}{method of calculating the rank: "meancellrank" (default): mean over cellrank of proxy crops, "meancroprank": rank over mean of proxy crops (normalized), "meanpricedcroprank": rank over mean of proxy crops (normalized using price), "watervalue": rank over value of irrigation water; and fullpotentail TRUE/FALSE separated by ":" (TRUE: Full irrigation potential (cell receives full irrigation requirements in total area). FALSE: reduced potential of cell receives at later stage in allocation algorithm)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm: TRUE (default): monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{potential_wat}{if TRUE: potential available water and areas used, if FALSE: currently reserved water on current irrigated cropland used}

\item{com_ag}{if TRUE: the currently already irrigated areas in initialization year are reserved for irrigation, if FALSE: no irrigation areas reserved (irrigation potential)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
plot minimum monetary yield gain achieved on irrigated area
}
\examples{
\dontrun{
plotMarginalReturnToIrrigationFull(y_axis_range = seq(0, 10000, by = 100), scenario = "ssp2")
}

}
\author{
Felicitas Beier
}
