% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterIrrigArea.R
\name{plotScatterIrrigArea}
\alias{plotScatterIrrigArea}
\title{plotScatterIrrigArea}
\usage{
plotScatterIrrigArea(
  region,
  scenario,
  lpjml,
  selectyears,
  climatetype,
  efrMethod,
  accessibilityrule,
  rankmethod,
  yieldcalib,
  allocationrule,
  gainthreshold,
  thresholdtype,
  irrigationsystem,
  landScen,
  cropmix,
  multicropping
)
}
\arguments{
\item{region}{GLO or one of the H12 regions}

\item{scenario}{EFP and non-agricultural water use scenarios separate by "." (e.g. "on.ssp2")}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{years for which irrigatable area is calculated}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Scalar value defining the strictness of accessibility restriction: discharge that is exceeded x percent of the time on average throughout a year (Qx). Default: 0.5 (Q50) (e.g. Q75: 0.25, Q50: 0.5)}

\item{rankmethod}{method of calculating the rank: "meancellrank" (default): mean over cellrank of proxy crops, "meancroprank": rank over mean of proxy crops (normalized), "meanpricedcroprank": rank over mean of proxy crops (normalized using price), "watervalue": rank over value of irrigation water; and fullpotentail TRUE/FALSE separated by ":" (TRUE: Full irrigation potential (cell receives full irrigation requirements in total area). FALSE: reduced potential of cell receives at later stage in allocation algorithm)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{gainthreshold}{Gainthreshold for allocation algorithm}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm: TRUE (default): monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{landScen}{Land availability scenario (currCropland, currIrrig, potCropland)
combination of land availability scenario and initialization year separated by ":".
Initialization year only relevant for curr scenarios.
protection scenario separated by "_" (only relevant when potCropland selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
scatterplot of projected irrigated areas on current cropland and actually irrigated areas reported by LUH
}
\examples{
\dontrun{
plotScatterIrrigArea()
}

}
\author{
Felicitas Beier
}
