% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAfforestationMask.R
\name{readAfforestationMask}
\alias{readAfforestationMask}
\title{readAff_noboreal}
\usage{
readAfforestationMask(subtype)
}
\arguments{
\item{subtype}{afforestation mask sub type}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Afforestation No Boreal dataset
}
\examples{

\dontrun{
  readSource("AfforestationMask", subtype="noboreal", convert="onlycorrect")
}

}
\author{
David Chen
}
