% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAtmosphericDepositionRates.R
\name{readAtmosphericDepositionRates}
\alias{readAtmosphericDepositionRates}
\title{readAtmosphericDepositionRates}
\usage{
readAtmosphericDepositionRates()
}
\value{
Magpie objects with results on cellular level.
}
\description{
Read Atmospheric deposition rate
}
\examples{

\dontrun{
  readSource("AtmosphericDepositionRates", convert="onlycorrect")
}

}
\author{
Marcos Alves
}
