% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAvlIrrig.R
\name{readAvlIrrig}
\alias{readAvlIrrig}
\title{readAvlIrrig}
\usage{
readAvlIrrig()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Available Irrigation
}
\examples{

\dontrun{
  readSource("AvlIrrig", convert="onlycorrect")
}

}
\author{
David Chen
}
