% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAvlLandSi.R
\name{readAvlLandSi}
\alias{readAvlLandSi}
\title{readAvl_Land_Si}
\usage{
readAvlLandSi()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read si0 and nsi0 areas based on Ramankutty dataset"
}
\examples{

\dontrun{
  readSource("AvlLandSi", convert="onlycorrect")
}

}
\author{
David Chen
}
