% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBphEffect.R
\name{readBphEffect}
\alias{readBphEffect}
\title{readBphEffect}
\usage{
readBphEffect()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Biogeophysical temperature change of afforestation (degree C). File is based on observation datasets of Bright et al. 2017 and Duveiller et al. 2018
}
\examples{

\dontrun{
  readSource("BphEffect", convert="onlycorrect")
}

}
\author{
Michael Windisch
}
