% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBphMask.R
\name{readBphMask}
\alias{readBphMask}
\title{readBphMask}
\usage{
readBphMask()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Mask of Datapoints of biogeophysical temperature change of afforestation (degree C) to be used as weight. File is based on observation datasets of Bright et al. 2017 and Duveiller et al. 2018
}
\examples{

\dontrun{
  readSource("readBphMask", convert="onlycorrect")
}

}
\author{
Michael Windisch
}
