% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBphTCRE.R
\name{readBphTCRE}
\alias{readBphTCRE}
\title{readBphTCRE}
\usage{
readBphTCRE()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Transient Climate Response to accumulated doubling of CO2. File based on CMIP5 +1perc CO2 per year experiment. To be used in the translation to carbon equivalents of BphEffect
}
\examples{

\dontrun{
  readSource("BphTCRE", convert="onlycorrect")
}

}
\author{
Michael Windisch
}
