% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCalibratedArea.R
\name{readCalibratedArea}
\alias{readCalibratedArea}
\title{readCalibratedArea}
\usage{
readCalibratedArea()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Calibrated Area
}
\examples{

\dontrun{
  readSource("CalibratedArea", convert="onlycorrect")
}

}
\author{
David Chen
}
