% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCO2Atmosphere.R
\name{readCO2Atmosphere}
\alias{readCO2Atmosphere}
\title{readCO2Atmosphere}
\usage{
readCO2Atmosphere(subtype = "rcp8p5")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
Magpie objects with results on global level
}
\description{
Read CO2 global atmospheric concentration
}
\examples{

\dontrun{
readSource("CO2Atmosphere", subtype="rcp8p5", convert=FALSE)
}

}
\seealso{
\code{\link{readCO2Atmosphere}}
}
\author{
Marcos Alves, Kristine Karstens
}
