% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCshareReleased.R
\name{readCshareReleased}
\alias{readCshareReleased}
\title{readCshareReleased}
\usage{
readCshareReleased()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read "Share of soil carbon that is released on cropland during the harvest. Calculated by Benjamin Bodirsky"
}
\examples{

\dontrun{
  readSource("CshareReleased", convert="onlycorrect")
}

}
\author{
David Chen
}
