% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDams.R
\name{readDams}
\alias{readDams}
\title{readDams}
\usage{
readDams()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read dam file (no source information available)
}
\examples{

\dontrun{
  readSource("Dams", convert="onlycorrect")
}

}
\author{
Kristine Karstens
}
