% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGCMClimate.R
\name{readGCMClimate}
\alias{readGCMClimate}
\title{readGCMClimate}
\usage{
readGCMClimate(subtype = "GCMClimate:rcp85:HadGEM2.temperature")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Climate data used as LPJmL inputs into MAgPIE objects
}
\examples{

\dontrun{
readSource("GCMClimate", subtype="HadGEM2:rcp85.temperature", convert="onlycorrect")
}

}
\seealso{
\code{\link{readGCMClimate}}
}
\author{
Marcos Alves, Kristine Karstens
}
