% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFAD.R
\name{readGFAD}
\alias{readGFAD}
\title{readGFAD}
\usage{
readGFAD()
}
\value{
magpie object in cellular resolution
}
\description{
Read GLobal Forest Age Dataset derived from MODIS and COPENICUS satellite data
}
\examples{

\dontrun{
  readSource("GFAD", convert="onlycorrect")
}
}
\author{
Abhijeet Mishra, Felicitas Beier
}
