% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFADArea.R
\name{readGFADArea}
\alias{readGFADArea}
\title{readGFADArea}
\usage{
readGFADArea()
}
\value{
magpie object in cellular resolution
}
\description{
Read GLobal Forest Age Dataset derived from MODIS and COPENICUS satellite data
}
\examples{

\dontrun{
  readSource("GFADArea", convert="onlycorrect")
}
}
\author{
Abhijeet Mishra, Felicitas Beier
}
