% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGLW3.R
\name{readGLW3}
\alias{readGLW3}
\title{readGLW3}
\usage{
readGLW3(subtype = "Da")
}
\arguments{
\item{subtype}{Subtype of file to be opened (either Da or Aw)}
}
\value{
Magpie objects
}
\description{
Read the gridded livestock of the world 3 dataset.
}
\examples{

\dontrun{
readSource("GLW3", subtype = "DA", convert="onlycorrect")
}

}
\author{
Marcos Alves
}
