% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPD.R
\name{readGPD}
\alias{readGPD}
\title{readGPD}
\usage{
readGPD()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
read GPD
Data from the Global Peatland Database provided by Alexandra Barthelmes.
The original xls file has been clean-up manually (country names). Turkey had two identical entries in the original xls file.
Sources:
"Inventory Reports and National Communications UNFCC 2014",
"soil and peatland science",
"European Mires Book" ,
"own estimates (incl. GIS data)",
}
\examples{

\dontrun{
  readSource("GPD", convert="onlycorrect")
}
}
\author{
Florian Humpenoeder
}
