% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGrassYldEmu.R
\name{readGrassYldEmu}
\alias{readGrassYldEmu}
\title{readGrassYldEmu}
\usage{
readGrassYldEmu(subtype = "109325f71e.inputs")
}
\arguments{
\item{subtype}{Subtype of file to be opened. Subtypes available:
'max_harvest', 'weights', 'inputs',  'stddevs' and 'means'.}
}
\value{
Magpie objects with a diverse inforamtion
}
\description{
Read files related to the training and optimization of the LPJml emulators.
}
\examples{

\dontrun{
readSource("GrassYldEmu", subtype = "GrassYldEmu:20f33a2280.weights", convert="onlycorrect")
}

}
\author{
Marcos Alves
}
