% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readISIMIP.R
\name{readISIMIP}
\alias{readISIMIP}
\title{readISIMIP}
\usage{
readISIMIP(subtype = "water_abstraction.past")
}
\arguments{
\item{subtype}{Data source to be read from and time frame (past, future) to be read in, separated by "."}
}
\value{
MAgPIE object of non-agricultural water demand at 0.5 cellular level in mio. m^3
}
\description{
Read in non-agricultural water demand data from ISIMIP3b input data
}
\examples{
\dontrun{ readSource("ISIMIP", subtype="water_abstraction.past", convert="onlycorrect") }

}
\author{
Felicitas Beier
}
