% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readISIMIPinputs.R
\name{readISIMIPinputs}
\alias{readISIMIPinputs}
\title{readISIMIPinputs}
\usage{
readISIMIPinputs(subtype = "ISIMIP3b:water:histsoc.waterabstraction")
}
\arguments{
\item{subtype}{Data source to be read from including path separated by ":", subtype separated by "."}
}
\value{
MAgPIE object of non-agricultural water demand at 0.5 cellular level in mio. m^3
}
\description{
Read in non-agricultural water demand data from ISIMIP inputs
}
\examples{
\dontrun{
readSource("ISIMIPinputs",
subtype="ISIMIP3b:water:histsoc.waterabstraction",
convert="onlycorrect")
}

}
\author{
Felicitas Beier
}
