% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKoeppen_geiger.R
\name{readKoeppen_geiger}
\alias{readKoeppen_geiger}
\title{readKoeppen_geiger}
\usage{
readKoeppen_geiger()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read koeppen geiger areas cellular
}
\examples{

\dontrun{
  readSource("Koeppen_geiger", convert="onlycorrect")
}

}
\author{
David Chen
}
