% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLeifeld2018.R
\name{readLeifeld2018}
\alias{readLeifeld2018}
\title{readLeifeld2018}
\usage{
readLeifeld2018()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
read potential peatland area from Leifeld2018
}
\examples{

\dontrun{
  readSource("Leifeld2018", convert="onlycorrect")
}
}
\author{
Florian Humpenoeder
}
