% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLuh2SideLayers.R
\name{readLuh2SideLayers}
\alias{readLuh2SideLayers}
\title{readLuh2SideLayers}
\usage{
readLuh2SideLayers()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read LUH2 land cover data of biodiversity loss
}
\examples{

\dontrun{
  readSource("Luh2SideLayers", convert="onlycorrect")
}

}
\author{
Michael Windisch
}
