% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLUH2UrbanFuture.R
\name{readLUH2UrbanFuture}
\alias{readLUH2UrbanFuture}
\title{readLUH2UrbanFuture}
\usage{
readLUH2UrbanFuture()
}
\value{
magpie object of gridded future urban land use in Mha, 2015-2100
}
\description{
read in gridded future urban land use datasets
}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
