% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNeighborCells.R
\name{readNeighborCells}
\alias{readNeighborCells}
\title{readNeighborCells}
\usage{
readNeighborCells()
}
\value{
empty magpie object with list data stored in attributes
}
\description{
read file with neighbor cell and distance information
             derived from LPJmL grid
}
\examples{
\dontrun{
readSource("NeighborCells", convert = FALSE)
}

}
\author{
Felicitas Beier
}
