% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNitrogenFixationRateNatural.R
\name{readNitrogenFixationRateNatural}
\alias{readNitrogenFixationRateNatural}
\title{NitrogenFixationRateNatural}
\usage{
readNitrogenFixationRateNatural()
}
\value{
Magpie objects with results on cellular level.
}
\description{
Read nitrogen fixation rate
}
\examples{

\dontrun{
  readSource("NitrogenFixationRateNatural", convert="onlycorrect")
}

}
\author{
Marcos Alves
}
