% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNpiNdcAdAolcPol.R
\name{readNpiNdcAdAolcPol}
\alias{readNpiNdcAdAolcPol}
\title{readNpiNdcAdAolcPol}
\usage{
readNpiNdcAdAolcPol()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read NPI/NDC dataset
}
\examples{

\dontrun{
  readSource("NpiNdcAdAolcPol", convert="onlycorrect")
}

}
\author{
Michael Windisch
}
