% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNpiNdcAffPol.R
\name{readNpiNdcAffPol}
\alias{readNpiNdcAffPol}
\title{readNpiNdcAffPol}
\usage{
readNpiNdcAffPol()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read dummy NPI/NDC policies
}
\examples{

\dontrun{
  readSource("NpiNdcAffPol", convert="onlycorrect")
}

}
\author{
Michael Windisch
}
