% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProtectArea.R
\name{readProtectArea}
\alias{readProtectArea}
\title{readProtectArea}
\usage{
readProtectArea()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read Protected Area
}
\examples{

\dontrun{
  readSource("ProtectArea", convert="onlycorrect")
}

}
\author{
David Chen
}
