% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRrLayer.R
\name{readRrLayer}
\alias{readRrLayer}
\title{readRrLayer}
\usage{
readRrLayer()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read range-rarity layer
}
\examples{

\dontrun{
  readSource("RrLayer", convert="onlycorrect")
}

}
\author{
Michael Windisch
}
