% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSoilClassification.R
\name{readSoilClassification}
\alias{readSoilClassification}
\title{readSoilClassification}
\usage{
readSoilClassification(subtype = "HWSD.soil")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
Magpie object with results on cellular level for soil types
}
\description{
Read soil classification data used as input for lpjml
}
\examples{

\dontrun{
readSource("SoilClassification", subtype="HWSD.soil", convert="onlycorrect")
}

}
\author{
Marcos Alves, Kristine Karstens
}
