% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTransport_Distance.R
\name{readTransport_Distance}
\alias{readTransport_Distance}
\title{readTransport_Distance}
\usage{
readTransport_Distance()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read transport distance (no source information available)
}
\examples{

\dontrun{
  readSource("Transport_Distance", convert="onlycorrect")
}

}
\author{
David Chen
}
