% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTransportDistance.R
\name{readTransportDistance}
\alias{readTransportDistance}
\title{readTransportDistance}
\usage{
readTransportDistance()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read transport distance (no source information available)
}
\examples{

\dontrun{
  readSource("TransportDistance", convert="onlycorrect")
}

}
\author{
David Chen
}
