% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUrbanLandFuture.R
\name{readUrbanLandFuture}
\alias{readUrbanLandFuture}
\title{readUrbanLandFuture}
\usage{
readUrbanLandFuture(subtype)
}
\arguments{
\item{subtype}{Currently only from LUH2v2 for all SSPs, LUH2v2 uses IMAGE}
}
\value{
magpie object of gridded future urban land use in Mha, 2015-2100
}
\description{
read in gridded future urban land use datasets
}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
