% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWATERGAP.R
\name{readWATERGAP}
\alias{readWATERGAP}
\title{readWATERGAP}
\usage{
readWATERGAP(subtype = "WATERGAP2013")
}
\arguments{
\item{subtype}{Data source to be read from}
}
\value{
MAgPIE object of non-agricultural water demand at 0.5 cellular level in mio. m^3
}
\description{
Read in non-agricultural water demand data from WATERGAP model with A2: WATERGAP WATCH project; B1: WATERGAP WATCH project; SSP2: WATERGAP ISIMIP project
}
\examples{
\dontrun{ readSource("WATERGAP", convert="onlycorrect") }

}
\author{
Felicitas Beier, Abhijeet Mishra
}
