% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEconOfIrrig.R
\name{reportEconOfIrrig}
\alias{reportEconOfIrrig}
\title{reportEconOfIrrig}
\usage{
reportEconOfIrrig(
  region = "GLO",
  output,
  gtRange,
  scenario,
  lpjml,
  iniyear,
  selectyears,
  climatetype,
  efrMethod,
  accessibilityrule,
  rankmethod,
  yieldcalib,
  allocationrule,
  irrigationsystem,
  landScen,
  cropmix,
  potentialWat = TRUE,
  comAg,
  multicropping,
  transDist
)
}
\arguments{
\item{region}{regional resolution (can be country iso-code,
region name and respective mapping "EUR:H12", "GLO" for global)}

\item{output}{output to be displayed: irrigated area "IrrigArea" or
available water volume "wat_ag_ww" "wat_ag_wc"}

\item{gtRange}{range of x-axis (gainthreshold) to be depicted on the curve}

\item{scenario}{non-agricultural water use scenario to be displayed in plot}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{iniyear}{initialization year}

\item{selectyears}{years for which irrigatable area is calculated}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Method used: Quantile method (Q) or
Coefficient of Variation (CV)
combined with scalar value defining the strictness
of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year
(Qx, e.g. Q75: 0.25, Q50: 0.5)
or base value for exponential curve separated by : (CV:2)}

\item{rankmethod}{Rank and optimization method consisting of
Unit according to which rank is calculated:
USD_ha (USD per hectare) for relative area return, or
USD_m3 (USD per cubic meter) for relative volumetric return;
USD for absolute return (total profit);
USD_m3ha (USD per hectare per cubic meter)
for relative return according to area and volume.
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices
and boolean indicating fullpotential (TRUE, i.e. cell
receives full irrigation requirements in total area)
or reduced potential (FALSE, reduced potential of cell
receives at later stage in allocation algorithm);
separated by ":"}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
     Also needs specification of refYields, separated by ":".
     Options: FALSE (for single cropping analyses) or
              "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst")}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{potentialWat}{if TRUE: potential available water and areas used,
if FALSE: currently reserved water on current irrigated cropland used}

\item{comAg}{if TRUE: the currently already irrigated areas
         in initialization year are reserved for irrigation,
if FALSE: no irrigation areas reserved (irrigation potential)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
               by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; TRUE:exogenous; FALSE)}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}
}
\value{
magpie object in cellular resolution
}
\description{
reports potentially irrigated area depending on gainthreshold,
             land constraint and water constraint
}
\examples{
\dontrun{
reportEconOfIrrig(gtRange = seq(0, 10000, by = 100), scenario = "ssp2")
}

}
\author{
Felicitas Beier
}
