% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAllocation.R
\name{toolAllocation}
\alias{toolAllocation}
\title{toolAllocation}
\usage{
toolAllocation(y, rs, l_inout, l_in, allocationrule, meancellrank)
}
\arguments{
\item{y}{Current year of loop}

\item{rs}{River structure (list of river structure details and cell numbers including ordered cell number)}

\item{l_inout}{list of inputs that are at the same time outputs: required_wat_min_allocation Minimum (water requirement reserved per grid cell (as magpie object of correct dimension)); discharge (Discharge to be allocated (as magpie object of correct dimension)); frac_fullirrig (fraction of fullirrigation requirements that can be fulfilled)}

\item{l_in}{list of inputs: irrig_yieldgainpotential (yield gain potential through irrigation of proxy crops: magpie object with cellular and year dimension (as magpie object of correct dimension)); required_wat_fullirrig_ww (required withdrawal for full irrigation in specific cell (as magpie object of correct dimension)); required_wat_fullirrig_wc (required consumption for full irrigation in specific cell (as magpie object of correct dimension)); gainthreshold (Minimum yield gain in USD/ha (as scalar value)); avl_wat_ww; avl_wat_wc}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{meancellrank}{cell ranking for different years (array). Note: only applicable when allocationrule "optimization" chosen}
}
\value{
magpie object in cellular resolution
}
\description{
This tool function executes the Allocation Algorithm in cell order of selected allocation rule
}
\author{
Felicitas Beier, Jan Philipp Dietrich
}
