% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolApplyRegionNames.R
\name{toolApplyRegionNames}
\alias{toolApplyRegionNames}
\title{Apply region names}
\usage{
toolApplyRegionNames(cdata, regionscode)
}
\arguments{
\item{cdata}{a cluster data file as produced by cluster_base}

\item{regionscode}{regionscode of the regional mapping to be used.
Must agree with the regionscode of the mapping mentioned in the madrat
config! Can be retrieved via \code{regionscode()}.}
}
\value{
the cluster data file with region names in spatial dimension
rather than country names
}
\description{
This tool function replaces country names with region names
in the spatial dimension of the object. To avoid mixing up
of cache files with different regional aggregation the
regioncode needs to supplied and checked as well.
Only if the supplied regions code agrees with the
region mapping currently chosen the function will return
the data.
}
\seealso{
\code{\link{calcClusterKMeans}}, \code{\link{calcClusterBase}}
}
\author{
Jan Philipp Dietrich
}
