% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolBWCregression.R
\name{toolBWCregression}
\alias{toolBWCregression}
\title{toolBWCregression}
\usage{
toolBWCregression(y, x)
}
\arguments{
\item{y}{Dependent variable (y) for blue water consumption regression
(crop- and system-specific crop blue water consumption for given system in main season)
as magpie object that is already filtered}

\item{x}{Independent variable (x) for blue water consumption regression
(crop- and system-specific grass blue water consumption in irrigated growing period of respective crop)
as magpie object that is already filtered}
}
\value{
list of regression coefficients and summary statistics
}
\description{
This function calculates the functional relationship between
             crop consumptive blue water use for and grass consumptive blue water use
             using a linear fit and returns the coefficients and summary statistics as a list
}
\examples{
\dontrun{
calcOutput("BlueWaterConsumptionOff", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
