% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolClusterPerRegion.R
\name{toolClusterPerRegion}
\alias{toolClusterPerRegion}
\title{Cluster per Region}
\usage{
toolClusterPerRegion(cells, ncluster, weight = NULL)
}
\arguments{
\item{cells}{spatial names as returned by \code{getCells}}

\item{ncluster}{The desired total number of clusters.}

\item{weight}{named vector with weighting factors for each region for the cluster distribution
,e.g. weight=c(AFR=3,EUR=0.5). weight > 1 will grant more cluster to a region and
weight < 1 less cluster than by default.}
}
\value{
A matrix with regions in rows and number of cells and clusters in
columns
}
\description{
This function calculates an appropriate number of clusters per region as it
is needed for ClusterKMeans
}
\seealso{
\code{\link{calcClusterKMeans}}, \code{\link{calcClusterBase}}
}
\author{
Jan Philipp Dietrich
}
