% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolClusterPerRegionManual.R
\name{toolClusterPerRegionManual}
\alias{toolClusterPerRegionManual}
\title{Cluster per Region set manually}
\usage{
toolClusterPerRegionManual(cells, ncluster, ncluster2reg)
}
\arguments{
\item{cells}{spatial names as returned by \code{getCells}}

\item{ncluster}{The desired total number of clusters.}

\item{ncluster2reg}{named vector with numbers per region}
}
\value{
A matrix with regions in rows and number of cells and clusters in
columns
}
\description{
This function translates weights into number of clusters per region as it
is needed for ClusterKMeans. Weights have to sum up to total number of clusters.
}
\seealso{
\code{\link{calcClusterKMeans}}, \code{\link{calcClusterBase}}
}
\author{
Kristine Karstens
}
