% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLPJarrayToMAgPIEmap.R
\name{toolLPJarrayToMAgPIEmap}
\alias{toolLPJarrayToMAgPIEmap}
\title{toolLPJarrayToMAgPIEmap}
\usage{
toolLPJarrayToMAgPIEmap(x)
}
\value{
magpie object with correct dimension for use in plotmap
}
\description{
transforms object with 67420 cells to be able to be used in plotmap function

This tool function selects the required cells from object with
67420 cells to be able to plot using plotmap or plotmap2 that
are applied to magpie object of size 59199 cells
}
\author{
Felicitas Beier
}
