% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLPJcell2MAgPIEcell.R
\name{toolLPJcell2MAgPIEcell}
\alias{toolLPJcell2MAgPIEcell}
\title{toolLPJcell2MAgPIEcell}
\usage{
toolLPJcell2MAgPIEcell(x)
}
\arguments{
\item{x}{magpie object to be transformed}
}
\value{
magpie object with correct dimension for use in plotmap
}
\description{
transforms magpie object with 67420 cells to object with 59199 cells

This tool function selects the required cells from object with
67420 cells to be able to plot using plotmap or plotmap2 that
are applied to magpie object of size 59199 cells
}
\author{
Felicitas Beier
}
