% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLPJcellCoordinates.R
\name{toolLPJcellCoordinates}
\alias{toolLPJcellCoordinates}
\title{toolLPJcellCoordinates}
\usage{
toolLPJcellCoordinates(x, type)
}
\arguments{
\item{x}{magpie object for which coordinatese should be remapped}

\item{type}{direction of transformation (coord2lpj or lpj2coord)}
}
\value{
magpie object with coordinates in LPJ cell ordering
}
\description{
maps LPJ cells and coordinates
}
\author{
Felicitas Beier
}
