% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolNeighborWaterProvision.R
\name{toolNeighborWaterProvision}
\alias{toolNeighborWaterProvision}
\title{toolNeighborWaterProvision}
\usage{
toolNeighborWaterProvision(transDist, c, rs, missingWW, missingWC, inoutLIST)
}
\arguments{
\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand}

\item{c}{Current local cell that requires additional water from
surrounding cells}

\item{rs}{River structure list including neighboring cells}

\item{missingWW}{Volume of water withdrawal
that needs to come from surrounding cells}

\item{missingWC}{Volume of water consumption
that needs to come from surrounding cells}

\item{inoutLIST}{List of objects that are updated
(discharge, watReserved, missingWW, missingWC)}
}
\value{
magpie object in cellular resolution
}
\description{
This tool function allows to draw water from surrounding cells
             if local water availability is not sufficient to fulfill demand
}
\author{
Felicitas Beier, Jens Heinke
}
