% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolNeuralNet.R
\name{toolNeuralNet}
\alias{toolNeuralNet}
\title{Neural Network Reconstruction}
\usage{
toolNeuralNet(inputs_ml, weights)
}
\arguments{
\item{inputs_ml}{Neural Network input features properly scaled with
the scale and center attributes of the scaled training set.}

\item{weights}{The learned weights and biases in a list format as outputed by the function \code{keras::get_weights()}. This function works with magpie objects}
}
\value{
The evaluated result of the neural network for the \code{input_ml} parameter.
}
\description{
Reconstructs and evaluate a neural network from the weights and biases provided as arguments
}
\author{
Marcos Alves
}
