% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRegionSums.R
\name{toolRegionSums}
\alias{toolRegionSums}
\title{toolRegionSums}
\usage{
toolRegionSums(x, region)
}
\arguments{
\item{x}{magpie object to be summed up over first dimension}

\item{region}{regional resolution (can be country iso-code, "GLO" for global,
or region name and respective mapping "EUR:H12")}
}
\value{
magpie object with regional sum
}
\description{
sum over regional aggregation chosen in region argument
}
\author{
Felicitas Beier
}
