% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRiverDischargeAllocation.R
\name{toolRiverDischargeAllocation}
\alias{toolRiverDischargeAllocation}
\title{toolRiverDischargeAllocation}
\usage{
toolRiverDischargeAllocation(
  rs,
  c,
  downCells,
  neighborSelectCells = NULL,
  iteration,
  transDist,
  inLIST,
  inoutLIST
)
}
\arguments{
\item{rs}{River structure with information on upstreamcells,
downstreamcells and neighboring cells and distances}

\item{c}{Current cell for which water shall be allocated}

\item{downCells}{Downstream cells of c}

\item{neighborSelectCells}{Optionally: precomputed named list of neighbor-cell
downstream cells (only required for main allocation
with transDist != 0)}

\item{iteration}{Currently active iteration of river discharge allocation.
Arguments:
"main" for case of main river cells
"neighbor" for case of neighboring cells of main river cells}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}

\item{inLIST}{List of objects that are inputs to the function
irrigGain, gainthreshold,}

\item{inoutLIST}{List of objects that are inputs to the function and
are updated by the function}
}
\value{
magpie object in cellular resolution
}
\description{
This tool function allocates discharge
             for grid cells respecting upstream-downstream relationships
             and various water constraints
}
\author{
Felicitas Beier, Jens Heinke, Jan Philipp Dietrich
}
