% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRiverDischargeUpdate.R
\name{toolRiverDischargeUpdate}
\alias{toolRiverDischargeUpdate}
\title{toolRiverDischargeUpdate}
\usage{
toolRiverDischargeUpdate(rs, runoffWOEvap, watCons, cellsCalc = NULL)
}
\arguments{
\item{rs}{River structure}

\item{runoffWOEvap}{Array that contains (runoff - lake evap)}

\item{watCons}{Array that contains water reserved
for consumptive use}

\item{cellsCalc}{Optional integer vector defining the upstream-to-downstream
cell calculation order}
}
\value{
array in cellular resolution and all year and scenario dimensions
}
\description{
This function calculates cellular discharge
             after reserving water uses for consumption
}
\examples{
\dontrun{
calcOutput("RiverHumanUseAccounting", aggregate = FALSE)
}
}
\author{
Felicitas Beier, Jens Heinke
}
