% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRiverUpDownBalance.R
\name{toolRiverUpDownBalance}
\alias{toolRiverUpDownBalance}
\title{toolRiverUpDownBalance}
\usage{
toolRiverUpDownBalance(inLIST, inoutLIST)
}
\arguments{
\item{inLIST}{List of objects that are inputs to the function:
previously reserved withdrawals and consumption in current cell;
currently requested withdrawal in current cell}

\item{inoutLIST}{List of objects that are inputs to the function and
are updated by the function:
discharge (including up- and downstream cells)
currently requested consumption (including upstream cells)}
}
\value{
list of arrays objects in cellular resolution
}
\description{
This function calculates the cell water balance
             under consideration of different reserved human uses
             (non-agricultural, neighbor water requirements,
              committed-agricultural uses)
}
\author{
Felicitas Beier, Jens Heinke
}
