% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSelectNeighborCell.R
\name{toolSelectNeighborCell}
\alias{toolSelectNeighborCell}
\title{toolSelectNeighborCell}
\usage{
toolSelectNeighborCell(transDist)
}
\arguments{
\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand}
}
\value{
magpie object in cellular resolution
}
\description{
Selects cells in certain radius of current cell
}
\author{
Felicitas Beier, Jens Heinke
}
