% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSelectRiverBasin.R
\name{toolSelectRiverBasin}
\alias{toolSelectRiverBasin}
\title{toolSelectRiverBasin}
\usage{
toolSelectRiverBasin(basinname)
}
\arguments{
\item{basinname}{river basin name as listed in mapping file RiverBasinMapping.csv}
}
\value{
list of cell names
}
\description{
returns coordinates of cells that belong to chosen basin
}
\author{
Felicitas Beier
}
