% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolTurnBackslash.R
\name{toolTurnBackslash}
\alias{toolTurnBackslash}
\title{toolTurnBackslash}
\usage{
toolTurnBackslash(x)
}
\value{
string
}
\description{
turns around backslashes in a string
}
\author{
Felicitas Beier
}
