% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcActualIrrigWatRequirements.R
\name{calcActualIrrigWatRequirements}
\alias{calcActualIrrigWatRequirements}
\title{calcActualIrrigWatRequirements}
\usage{
calcActualIrrigWatRequirements(
  selectyears = "all",
  climatetype = "GSWP3-W5E5:historical",
  iniyear = 1995
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{iniyear}{Initialization year (for weight by cropland)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates actual irrigation water requirements per cell given a certain irrigation system
}
\examples{
\dontrun{ calcOutput("ActualIrrigWatRequirements", aggregate=FALSE) }

}
\seealso{
\code{\link{calcIrrigationSystem}}, \code{\link{calcIrrigWatRequirements}}
}
\author{
Felicitas Beier
}
