% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcActualIrrigWatRequirements.R
\name{calcActualIrrigWatRequirements}
\alias{calcActualIrrigWatRequirements}
\title{calcActualIrrigWatRequirements}
\usage{
calcActualIrrigWatRequirements(
  selectyears,
  iniyear,
  lpjml,
  climatetype,
  irrigationsystem,
  multicropping
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year (for weight by cropland)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; FALSE:NULL)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates actual irrigation water requirements
             per cell given the chosen irrigation system
}
\examples{
\dontrun{
calcOutput("ActualIrrigWatRequirements", aggregate = FALSE)
}

}
\seealso{
\code{\link{calcIrrigationSystem}}, \code{\link{calcIrrigWatRequirements}}
}
\author{
Felicitas Beier
}
