% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcActualIrrigWatRequirements.R
\name{calcActualIrrigWatRequirements}
\alias{calcActualIrrigWatRequirements}
\title{calcActualIrrigWatRequirements}
\usage{
calcActualIrrigWatRequirements(lpjml, selectyears, climatetype, iniyear)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{Years to be returned}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{iniyear}{Initialization year (for weight by cropland)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates actual irrigation water requirements
             per cell given a certain irrigation system
}
\examples{
\dontrun{
calcOutput("ActualIrrigWatRequirements", aggregate = FALSE)
}

}
\seealso{
\code{\link{calcIrrigationSystem}}, \code{\link{calcIrrigWatRequirements}}
}
\author{
Felicitas Beier
}
