% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcActualIrrigWatRequirements.R
\name{calcActualIrrigWatRequirements}
\alias{calcActualIrrigWatRequirements}
\title{calcActualIrrigWatRequirements}
\usage{
calcActualIrrigWatRequirements(
  selectyears = "all",
  iniyear = 1995,
  crops = "magpie",
  version = "LPJmL5",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  harmonize_baseline = FALSE,
  ref_year = NULL
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year (for weight by cropland)}

\item{crops}{Selects "magpie" (default) or "lpjml" crops}

\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates actual irrigation water requirements per cell given a certain irrigation system
}
\examples{
\dontrun{ calcOutput("ActualIrrigWatRequirements", aggregate=FALSE) }

}
\seealso{
\code{\link{calcIrrigationSystem}}, \code{\link{calcIrrigWatRequirements}}
}
\author{
Felicitas Beier
}
