% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaPotIrrig.R
\name{calcAreaPotIrrig}
\alias{calcAreaPotIrrig}
\title{calcAreaPotIrrig}
\usage{
calcAreaPotIrrig(selectyears, comagyear, landScen)
}
\arguments{
\item{selectyears}{years to be returned}

\item{comagyear}{if NULL: total potential croparea is used;
if !NULL: already irrigated area is subtracted;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea (e.g. NULL, 1995, 2010)}

\item{landScen}{Land availability scenario (currCropland, currIrrig, potCropland)
combination of land availability scenario and initialization year separated by ":".
Initialization year only relevant for curr scenarios.
protection scenario separated by "_" (only relevant when potCropland selected):
WDPA, BH, FF, CPD, LW, HalfEarth}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates land that is potentially available
             for irrigated agriculture
}
\examples{
\dontrun{
calcOutput("AreapotCropland", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
