% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaPotIrrig.R
\name{calcAreaPotIrrig}
\alias{calcAreaPotIrrig}
\title{calcAreaPotIrrig}
\usage{
calcAreaPotIrrig(selectyears, comagyear, avlland_scen)
}
\arguments{
\item{selectyears}{years to be returned}

\item{comagyear}{if NULL: total potential croparea is used;
if !NULL: already irrigated area is subtracted;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea (e.g. NULL, 1995, 2010)}

\item{avlland_scen}{Land availability scenario (currCropland, currIrrig, potIrrig)
combination of land availability scenario and initialization year separated by ":".
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates area that can potentially be used for
             irrigation given assumptions defined in arguments
}
\examples{
\dontrun{
calcOutput("AreaPotIrrig", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
