% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaPotIrrig.R
\name{calcAreaPotIrrig}
\alias{calcAreaPotIrrig}
\title{calcAreaPotIrrig}
\usage{
calcAreaPotIrrig(selectyears, comagyear, iniyear, landScen)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{comagyear}{If NULL: total potential croparea is used;
if !NULL: already irrigated area is subtracted;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea (e.g. NULL, 1995, 2010)}

\item{iniyear}{Initialization year for current cropland area}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_FF, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates land that is potentially available
             for irrigated agriculture
}
\examples{
\dontrun{
calcOutput("AreaPotIrrig", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
