% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaPotIrrig.R
\name{calcAreaPotIrrig}
\alias{calcAreaPotIrrig}
\title{calcAreaPotIrrig}
\usage{
calcAreaPotIrrig(selectyears, iniareayear, protect_scen)
}
\arguments{
\item{selectyears}{years to be returned}

\item{iniareayear}{if !NULL: already irrigated area is subtracted; if NULL: total potential land area is used; year specified here is the year of the initialization used for cropland area initialization in calcIrrigatedArea}

\item{protect_scen}{land protection scenario: NULL (no irrigation limitation in protected areas), WDPA, BH, FF, CPD, LW, HalfEarth. Areas where no irrigation water withdrawals are allowed due to biodiversity protection}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates area that can potentially be used for irrigation given assumptions defined in arguments
}
\examples{
\dontrun{ calcOutput("AreaPotIrrig", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
