% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaPotIrrig.R
\name{calcAreaPotIrrig}
\alias{calcAreaPotIrrig}
\title{calcAreaPotIrrig}
\usage{
calcAreaPotIrrig(
  selectyears,
  comagyear,
  iniyear,
  landScen,
  lpjml,
  climatetype,
  efrMethod,
  fossilGW,
  multicropping,
  transDist,
  scenarios = c("on.ISIMIP", "on.ssp1", "on.ssp2", "on.ssp3", "off.ISIMIP", "off.ssp1",
    "off.ssp2", "off.ssp3")
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{comagyear}{If NULL: total potential croparea is used;
if !NULL: already irrigated area is subtracted;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea (e.g. NULL, 1995, 2010)}

\item{iniyear}{Initialization year for current cropland area}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{lpjml}{If comagyear != NULL: LPJmL version used to calculate committed
agricultural use}

\item{climatetype}{If comagyear != NULL: climate scenario used to calculate committed
agricultural use
or historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{If comagyear != NULL: EFR method used to calculate committed
agricultural use (e.g., Smakhtin:good, VMF:fair)}

\item{fossilGW}{This argument is only relevant when comagyear != NULL
as it determines the area that is set aside as it is already
reserved by the committed agriculture iteration.
If TRUE: non-renewable groundwater can be used.
If FALSE: non-renewable groundwater cannot be used.}

\item{multicropping}{TRUE or FALSE (for committed agricultural area accounting)}

\item{transDist}{If comagyear != NULL: Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability
of committed agricultural uses}

\item{scenarios}{Scenarios for object dimension consisting of "EFP.scenario".
Default is c("on.ISIMIP",  "on.ssp1", "on.ssp2", "on.ssp3",
             "off.ISIMIP", "off.ssp1", "off.ssp2", "off.ssp3")
If scenarios change, they have to be adjusted in the default
setting of this function.}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates land that is potentially available
             for irrigated agriculture
}
\examples{
\dontrun{
calcOutput("AreaPotIrrig", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
