% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaPotIrrig.R
\name{calcAreaPotIrrig}
\alias{calcAreaPotIrrig}
\title{calcAreaPotIrrig}
\usage{
calcAreaPotIrrig(selectyears, iniyear, iniarea, protect_scen)
}
\arguments{
\item{selectyears}{years to be returned}

\item{iniyear}{year of initialization for cropland area initialization}

\item{iniarea}{if TRUE: already irrigated area is subtracted, if FALSE: total potential land area is used}

\item{protect_scen}{land protection scenario: NULL (no irrigation limitation in protected areas), WDPA, BH, FF, CPD, LW, HalfEarth. Areas where no irrigation water withdrawals are allowed due to biodiversity protection}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates area that can potentially be used for irrigation given assumptions defined in arguments
}
\examples{
\dontrun{ calcOutput("AreaPotIrrig", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
