% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaPotIrrig.R
\name{calcAreaPotIrrig}
\alias{calcAreaPotIrrig}
\title{calcAreaPotIrrig}
\usage{
calcAreaPotIrrig(
  selectyears,
  comagyear,
  iniyear,
  landScen,
  lpjml,
  climatetype,
  efrMethod,
  multicropping,
  transDist
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{comagyear}{If NULL: total potential croparea is used;
if !NULL: already irrigated area is subtracted;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea (e.g. NULL, 1995, 2010)}

\item{iniyear}{Initialization year for current cropland area}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{lpjml}{if comagyear != NULL: LPJmL version used to calculate committed
agricultural use}

\item{climatetype}{if comagyear != NULL: climate scenario used to calculate committed
agricultural use
or historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{if comagyear != NULL: EFR method used to calculate committed
agricultural use (e.g., Smakhtin:good, VMF:fair)}

\item{multicropping}{if comagyear != NULL: multicropping argument used to calculate committed
agricultural use.
Two components (separated by ":"):
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{transDist}{if comagyear != NULL: Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability
of committed agricultural uses}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates land that is potentially available
             for irrigated agriculture
}
\examples{
\dontrun{
calcOutput("AreaPotIrrig", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
