% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlWater.R
\name{calcAvlWater}
\alias{calcAvlWater}
\title{calcAvlWater}
\usage{
calcAvlWater(
  lpjml = "lpjml5.9.5-m1",
  climatetype = "MRI-ESM2-0:ssp370",
  stage = "harmonized2020",
  seasonality = "grper"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{stage}{Degree of processing: raw, smoothed, harmonized, harmonized2020}

\item{seasonality}{grper (default): water available in growing period per year;
total: total water available throughout the year;
monthly: monthly water availability (for further processing, e.g. in calcEnvmtlFlow)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water availability for MAgPIE retrieved from LPJmL
}
\examples{
\dontrun{
calcOutput("AvlWater", aggregate = FALSE)
}
}
\author{
Felicitas Beier, Kristine Karstens, Abhijeet Mishra
}
