% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlWater.R
\name{calcAvlWater}
\alias{calcAvlWater}
\title{calcAvlWater}
\usage{
calcAvlWater(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb"),
  climatetype = "GSWP3-W5E5:historical",
  cells = "magpiecell",
  stage = "harmonized2020",
  seasonality = "grper"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{cells}{Number of cells to be returned
(select "magpiecell" for 59199 cells or "lpjcell" for 67420 cells)}

\item{stage}{Degree of processing: raw, smoothed, harmonized, harmonized2020}

\item{seasonality}{grper (default): water available in growing period per year;
total: total water available throughout the year;
monthly: monthly water availability (for further processing, e.g. in calcEnvmtlFlow)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water availability for MAgPIE retrieved from LPJmL
}
\examples{
\dontrun{
calcOutput("AvlWater", aggregate = FALSE)
}
}
\author{
Felicitas Beier, Kristine Karstens, Abhijeet Mishra
}
