% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBlueWaterConsumption.R
\name{calcBlueWaterConsumption}
\alias{calcBlueWaterConsumption}
\title{calcBlueWaterConsumption}
\usage{
calcBlueWaterConsumption(
  selectyears,
  iniyear,
  lpjml,
  climatetype,
  fallowFactor = 0.75,
  areaMask,
  season
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year for filtering rules of data for regression
that determines second season yield}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Climate model (e.g., "MRI-ESM2-0:ssp370")
or historical baseline (e.g., "GSWP3-W5E5:historical")}

\item{fallowFactor}{Factor determining water requirement reduction in off season due to
fallow period between harvest of first (main) season and
sowing of second (off) season}

\item{areaMask}{Multicropping area mask to be used
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLandInG
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification}

\item{season}{Season to be returned: "main" (LPJmL growing period),
"year" (entire year))}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates consumptive blue water use for the whole year based on
             LPJmL blue water consumption of crops and the difference between rainfed and irrigated
             evapotranspiration of grass
}
\examples{
\dontrun{
calcOutput("BlueWaterConsumption", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
