% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBlueWaterConsumption.R
\name{calcBlueWaterConsumption}
\alias{calcBlueWaterConsumption}
\title{calcBlueWaterConsumption}
\usage{
calcBlueWaterConsumption(selectyears, lpjml, climatetype, multicropping)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates consumptive blue water use for the whole year based on
             LPJmL blue water consumption of crops and the difference between rainfed and irrigated
             evapotranspiration of grass
}
\examples{
\dontrun{
calcOutput("BlueWaterConsumption", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
