% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBlueWaterConsumption.R
\name{calcBlueWaterConsumption}
\alias{calcBlueWaterConsumption}
\title{calcBlueWaterConsumption}
\usage{
calcBlueWaterConsumption(
  selectyears,
  lpjml,
  climatetype,
  fallowFactor = 0.75,
  areaMask,
  output,
  proxycrops = "standard"
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{fallowFactor}{Factor determining water requirement reduction in off season due to
fallow period between harvest of first (main) season and
sowing of second (off) season}

\item{areaMask}{Multicropping area mask to be used
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification}

\item{output}{output to be returned by the function: combination of
crop type ("crops" or "grass") and
season ("main" (LPJmL growing period), "year" (entire year)),
separated by ":"
("crops:main", "crops:year", "grass:main", "grass:year")}

\item{proxycrops}{standard: FALSE
if TRUE: proxy crops of LPJmL for MAgPIE perennials selected}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates consumptive blue water use for the whole year based on
             LPJmL blue water consumption of crops and the difference between rainfed and irrigated
             evapotranspiration of grass
}
\examples{
\dontrun{
calcOutput("BlueWaterConsumption", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
