% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBlueWaterConsumption.R
\name{calcBlueWaterConsumption}
\alias{calcBlueWaterConsumption}
\title{calcBlueWaterConsumption}
\usage{
calcBlueWaterConsumption(
  selectyears,
  lpjml,
  climatetype,
  fallowFactor = 0.75,
  output
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{fallowFactor}{Factor determining water requirement reduction in off season due to
fallow period between harvest of first (main) season and
sowing of second (off) season}

\item{output}{output to be returned by the function: combination of
crop type ("crops" or "grass") and
season ("main" (LPJmL growing period), "year" (entire year)),
separated by ":"
("crops:main", "crops:year", "grass:main", "grass:year")}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates consumptive blue water use for the whole year based on
             LPJmL blue water consumption of crops and the difference between rainfed and irrigated
             evapotranspiration of grass
}
\examples{
\dontrun{
calcOutput("BlueWaterConsumption", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
