% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBlueWaterConsumptionOff.R
\name{calcBlueWaterConsumptionOff}
\alias{calcBlueWaterConsumptionOff}
\title{calcBlueWaterConsumptionOff}
\usage{
calcBlueWaterConsumptionOff(selectyears, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Climate model (e.g., "MRI-ESM2-0:ssp370")
or historical baseline (e.g., "GSWP3-W5E5:historical")}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates consumptive blue water use for the whole year based on
             LPJmL blue water consumption of crops and the difference between rainfed and irrigated
             evapotranspiration of grass
}
\examples{
\dontrun{
calcOutput("BlueWaterConsumptionOff", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
