% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCO2Atmosphere.R
\name{calcCO2Atmosphere}
\alias{calcCO2Atmosphere}
\title{calcCO2Atmosphere}
\usage{
calcCO2Atmosphere(rcp, level = "cellular")
}
\arguments{
\item{rcp}{specify the RCP (rcp8p5, rcp6p0, rcp4p5 or rcp2p6)}

\item{level}{specify the spatial output level}
}
\value{
magpie object in cellular resolution
}
\description{
Disaggregate CO2 global atmospheric concentration to cellular level
}
\examples{
\dontrun{ calcOutput("CO2Atmosphere", aggregate = FALSE) }

}
\author{
Marcos Alves, Kristine Karstens
}
