% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropareaAdjusted.R
\name{calcCropareaAdjusted}
\alias{calcCropareaAdjusted}
\title{calcCropareaAdjusted}
\usage{
calcCropareaAdjusted(iniyear, dataset = "Toolbox", sectoral = "kcr")
}
\arguments{
\item{iniyear}{initialization year}

\item{dataset}{LUH or Toolbox.
Note: once migration to Toolbox data is complete, this function
can be replaced with calcCropareaToolbox}

\item{sectoral}{crops to be reported: "kcr" for MAgPIE items, and "lpj" for LPJmL items}
}
\value{
magpie object in cellular resolution
}
\description{
This returns croparea as reported by FAO and LUH for the
             initialization year and
}
\examples{
\dontrun{
calcOutput("CropareaAdjusted", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
