% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropareaAdjusted.R
\name{calcCropareaAdjusted}
\alias{calcCropareaAdjusted}
\title{calcCropareaAdjusted}
\usage{
calcCropareaAdjusted(iniyear)
}
\arguments{
\item{iniyear}{initialization year}
}
\value{
magpie object in cellular resolution
}
\description{
This returns croparea as reported by FAO and LUH for the
             initialization year and
}
\examples{
\dontrun{
calcOutput("CropareaAdjusted", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
