% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropAreaPotIrrig.R
\name{calcCropAreaPotIrrig}
\alias{calcCropAreaPotIrrig}
\title{calcCropAreaPotIrrig}
\usage{
calcCropAreaPotIrrig(selectyears, comagyear, iniyear, cropmix, landScen)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{comagyear}{If NULL: total potential croparea is used;
if !NULL: already irrigated area is subtracted;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea (e.g. NULL, 1995, 2010)}

\item{iniyear}{Initialization year for current cropland area}

\item{cropmix}{Cropmix for which croparea share is calculated
(options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops as vector)}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates croparea that is potentially available
             for irrigated agriculture per crop given a chosen cropmix
}
\examples{
\dontrun{
calcOutput("CropAreaPotIrrig", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
