% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDischargeAccessibilityShare.R
\name{calcDischargeAccessibilityShare}
\alias{calcDischargeAccessibilityShare}
\title{calcDischargeAccessibilityShare}
\usage{
calcDischargeAccessibilityShare(
  lpjml,
  selectyears,
  climatetype,
  accessibilityrule
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{accessibilityrule}{Method used: Quantile method (Q) or coefficient of variation (CV)
combined with Scalar value defining the strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year (Qx, e.g. Q75: 0.25, Q50: 0.5)
or base value for expontential curve separated by : (CV:2)}
}
\value{
magpie object in cellular resolution representing share of discharge that is reserved for environmental flows
}
\description{
This function calculates the share of discharge that is useable
             given the variability of monthly flows. If discharge is highly
             variable, it is harder to bring into productive use and therefore
             water availability (for human use) is reduced.
}
\examples{
\dontrun{ calcOutput("DischargeAccessibilityShare", aggregate=FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
