% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDischargeAccessibilityShare.R
\name{calcDischargeAccessibilityShare}
\alias{calcDischargeAccessibilityShare}
\title{calcDischargeAccessibilityShare}
\usage{
calcDischargeAccessibilityShare(
  lpjml,
  selectyears,
  climatetype,
  accessibilityrule
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{selectyears}{Years to be returned
(Note: does not affect years of harmonization or smoothing)}

\item{climatetype}{Switch between different climate scenarios
or historical baseline "GSWP3-W5E5:historical"}

\item{accessibilityrule}{Method used: Quantile method (Q) or Coefficient of Variation (CV)
combined with scalar value defining the strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year
is inaccessible (e.g. Q:1 all discharge accessible,
Q:0.75 0.75-quantile is accessible everything that is more variable inaccessible)
or base value for exponential curve, separated by : (CV:2)}
}
\value{
magpie object in cellular resolution representing share of discharge
        that is accessible to humans
}
\description{
This function calculates the share of discharge that is useable
             given the variability of monthly flows. If discharge is highly
             variable, it is harder to bring into productive use and therefore
             water availability (for human use) is reduced.
}
\examples{
\dontrun{
calcOutput("DischargeAccessibilityShare", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
