% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDischargeInaccessible.R
\name{calcDischargeInaccessible}
\alias{calcDischargeInaccessible}
\title{calcDischargeInaccessible}
\usage{
calcDischargeInaccessible(
  lpjml,
  selectyears,
  climatetype,
  variabilitythreshold
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{variabilitythreshold}{Scalar value defining the strictness of accessibility restriction: discharge that is exceeded x percent of the time on average throughout a year (Qx). Default: 0.5 (Q50) (e.g. Q75: 0.25, Q50: 0.5)}
}
\value{
magpie object in cellular resolution representing share of discharge that is reserved for environmental flows
}
\description{
This function calculates the discharge that is inaccessible to humans
             based on the variability of monthly flows and natural discharge.
}
\examples{
\dontrun{ calcOutput("DischargeInaccessible", aggregate=FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
