% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEFRviolations.R
\name{calcEFRviolations}
\alias{calcEFRviolations}
\title{calcEFRviolations}
\usage{
calcEFRviolations(
  lpjml,
  selectyears,
  climatetype,
  efrMethod,
  accessibilityrule,
  rankmethod,
  yieldcalib,
  allocationrule,
  gainthreshold,
  irrigationsystem,
  iniyear,
  landScen,
  cropmix,
  comAg,
  multicropping,
  scenario,
  cellular = TRUE
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{selectyears}{Years for which irrigatable area is calculated}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year (Qx).
(e.g. Q75: 0.25, Q50: 0.5)}

\item{rankmethod}{Rank and optimization method consisting of
Unit according to which rank is calculated:
USD_ha (USD per hectare) for relative area return, or
USD_m3 (USD per cubic meter) for relative volumetric return;
USD for absolute return (total profit);
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices
and boolean indicating fullpotential (TRUE, i.e. cell
receives full irrigation requirements in total area)
or reduced potential (FALSE, reduced potential of cell
receives at later stage in allocation algorithm);
separated by ":"}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
     Also needs specification of refYields, separated by ":".
     Options: FALSE (for single cropping analyses) or
              "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst", "equality")}

\item{gainthreshold}{Threshold of yield improvement potential required
(same unit as in rankmethod)}

\item{irrigationsystem}{Irrigation system used
("surface", "sprinkler", "drip", "initialization")}

\item{iniyear}{Initialization year of irrigation system}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{comAg}{If TRUE: currently already irrigated areas in
         initialization year are reserved for irrigation,
if FALSE: no irrigation areas reserved (irrigation potential)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
            GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; TRUE:exogenous; FALSE)}

\item{scenario}{Combination of EFP ("on", "off") and
non-ag. water use scenario ("ssp2", "ISIMIP", ...)
separated by "."}

\item{cellular}{If TRUE: cellular data returned.
If FALSE: aggregated to basins}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates grid cell specific violation of
             environmental flow requirements
}
\examples{
\dontrun{
calcOutput("EFRviolations", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Benjamin Leon Bodirsky
}
