% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmulatorPastCorrectionFactor.R
\name{calcEmuPastCorrectFactor}
\alias{calcEmuPastCorrectFactor}
\title{calcEmuPastCorrectFactor}
\usage{
calcEmuPastCorrectFactor(model = "109325f71e")
}
\arguments{
\item{model}{grass harvest machine learnning model ID}
}
\value{
MAgPIE objects with pasture correction factor on a celullar level.
}
\description{
Loads the ANN trained grass emulator with historical data and calculates the pasture correction factor
}
\examples{

\dontrun{
calOutput("EmuPastCorrectFactor", model_id = "20f33a2280", folder = "GrassYieldsEmulator")
}

}
\author{
Marcos Alves
}
