% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnvmtlFlowRequirementsShare.R
\name{calcEnvmtlFlowRequirementsShare}
\alias{calcEnvmtlFlowRequirementsShare}
\title{calcEnvmtlFlowRequirementsShare}
\usage{
calcEnvmtlFlowRequirementsShare(
  version = "LPJmL4",
  EFRyears = c(1985:2015),
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  LFR_val = 0.1,
  HFR_LFR_less10 = 0.2,
  HFR_LFR_10_20 = 0.15,
  HFR_LFR_20_30 = 0.07,
  HFR_LFR_more30 = 0
)
}
\arguments{
\item{version}{Switch between LPJmL4 and LPJmL5}

\item{EFRyears}{Long-term reference time frame for EFR calculation}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{LFR_val}{Strictness of environmental flow requirements}

\item{HFR_LFR_less10}{High flow requirements (share of total water for cells) with LFR<10percent of total water}

\item{HFR_LFR_10_20}{High flow requirements (share of total water for cells) with 10percent < LFR < 20percent of total water}

\item{HFR_LFR_20_30}{High flow requirements (share of total water for cells) with 20percent < LFR < 30percent of total water}

\item{HFR_LFR_more30}{High flow requirements (share of total water for cells) with LFR>30percent of total water}
}
\value{
magpie object in cellular resolution representing share of discharge that is reserved for environmental flows
}
\description{
This function calculates environmental flow requirements (EFR) (as share of discharge) based on LPJmL monthly discharge following Smakthin et al. (2004)
}
\examples{
\dontrun{ calcOutput("EnvmtlFlowRequirementsShare", aggregate=FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
