% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnvmtlFlowRequirementsShare.R
\name{calcEnvmtlFlowRequirementsShare}
\alias{calcEnvmtlFlowRequirementsShare}
\title{calcEnvmtlFlowRequirementsShare}
\usage{
calcEnvmtlFlowRequirementsShare(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype = "GSWP3-W5E5:historical"
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop. Note: Default version arguments need to be updated when new versions are used!}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution representing share of discharge that is reserved for environmental flows
}
\description{
This function calculates environmental flow requirements (EFR) (as share of discharge) based on LPJmL monthly discharge following Smakthin et al. (2004)
}
\examples{
\dontrun{ calcOutput("EnvmtlFlowRequirementsShare", aggregate=FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
