% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFullIrrigationRequirement.R
\name{calcFullIrrigationRequirement}
\alias{calcFullIrrigationRequirement}
\title{calcFullIrrigationRequirement}
\usage{
calcFullIrrigationRequirement(
  lpjml,
  climatetype,
  selectyears,
  comagyear,
  irrigationsystem,
  landScen,
  cropmix,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned}

\item{comagyear}{if !NULL: already irrigated area is subtracted;
if NULL: total potential land area is used;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{landScen}{Land availability scenario (currCropland, currIrrig, potCropland)
combination of land availability scenario and initialization year separated by ":".
Initialization year only relevant for curr scenarios.
protection scenario separated by "_" (only relevant when potCropland selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the water requirements for full
             irrigation per cell per crop given potentially available land
}
\examples{
\dontrun{
calcOutput("FullIrrigationRequirement", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
