% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFullIrrigationRequirement.R
\name{calcFullIrrigationRequirement}
\alias{calcFullIrrigationRequirement}
\title{calcFullIrrigationRequirement}
\usage{
calcFullIrrigationRequirement(
  climatetype = "GSWP3-W5E5:historical",
  harmonize_baseline = FALSE,
  time = "spline",
  dof = 4,
  averaging_range = NULL,
  ref_year = NULL,
  selectyears = seq(1995, 2095, by = 5),
  iniyear = 1995,
  iniarea = TRUE,
  irrigationsystem = "initialization",
  protect_scen,
  proxycrop
)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{time}{time smoothing: average, spline or raw (default)}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{ref_year}{reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}

\item{selectyears}{years to be returned}

\item{iniyear}{year of initialization for cropland area}

\item{iniarea}{if TRUE (default): already irrigated area is subtracted, if FALSE: total potential land area is used}

\item{irrigationsystem}{irrigation system used: system share as in initialization year (default) or drip, surface, sprinkler for full irrigation by selected system}

\item{protect_scen}{land protection scenario: NULL (no irrigation limitation in protected areas), WDPA, BH, FF, CPD, LW, HalfEarth. Areas where no irrigation water withdrawals are allowed due to biodiversity protection}

\item{proxycrop}{historical crop mix pattern ("historical") or list of proxycrop(s)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the water requirements for full irrigation per cell per crop given potentially available land
}
\examples{
\dontrun{ calcOutput("FullIrrigationRequirement", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
