% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFullIrrigationRequirement.R
\name{calcFullIrrigationRequirement}
\alias{calcFullIrrigationRequirement}
\title{calcFullIrrigationRequirement}
\usage{
calcFullIrrigationRequirement(
  lpjml,
  climatetype,
  selectyears,
  comagyear,
  iniyear,
  irrigationsystem,
  landScen,
  cropmix,
  multicropping,
  yieldcalib
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned}

\item{comagyear}{if !NULL: already irrigated area is subtracted;
if NULL: total potential land area is used;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea}

\item{iniyear}{Croparea initialization year}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the water requirements for full
             irrigation per cell per crop given potentially available land
}
\examples{
\dontrun{
calcOutput("FullIrrigationRequirement", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
