% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFullIrrigationRequirement.R
\name{calcFullIrrigationRequirement}
\alias{calcFullIrrigationRequirement}
\title{calcFullIrrigationRequirement}
\usage{
calcFullIrrigationRequirement(
  climatetype,
  selectyears,
  iniyear,
  iniareayear,
  irrigationsystem,
  protect_scen,
  proxycrop
)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{years to be returned}

\item{iniyear}{year of initialization for cropland area}

\item{iniareayear}{if !NULL: already irrigated area is subtracted; if NULL: total potential land area is used; year specified here is the year of the initialization used for cropland area initialization in calcIrrigatedArea}

\item{irrigationsystem}{irrigation system used: system share as in initialization year (default) or drip, surface, sprinkler for full irrigation by selected system}

\item{protect_scen}{land protection scenario: NULL (no irrigation limitation in protected areas), WDPA, BH, FF, CPD, LW, HalfEarth. Areas where no irrigation water withdrawals are allowed due to biodiversity protection}

\item{proxycrop}{historical crop mix pattern ("historical") or list of proxycrop(s)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the water requirements for full irrigation per cell per crop given potentially available land
}
\examples{
\dontrun{ calcOutput("FullIrrigationRequirement", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
