% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFullIrrigationRequirement.R
\name{calcFullIrrigationRequirement}
\alias{calcFullIrrigationRequirement}
\title{calcFullIrrigationRequirement}
\usage{
calcFullIrrigationRequirement(
  version = "LPJmL5",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  harmonize_baseline = FALSE,
  time = "spline",
  dof = 4,
  cells = "lpjcell",
  selectyears = seq(1995, 2095, by = 5),
  iniyear = 1995,
  iniarea = TRUE,
  irrig_requirement = "withdrawal"
)
}
\arguments{
\item{version}{switch between LPJmL4 and LPJmL5}

\item{climatetype}{switch between different climate scenarios (default: "CRU_4")}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{time}{time smoothing: average, spline or raw (default)}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{cells}{switch between "lpjcell" (67420) and "magpiecell" (59199)}

\item{selectyears}{years to be returned}

\item{iniyear}{Year of initialization for cropland area}

\item{iniarea}{if TRUE (default): already irrigated area is subracted, if FALSE: total potential land area is used}

\item{irrig_requirement}{consumptive (consumption) or non-consumptive (withdrawals) irrigation water requirements}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{ref_year}{reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the water requirements for full irrigation per cell per crop given potentially available land
}
\examples{
\dontrun{ calcOutput("FullIrrigationRequirement", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
