% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGCMClimate.R
\name{calcGCMClimate}
\alias{calcGCMClimate}
\title{calcGCMClimate}
\usage{
calcGCMClimate(
  GCMModel = "HadGEM2",
  ClimateVariable = "temperature",
  rcp = "rcp85"
)
}
\arguments{
\item{rcp}{specify the RCP (rcp85, rcp60, rcp45 or rcp26)}

\item{level}{specify the spatial output level}
}
\value{
magpie object in cellular resolution
}
\description{
Disaggregate CO2 global atmospheric concentration to cellular level
}
\examples{
\dontrun{ calcOutput("GCMClimate", aggregate = FALSE) }

}
\author{
Marcos Alves, Kristine Karstens
}
