% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrassET.R
\name{calcGrassET}
\alias{calcGrassET}
\title{calcGrassET}
\usage{
calcGrassET(selectyears, lpjml, climatetype, season)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{season}{"wholeYear":  grass ET in the entire year (main + off season)
"mainSeason": grass ET in the crop-specific growing
              period of LPJmL (main season)}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates evapotranspiration (ET) of grassland
             under irrigated and rainfed conditions based on LPJmL inputs.
}
\examples{
\dontrun{
calcOutput("GrassET", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
