% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingPeriod.R
\name{calcGrowingPeriod}
\alias{calcGrowingPeriod}
\title{calcGrowingPeriod}
\usage{
calcGrowingPeriod(
  lpjml = "lpjml5.9.5-m1",
  climatetype = "MRI-ESM2-0:ssp370",
  stage = "harmonized2020",
  yield_ratio = 0.1
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{stage}{Degree of processing: raw, smoothed, harmonized, harmonized2020}

\item{yield_ratio}{threshold for cell yield over global average. crops in cells below threshold will be ignored}
}
\value{
magpie object in cellular resolution
}
\description{
This function determines a mean sowing date and a mean growing period
             for each cell in order to determine when irrigation can take place.
}
\examples{
\dontrun{
calcOutput("GrowingPeriod", aggregate = FALSE)
}

}
\author{
Kristine Karstens, Felicitas Beier
}
