% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingPeriod.R
\name{calcGrowingPeriod}
\alias{calcGrowingPeriod}
\title{calcGrowingPeriod}
\usage{
calcGrowingPeriod()
}
\value{
magpie object in cellular resolution
}
\description{
This function determine a mean sowing date and a mean growing period for each cell
in order to determine when irrigation can take place.
}
\examples{
\dontrun{ calcOutput("GrowingPeriod", aggregate = FALSE) }
}
\author{
Kristine Karstens
}
