% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingPeriod.R
\name{calcGrowingPeriod}
\alias{calcGrowingPeriod}
\title{calcGrowingPeriod}
\usage{
calcGrowingPeriod(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb"),
  climatetype = "GSWP3-W5E5:historical",
  stage = "harmonized2020",
  yield_ratio = 0.1,
  cells = "magpiecell"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{stage}{Degree of processing: raw, smoothed, harmonized, harmonized2020}

\item{yield_ratio}{threshold for cell yield over global average. crops in cells below threshold will be ignored}

\item{cells}{Number of cells to be returned
(select "magpiecell" for 59199 cells or "lpjcell" for 67420 cells)}
}
\value{
magpie object in cellular resolution
}
\description{
This function determines a mean sowing date and a mean growing period
             for each cell in order to determine when irrigation can take place.
}
\examples{
\dontrun{
calcOutput("GrowingPeriod", aggregate = FALSE)
}

}
\author{
Kristine Karstens, Felicitas Beier
}
