% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInitialLsu.R
\name{calcInitialLsu}
\alias{calcInitialLsu}
\title{calcInitialLsu}
\usage{
calcInitialLsu(model = "f41f19be67")
}
\arguments{
\item{model}{Grass harvest machine learning model ID}
}
\value{
MAgPIE objects with optimal lsu on a cellular level.
}
\description{
Loads the LSU that provides the maximum grass harvest as a initial values for MAgPIE
}
\examples{

\dontrun{
calOutput("InitialLsu", model = "f41f19be67")
}

}
\author{
Marcos Alves
}
