% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigAreaCommitted.R
\name{calcIrrigAreaCommitted}
\alias{calcIrrigAreaCommitted}
\title{calcIrrigAreaCommitted}
\usage{
calcIrrigAreaCommitted(selectyears, iniyear, depreciation = 0.1)
}
\arguments{
\item{selectyears}{select years}

\item{iniyear}{initialization year}

\item{depreciation}{parameter defining yearly depreciation rate at which previously cropland committed to irrigation becomes "unreserved" for irrigation (default: 0.1)}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area reserved for irrigation based on area irrigated in initialization year and depreciation parameter
}
\examples{
\dontrun{ calcOutput("IrrigAreaCommitted", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
