% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigAreaCommitted.R
\name{calcIrrigAreaCommitted}
\alias{calcIrrigAreaCommitted}
\title{calcIrrigAreaCommitted}
\usage{
calcIrrigAreaCommitted(
  selectyears = seq(1995, 2100, by = 5),
  iniyear = 1995,
  depreciation = 0.1
)
}
\arguments{
\item{selectyears}{select years}

\item{iniyear}{initialization year}

\item{depreciation}{parameter defining yearly depreciation rate at which previously irrigated cropland becomes "unreserved" for irrigation}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area reserved for irrigation based on area irrigated in initialization year and depreciation parameter
}
\examples{
\dontrun{ calcOutput("IrrigAreaCommitted", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
