% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigatableAreaUnlimited.R
\name{calcIrrigatableAreaUnlimited}
\alias{calcIrrigatableAreaUnlimited}
\title{calcIrrigatableAreaUnlimited}
\usage{
calcIrrigatableAreaUnlimited(
  selectyears,
  avlland_scen,
  lpjml,
  climatetype,
  cropmix,
  yieldcalib,
  thresholdtype,
  gainthreshold,
  multicropping
)
}
\arguments{
\item{selectyears}{years for which irrigatable area is calculated}

\item{avlland_scen}{Land availability scenario: current or potential;
optional additionally: protection scenario in case of potential
and initialization year of cropland area
combination of land availability scenario and initialization year separated by ":"
land availability scenario: currIrrig, currCropland, potIrrig
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{yieldcalib}{Calibrated (LPJmL yield potentials smoothed and harmonized
to baseline and calibrated with global FAO calibration factor
for proxycrops where LPJmL crops mapped multiple times to MAgPIE crops) or
FAO (LPJmL yields calibrated with current FAO yield)}

\item{thresholdtype}{Unit of yield improvement potential used as threshold:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{gainthreshold}{Threshold of yield improvement potential
(same unit as thresholdtype)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area that can potentially be irrigated given chosen
             land scenario and gainthreshold
}
\examples{
\dontrun{
calcOutput("IrrigatableAreaUnlimited", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
