% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigatableAreaUnlimited.R
\name{calcIrrigatableAreaUnlimited}
\alias{calcIrrigatableAreaUnlimited}
\title{calcIrrigatableAreaUnlimited}
\usage{
calcIrrigatableAreaUnlimited(
  selectyears,
  landScen,
  lpjml,
  climatetype,
  cropmix,
  yieldcalib,
  thresholdtype,
  gainthreshold,
  multicropping
)
}
\arguments{
\item{selectyears}{years for which irrigatable area is calculated}

\item{landScen}{Land availability scenario (currCropland, currIrrig, potCropland)
combination of land availability scenario and initialization year separated by ":".
Initialization year only relevant for curr scenarios.
protection scenario separated by "_" (only relevant when potCropland selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{thresholdtype}{Unit of yield improvement potential used as threshold:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{gainthreshold}{Threshold of yield improvement potential
(same unit as thresholdtype)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area that can potentially be irrigated given chosen
             land scenario and gainthreshold
}
\examples{
\dontrun{
calcOutput("IrrigatableAreaUnlimited", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
