% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigatableAreaUnlimited.R
\name{calcIrrigatableAreaUnlimited}
\alias{calcIrrigatableAreaUnlimited}
\title{calcIrrigatableAreaUnlimited}
\usage{
calcIrrigatableAreaUnlimited(
  selectyears,
  iniyear,
  landScen,
  lpjml,
  climatetype,
  cropmix,
  yieldcalib,
  thresholdtype,
  gainthreshold,
  multicropping
)
}
\arguments{
\item{selectyears}{years for which irrigatable area is calculated}

\item{iniyear}{initialization year}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_FF, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{thresholdtype}{Unit of yield improvement potential used as threshold:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{gainthreshold}{Threshold of yield improvement potential
(same unit as thresholdtype)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area that can potentially be irrigated given chosen
             land scenario and gainthreshold
}
\examples{
\dontrun{
calcOutput("IrrigatableAreaUnlimited", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
