% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigatedArea.R
\name{calcIrrigatedArea}
\alias{calcIrrigatedArea}
\title{calcIrrigatedArea}
\usage{
calcIrrigatedArea(
  selectyears = seq(1995, 2100, by = 5),
  iniyear = 1995,
  depreciation = 0.1,
  cells = "lpjcell"
)
}
\arguments{
\item{selectyears}{select years}

\item{iniyear}{initialization year}

\item{depreciation}{parameter defining rate at which previously irrigated cropland becomes "unreserved" for irrigation}

\item{cells}{cells to be returned by the function (lpjcell or magpiecell)}
}
\value{
magpie object in cellular resolution
}
\description{
calculates area reserved for irrigation based on area irrigated in initialization year and depreciation parameter
}
\examples{
\dontrun{ calcOutput("IrrigatedArea", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
