% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigationSystem.R
\name{calcIrrigationSystem}
\alias{calcIrrigationSystem}
\title{calcIrrigationSystem}
\usage{
calcIrrigationSystem(source = "Jaegermeyr_lpjcell")
}
\arguments{
\item{source}{Data source to be used: Jaegermeyr (irrigation system share based on FAO 2014, ICID 2012 and Rohwer et al. 2007) or LPJmL (dominant irrigation system per country) and number of cells (lpjcell or magpiecell) separated by _}
}
\value{
magpie object in cellular resolution
}
\description{
This function returns the irrigation system share initialization
}
\examples{
\dontrun{ calcOutput("IrrigationSystem",source="Jaegermeyr_lpjcell",aggregate = FALSE) }

}
\author{
Felicitas Beier
}
