% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigCellranking.R
\name{calcIrrigCellranking}
\alias{calcIrrigCellranking}
\title{calcIrrigCellranking}
\usage{
calcIrrigCellranking(
  lpjml,
  climatetype,
  cellrankyear,
  method,
  cropmix,
  iniyear,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical" for yields}

\item{cellrankyear}{Year(s) for which cell rank is calculated}

\item{method}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return;
and boolean indicating fullpotential (TRUE) or reduced potential (FALSE)}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{iniyear}{Initialization year for price}

\item{yieldcalib}{Calibrated (LPJmL yield potentials smoothed and harmonized
to baseline and calibrated with global FAO calibration factor
for proxycrops where LPJmL crops mapped multiple times to MAgPIE crops) or
FAO (LPJmL yields calibrated with current FAO yield)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates a cellranking
             for the river basin discharge allocation
             based on yield improvement potential through irrigation
}
\examples{
\dontrun{
calcOutput("IrrigCellranking", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
