% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigCellranking.R
\name{calcIrrigCellranking}
\alias{calcIrrigCellranking}
\title{calcIrrigCellranking}
\usage{
calcIrrigCellranking(
  lpjml,
  climatetype,
  cellrankyear,
  method,
  cropmix,
  iniyear,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical" for yields}

\item{cellrankyear}{Year(s) for which cell rank is calculated}

\item{method}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return;
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices;
and boolean indicating fullpotential (TRUE) or reduced potential (FALSE)}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{iniyear}{Initialization year for price}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
         Also needs specification of refYields, separated by ":".
         Options: FALSE (for single cropping analyses) or
                  "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; TRUE:exogenous; FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates a cellranking
             for the river basin discharge allocation
             based on yield improvement potential through irrigation
}
\examples{
\dontrun{
calcOutput("IrrigCellranking", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
