% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigCellranking.R
\name{calcIrrigCellranking}
\alias{calcIrrigCellranking}
\title{calcIrrigCellranking}
\usage{
calcIrrigCellranking(
  lpjml,
  climatetype,
  cellrankyear,
  iniyear,
  comagyear,
  irrigationsystem,
  landScen,
  method,
  efrMethod,
  transDist,
  cropmix,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical" for yields}

\item{cellrankyear}{Year(s) for which cell rank is calculated}

\item{iniyear}{Initialization year for price}

\item{comagyear}{if !NULL: already irrigated area is subtracted;
if NULL: total potential land area is used;
year specified here is the year of the initialization
used for cropland area initialization in calcIrrigatedArea}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{method}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for relative area return, or
USD_m3 (USD per cubic meter) for relative volumetric return;
USD for absolute return (total profit);
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices;
and boolean indicating fullpotential (TRUE) or reduced potential (FALSE)}

\item{efrMethod}{if method USD or US_m3: EFR method used to calculate committed
agricultural use relevant for chosen land scenario (e.g., Smakhtin:good, VMF:fair)}

\item{transDist}{if method USD or US_m3: Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability
of committed agricultural uses relevant for chosen land scenario}

\item{cropmix}{Selected cropmix for which yield improvement potential
is calculated (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)
NULL returns all crops individually}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
         Also needs specification of refYields, separated by ":".
         Options: FALSE (for single cropping analyses) or
                  "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates a cellranking
             for the river basin discharge allocation
             based on yield improvement potential through irrigation
}
\examples{
\dontrun{
calcOutput("IrrigCellranking", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
