% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigCellranking.R
\name{calcIrrigCellranking}
\alias{calcIrrigCellranking}
\title{calcIrrigCellranking}
\usage{
calcIrrigCellranking(
  climatetype = "GSWP3-W5E5:historical",
  cellrankyear = "y1995",
  method = "meancellrank",
  proxycrop = c("maiz", "rapeseed", "puls_pro"),
  iniyear = 1995
)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical" for yields}

\item{cellrankyear}{year(s) for which cell rank is calculated}

\item{method}{method of calculating the rank: "meancellrank" (default): mean over cellrank of proxy crops, "meancroprank": rank over mean of proxy crops (normalized), "meanpricedcroprank": rank over mean of proxy crops (normalized using price), "watervalue": rank over value of irrigation water}

\item{proxycrop}{proxycrop(s) selected for crop mix specific calculations: list of proxycrops or "historical"}

\item{iniyear}{initialization year for price in price-weighted normalization of meanpricedcroprank and case of monetary=TRUE year of prices}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates a cellranking for the river basin discharge allocation based on yield improvement potential through irrigation
}
\examples{
\dontrun{ calcOutput("IrrigCellranking", aggregate=FALSE) }
}
\author{
Felicitas Beier
}
