% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigCellranking.R
\name{calcIrrigCellranking}
\alias{calcIrrigCellranking}
\title{calcIrrigCellranking}
\usage{
calcIrrigCellranking(
  version = "LPJmL5",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = FALSE,
  ref_year = "y2015",
  cellrankyear = "y1995",
  cells = "lpjcell",
  crops = "magpie",
  method = "meancellrank",
  proxycrop = c("maiz", "rapeseed", "puls_pro")
)
}
\arguments{
\item{version}{switch between LPJmL version for yields}

\item{climatetype}{switch between different climate scenarios for yields}

\item{time}{time smoothing: average, spline (default) or raw}

\item{averaging_range}{just specify for time=="average": number of time steps to average}

\item{dof}{just specify for time=="spline": degrees of freedom}

\item{harmonize_baseline}{FALSE (default) no harmonization, harmonization: if a baseline is specified here data is harmonized to that baseline (from ref_year onwards)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}

\item{cellrankyear}{year for which cell rank is calculated}

\item{cells}{switch between "lpjcell" (67420) and "magpiecell" (59199)}

\item{crops}{switch between "magpie" and "lpjml" (default) crops}

\item{method}{method of calculating the rank: "meancellrank" (default): mean over cellrank of proxy crops, "meancroprank": rank over mean of proxy crops}

\item{proxycrop}{proxycrop(s) selected for rank calculation}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates a cellranking for the river basin discharge allocation based on yield improvement potential through irrigation
}
\examples{
\dontrun{ calcOutput("calcIrrigCellranking", aggregate=FALSE) }
}
\author{
Felicitas Beier
}
