% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigWatRequirements.R
\name{calcIrrigWatRequirements}
\alias{calcIrrigWatRequirements}
\title{calcIrrigWatRequirements}
\usage{
calcIrrigWatRequirements(
  selectyears = "all",
  lpjml = c(natveg = "LPJmL4", crop = "LPJmL5"),
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = FALSE,
  ref_year = NULL
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop. Note: Default version arguments need to be updated when new versions are used!}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default) of input data to this function}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization of input data to this function, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline of input data to this function (just specify when harmonize_baseline=TRUE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates irrigation water requirements based on LPJmL blue water consumption of plants and considering irrigation efficiencies
}
\examples{
\dontrun{ calcOutput("IrrigWatRequirements", aggregate=FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
