% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigWatRequirements.R
\name{calcIrrigWatRequirements}
\alias{calcIrrigWatRequirements}
\title{calcIrrigWatRequirements}
\usage{
calcIrrigWatRequirements(
  selectyears,
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop. Note: Default version arguments need to be updated when new versions are used!}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates irrigation water requirements based on LPJmL blue water consumption of plants and considering irrigation efficiencies
}
\examples{
\dontrun{ calcOutput("IrrigWatRequirements", aggregate=FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
