% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigWatRequirements.R
\name{calcIrrigWatRequirements}
\alias{calcIrrigWatRequirements}
\title{calcIrrigWatRequirements}
\usage{
calcIrrigWatRequirements(selectyears, lpjml, climatetype, multicropping)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; FALSE:NULL)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates irrigation water requirements based on
             LPJmL blue water consumption of crops and
             considering irrigation efficiencies
}
\examples{
\dontrun{
calcOutput("IrrigWatRequirements", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
