% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigWatRequirements.R
\name{calcIrrigWatRequirements}
\alias{calcIrrigWatRequirements}
\title{calcIrrigWatRequirements}
\usage{
calcIrrigWatRequirements(
  selectyears,
  lpjml,
  climatetype,
  multicropping,
  proxycrops = "combined"
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Climate model or historical baseline "GSWP3-W5E5:historical"}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from LandInG
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{proxycrops}{only relevant for multicropping == TRUE
(must be set to "combined" otherwise, default);
required for special treatment of LPJmL annual crops that serve as
proxy for MAgPIE perennials;
Arguments: TRUE, FALSE, combined
FALSE: standard crops
TRUE: proxy crops of LPJmL for MAgPIE perennials selected
combined: final output that includes special treatment of perennials
and standard crop aggregation}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates irrigation water requirements based on
             LPJmL blue water consumption of crops and
             considering irrigation efficiencies
}
\examples{
\dontrun{
calcOutput("IrrigWatRequirements", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
