% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigWatRequirements.R
\name{calcIrrigWatRequirements}
\alias{calcIrrigWatRequirements}
\title{calcIrrigWatRequirements}
\usage{
calcIrrigWatRequirements(
  selectyears = "all",
  cells = "lpjcell",
  crops = "magpie",
  version = "LPJmL5",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = FALSE,
  ref_year = NULL,
  irrig_requirement = "withdrawal"
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{cells}{Switch between "lpjcell" (67420) and "magpiecell" (59199)}

\item{crops}{Selects "magpie" (default) or "lpjml" crops}

\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default) of input data to this function}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization of input data to this function, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline of input data to this function (just specify when harmonize_baseline=TRUE)}

\item{irrig_requirement}{Consumptive (consumption) or non-consumptive (withdrawals) irrigation water requirements}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates irrigation water requirements based on LPJmL blue water consumption of plants and considering irrigation efficiencies
}
\examples{
\dontrun{ calcOutput("IrrigWatRequirements", aggregate=FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
