% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigWatValue.R
\name{calcIrrigWatValue}
\alias{calcIrrigWatValue}
\title{calcIrrigWatValue}
\usage{
calcIrrigWatValue(
  selectyears = c(1995, 2100, 5),
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = FALSE,
  ref_year = "y2015",
  cells = "lpjcell",
  iniyear = 1995
)
}
\arguments{
\item{selectyears}{years to be returned}

\item{climatetype}{switch between different climate scenarios for yields and irrigation water requirements}

\item{time}{time smoothing: average, spline (default) or raw}

\item{averaging_range}{just specify for time=="average": number of time steps to average}

\item{dof}{just specify for time=="spline": degrees of freedom}

\item{harmonize_baseline}{FALSE (default) no harmonization, harmonization: if a baseline is specified here data is harmonized to that baseline (from ref_year onwards)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}

\item{cells}{switch between "lpjcell" (67420) and "magpiecell" (59199)}

\item{iniyear}{initialization year for price in price-weighted normalization of meanpricedcroprank}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the value of irrigation water (added value of water to the production process) based on yield gain, crop prices and irrigation water requirements
}
\examples{
\dontrun{ calcOutput("IrrigWatValue", aggregate=FALSE) }
}
\author{
Felicitas Beier
}
