% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigWatValue.R
\name{calcIrrigWatValue}
\alias{calcIrrigWatValue}
\title{calcIrrigWatValue}
\usage{
calcIrrigWatValue(
  lpjml,
  selectyears,
  climatetype,
  iniyear,
  proxycrop,
  FAOyieldcalib
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{years to be returned}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{iniyear}{initialization year for price in price-weighted normalization of meanpricedcroprank and yield improvement potential prices}

\item{proxycrop}{historical crop mix pattern ("historical") or list of proxycrop(s) or NULL for all crops}

\item{FAOyieldcalib}{TRUE (LPJmL yields scaled with current FAO yield) or FALSE (LPJmL yield potentials)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the value of irrigation water (added value of water to the production process) based on yield gain, crop prices and irrigation water requirements
}
\examples{
\dontrun{ calcOutput("IrrigWatValue", aggregate=FALSE) }
}
\author{
Felicitas Beier
}
