% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigYieldImprovementPotential.R
\name{calcIrrigYieldImprovementPotential}
\alias{calcIrrigYieldImprovementPotential}
\title{calcIrrigYieldImprovementPotential}
\usage{
calcIrrigYieldImprovementPotential(
  lpjml,
  climatetype,
  unit,
  iniyear,
  selectyears,
  comAg,
  cropmix,
  landScen,
  irrigationsystem,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{unit}{Unit of yield improvement potential to be returned and
level of price aggregation used, separated by ":".
Unit:
USD_ha (USD per hectare) for relative area return, or
USD_m3 (USD per cubic meter) for relative volumetric return;
USD for absolute return (total profit);
for relative return according to area and volume.
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices}

\item{iniyear}{initialization year for food price and cropmix area}

\item{selectyears}{Years to be returned by the function}

\item{comAg}{If TRUE: committed irrigated areas are subtracted,
if FALSE: total potential croparea is used}

\item{cropmix}{Selected cropmix for which yield improvement potential
is calculated (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)
NULL returns all crops individually}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, or one of the scenarios
available in calcConservationPriorities,
   e.g., 30by20, BH, BH_IFL, PBL_HalfEarth,
   or NA for no protection).
For case of no land protection select "NA" in second part of argument
or do not specify second part of the argument}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
Also needs specification of refYields, separated by ":".
Options: FALSE (for single cropping analyses) or
         "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the yield improvement potential
             through irrigation for all grid cells given a certain crop mix
}
\examples{
\dontrun{
calcOutput("IrrigYieldImprovementPotential", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
