% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigYieldImprovementPotential.R
\name{calcIrrigYieldImprovementPotential}
\alias{calcIrrigYieldImprovementPotential}
\title{calcIrrigYieldImprovementPotential}
\usage{
calcIrrigYieldImprovementPotential(
  lpjml,
  climatetype,
  unit,
  iniyear,
  selectyears,
  cropmix,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{unit}{Unit of yield improvement potential to be returned:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{iniyear}{initialization year for food price and cropmix area}

\item{selectyears}{Years to be returned by the function}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix
NULL returns all crops individually}

\item{yieldcalib}{Calibrated (LPJmL yield potentials smoothed and harmonized
to baseline and calibrated with global FAO calibration factor
for proxycrops where LPJmL crops mapped multiple times to MAgPIE crops) or
FAO (LPJmL yields calibrated with current FAO yield)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the yield improvement potential
             through irrigation for different crops
}
\examples{
\dontrun{
calcOutput("IrrigYieldImprovementPotential", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
