% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigYieldImprovementPotential.R
\name{calcIrrigYieldImprovementPotential}
\alias{calcIrrigYieldImprovementPotential}
\title{calcIrrigYieldImprovementPotential}
\usage{
calcIrrigYieldImprovementPotential(
  climatetype,
  monetary,
  iniyear,
  selectyears,
  proxycrop,
  FAOyieldcalib
)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical" for yields}

\item{monetary}{yield improvement potential in tDM (FALSE, default) or priced yield improvement potential in USD05 (TRUE)}

\item{iniyear}{year to be used when monetary activated}

\item{selectyears}{years to be returned by the function}

\item{proxycrop}{proxycrop(s) selected for crop mix specific calculations: average over proxycrop(s) yield gain. NULL returns all crops individually}

\item{FAOyieldcalib}{TRUE (LPJmL yields scaled with current FAO yield) or FALSE (LPJmL yield potentials)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the yield improvement potential of irrigation for different crops
}
\examples{
\dontrun{ calcOutput("IrrigYieldImprovementPotential", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
