% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigYieldImprovementPotential.R
\name{calcIrrigYieldImprovementPotential}
\alias{calcIrrigYieldImprovementPotential}
\title{calcIrrigYieldImprovementPotential}
\usage{
calcIrrigYieldImprovementPotential(
  lpjml,
  climatetype,
  monetary,
  iniyear,
  selectyears,
  cropmix,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version required for yields}

\item{climatetype}{Climate scenarios or historical baseline "GSWP3-W5E5:historical" for yields}

\item{monetary}{Yield improvement potential in tDM (FALSE, default) or
priced yield improvement potential in USD05 (TRUE)}

\item{iniyear}{initialization year for food price and cropmix area}

\item{selectyears}{Years to be returned by the function}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix
NULL returns all crops individually}

\item{yieldcalib}{FAO (LPJmL yields calibrated with current FAO yield) or
calibrated (LPJmL yield potentials harmonized to baseline and calibrated for proxycrops) or
smoothed (smoothed LPJmL yield potentials, not harmonized, not calibrated) or
smoothed_calib (not harmonized, but calibrated)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the yield improvement potential of irrigation for different crops
}
\examples{
\dontrun{
calcOutput("IrrigYieldImprovementPotential", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
