% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigYieldImprovementPotential.R
\name{calcIrrigYieldImprovementPotential}
\alias{calcIrrigYieldImprovementPotential}
\title{calcIrrigYieldImprovementPotential}
\usage{
calcIrrigYieldImprovementPotential(
  lpjml,
  climatetype,
  unit,
  iniyear,
  selectyears,
  cropmix,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{unit}{Unit of yield improvement potential to be returned and
level of price aggregation used, separated by ":".
Unit:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return.
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices}

\item{iniyear}{initialization year for food price and cropmix area}

\item{selectyears}{Years to be returned by the function}

\item{cropmix}{Selected cropmix for which yield improvement potential
is calculated (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)
NULL returns all crops individually}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:cropIrrig" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the yield improvement potential
             through irrigation for all grid cells given a certain crop mix
}
\examples{
\dontrun{
calcOutput("IrrigYieldImprovementPotential", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
