% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigYieldImprovementPotential.R
\name{calcIrrigYieldImprovementPotential}
\alias{calcIrrigYieldImprovementPotential}
\title{calcIrrigYieldImprovementPotential}
\usage{
calcIrrigYieldImprovementPotential(
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  monetary = FALSE,
  iniyear = 1995,
  harmonize_baseline = FALSE,
  ref_year = NULL,
  selectyears = seq(1995, 2095, by = 5),
  cells = "magpiecell",
  proxycrop
)
}
\arguments{
\item{climatetype}{switch between different climate scenarios (default: "CRU_4") of calcYields function}

\item{time}{time smoothing of calcYields function: average, spline (default) or raw}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{monetary}{yield improvement potential in tDM (FALSE, default) or priced yield improvement potential in USD05 (TRUE)}

\item{iniyear}{year to be used when monetary activated}

\item{harmonize_baseline}{harmonization in calcYields function: FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year onwards)}

\item{ref_year}{reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}

\item{selectyears}{years to be returned by the function}

\item{cells}{switch between "lpjcell" (67420) and "magpiecell" (59199)}

\item{proxycrop}{proxycrop(s) selected for crop mix specific calculations: average over proxycrop(s) yield gain. NULL returns all crops individually}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the yield improvement potential of irrigation for different crops
}
\examples{
\dontrun{ calcOutput("IrrigYieldImprovementPotential", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
