% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationRateNatural.R
\name{calcNitrogenFixationRateNatural}
\alias{calcNitrogenFixationRateNatural}
\title{calcNitrogenFixationRateNatural}
\usage{
calcNitrogenFixationRateNatural()
}
\value{
magpie object in cellular resolution
}
\description{
Function that extracts Nitrogen fixation rate of natural vegetation
}
\examples{
\dontrun{ calcOutput("NitrogenFixationRateNatural", aggregate = FALSE) }

}
\author{
Marcos Alves
}
