% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonAgWaterDemand.R
\name{calcNonAgWaterDemand}
\alias{calcNonAgWaterDemand}
\title{calcNonAgWaterDemand}
\usage{
calcNonAgWaterDemand(
  selectyears = "all",
  source = "WATCH_ISIMIP_WATERGAP",
  seasonality = "grper",
  waterusetype = "withdrawal"
)
}
\arguments{
\item{selectyears}{years to be returned}

\item{source}{data source to be used}

\item{seasonality}{grper (default): non-agricultural water demand in growing period per year; total: non-agricultural water demand throughout the year}

\item{waterusetype}{withdrawal (default) or consumption}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts non-agricultural water demand
}
\examples{
\dontrun{ calcOutput("NonAgWaterDemand", aggregate = FALSE) }

}
\author{
Felicitas Beier
}
