% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNpiNdcAdAolcPol.R
\name{calcNpiNdcAdAolcPol}
\alias{calcNpiNdcAdAolcPol}
\title{calcNpiNdcAdAolcPol}
\usage{
calcNpiNdcAdAolcPol()
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts NPI/NDC forestry (area?)
}
\examples{
\dontrun{ calcOutput("NpiNdcAdAolcPol", aggregate = FALSE) }

}
\author{
Michael Windisch
}
