% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialIrrigWatRequirements.R
\name{calcPotentialIrrigWatRequirements}
\alias{calcPotentialIrrigWatRequirements}
\title{calcPotentialIrrigWatRequirements}
\usage{
calcPotentialIrrigWatRequirements(selectyears, iniyear, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year (for weight by cropland)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates actual irrigation water requirements
             per cell and season given a certain irrigation system
}
\examples{
\dontrun{
calcOutput("PotentialIrrigWatRequirements", aggregate = FALSE)
}

}
\seealso{
\code{\link{calcActualIrrigWatRequirements}}, \code{\link{calcIrrigWatRequirements}}
}
\author{
Felicitas Beier
}
