% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverDischargeNatAndHuman.R
\name{calcRiverDischargeNatAndHuman}
\alias{calcRiverDischargeNatAndHuman}
\title{calcRiverDischargeNatAndHuman}
\usage{
calcRiverDischargeNatAndHuman(
  lpjml,
  selectyears,
  iniyear,
  climatetype,
  EFRmethod
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{iniyear}{Initialization year of irrigation system}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates cellular discharge after considering known human consumption (non-agricultural and committed agricultural) along the river calculated and accounted for in previous river routings (see calcRiverNaturalFlows and calcRiverHumanUses)
}
\examples{
\dontrun{ calcOutput("RiverDischargeNatAndHuman", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
