% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverHumanUses.R
\name{calcRiverHumanUses}
\alias{calcRiverHumanUses}
\title{calcRiverHumanUses}
\usage{
calcRiverHumanUses(
  humanuse,
  lpjml,
  climatetype,
  selectyears,
  iniyear,
  EFRmethod
)
}
\arguments{
\item{humanuse}{Human use type to which river routing shall be applied
(non_agriculture or committed_agriculture).}

\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios
or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned
(Note: does not affect years of harmonization or smoothing)}

\item{iniyear}{Initialization year of irrigation system}

\item{EFRmethod}{EFR method used including selected strictness of EFRs
(Smakhtin:good, VMF:fair)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates human uses and reserved water along
             the river
}
\examples{
\dontrun{
calcOutput("RiverHumanUses", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
