% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverHumanUses.R
\name{calcRiverHumanUses}
\alias{calcRiverHumanUses}
\title{calcRiverHumanUses}
\usage{
calcRiverHumanUses(selectyears, humanuse, iniyear, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{humanuse}{Human use type to which river routing shall be applied (non_agriculture or committed_agriculture). Note: non_agriculture must be run prior to committed_agriculture}

\item{iniyear}{Initialization year of irrigation system}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates natural discharge for the river routing derived from inputs from LPJmL
}
\examples{
\dontrun{ calcOutput("RiverHumanUses", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
