% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverNaturalFlows2.R
\name{calcRiverNaturalFlows2}
\alias{calcRiverNaturalFlows2}
\title{calcRiverNaturalFlows_magpie_magpie}
\usage{
calcRiverNaturalFlows2(
  selectyears,
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype = "HadGEM2_ES:rcp2p6:co2"
)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop. Note: Default version arguments need to be updated when new versions are used!}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates natural discharge for the river routing derived from inputs from LPJmL
}
\examples{
\dontrun{ calcOutput("RiverNaturalFlows_magpie", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
