% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverSurplusDischargeAllocation.R
\name{calcRiverSurplusDischargeAllocation}
\alias{calcRiverSurplusDischargeAllocation}
\title{calcRiverSurplusDischargeAllocation}
\usage{
calcRiverSurplusDischargeAllocation(
  lpjml,
  climatetype,
  selectyears,
  output,
  EFRmethod,
  accessibilityrule,
  rankmethod,
  allocationrule,
  thresholdtype,
  gainthreshold,
  irrigationsystem,
  iniyear,
  avlland_scen,
  cropmix,
  yieldcalib,
  com_ag,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios
or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned
(Note: does not affect years of harmonization or smoothing)}

\item{output}{Output to be reported}

\item{EFRmethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Method used: Quantile method (Q) or Coefficient of Variation (CV)
combined with scalar value defining the strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year
(Qx, e.g. Q75: 0.25, Q50: 0.5)
or base value for exponential curve separated by : (CV:2)}

\item{rankmethod}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return;
and boolean indicating fullpotential
(TRUE, i.e. cell receives full irrigation requirements in total area)
or reduced potential
(FALSE, reduced potential of cell receives at later stage in allocation algorithm);
separated by ":"}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst")}

\item{thresholdtype}{Unit of yield improvement potential used as threshold:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{gainthreshold}{Threshold of yield improvement potential required for
water allocation in upstreamfirst algorithm
(in same unit as thresholdtype)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge
allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{iniyear}{Initialization year of irrigation system}

\item{avlland_scen}{Land availability scenario: current or potential;
optional additionally: protection scenario in case of potential (when left empty: no protection)
and initialization year of cropland area
combination of land availability scenario and initialization year separated by ":".
land availability scenario: currIrrig (only currently irrigated cropland available for irrigated agriculture),
currCropland (only current cropland areas available for irrigated agriculture),
potIrrig (suitable land is available for irrigated agriculture, potentially land restrictions activated through protect_scen argument)
protection scenario separated by "_" (only relevant when potIrrig selected): WDPA, BH, FF, CPD, LW, HalfEarth}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{yieldcalib}{Calibrated (LPJmL yield potentials smoothed and harmonized
to baseline and calibrated with global FAO calibration factor
for proxycrops where LPJmL crops mapped multiple times to MAgPIE crops) or
FAO (LPJmL yields calibrated with current FAO yield)}

\item{com_ag}{if TRUE: the currently already irrigated areas in
         initialization year are reserved for irrigation,
if FALSE: no irrigation areas reserved (irrigation potential)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function distributes surplus basin discharge after
             previous river routings following certain management assumptions
}
\examples{
\dontrun{
calcOutput("RiverSurplusDischargeAllocation", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
