% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverSurplusDischargeAllocation.R
\name{calcRiverSurplusDischargeAllocation}
\alias{calcRiverSurplusDischargeAllocation}
\title{calcRiverSurplusDischargeAllocation}
\usage{
calcRiverSurplusDischargeAllocation(
  selectyears = "all",
  output,
  version = "LPJmL4",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = "CRU_4",
  ref_year = "y2015",
  allocationrule = "optimization",
  thresholdtype = TRUE,
  gainthreshold = 10,
  irrigationsystem = "initialization",
  iniyear = 1995,
  protect_scen,
  proxycrop
)
}
\arguments{
\item{selectyears}{Years to be returned (Note: does not affect years of harmonization or smoothing)}

\item{output}{output to be reported}

\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm: TRUE (default): monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{gainthreshold}{Threshold of yield improvement potential required for water allocation in upstreamfirst algorithm (in tons per ha)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{iniyear}{Initialization year of irrigation system}

\item{protect_scen}{land protection scenario: NULL (no irrigation limitation in protected areas), WDPA, BH, FF, CPD, LW, HalfEarth. Areas where no irrigation water withdrawals are allowed due to biodiversity protection}

\item{proxycrop}{historical crop mix pattern ("historical") or list of proxycrop(s)}
}
\value{
magpie object in cellular resolution
}
\description{
This function distributes surplus basin discharge after the previous river routings following certain management assumptions
}
\examples{
\dontrun{ calcOutput("RiverSurplusDischargeAllocation", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
