% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRrLayer.R
\name{calcRRLayer}
\alias{calcRRLayer}
\title{calcRRLayer}
\usage{
calcRRLayer()
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts range-rarity as used for biodiversity loss
}
\examples{
\dontrun{ calcOutput("RRLayer", aggregate = FALSE) }

}
\author{
Michael Windisch, Patrick v. Jeetze
}
