% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShrCurrIrrigFulfilled.R
\name{calcShrCurrIrrigFulfilled}
\alias{calcShrCurrIrrigFulfilled}
\title{calcShrCurrIrrigFulfilled}
\usage{
calcShrCurrIrrigFulfilled(
  lpjml,
  climatetype,
  multicropping,
  selectyears,
  iniyear,
  efrMethod
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}
}
\value{
cellular magpie object
}
\description{
calculates of share of current irrigation that can be fulfilled
             given renewable water availability of the algorithm
}
\examples{
\dontrun{
calcShrCurrIrrigFulfilled()
}

}
\author{
Felicitas Beier
}
