% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShrHumanUsesFulfilled.R
\name{calcShrHumanUsesFulfilled}
\alias{calcShrHumanUsesFulfilled}
\title{calcShrHumanUsesFulfilled}
\usage{
calcShrHumanUsesFulfilled(
  lpjml,
  climatetype,
  transDist,
  multicropping,
  selectyears,
  iniyear,
  efrMethod
)
}
\arguments{
\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
("endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set),
separated by ":"
(e.g. TRUE:endogenous; TRUE:exogenous; FALSE)}

\item{selectyears}{Years to be returned}

\item{iniyear}{Initialization year}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}
}
\value{
cellular magpie object
}
\description{
calculates of share of current non-agricultural and irrigation 
             that can be fulfilled
             given renewable water availability of the algorithm
}
\examples{
\dontrun{
calcOutput("ShrHumanUsesFulfilled", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
