% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterAllocation.R
\name{calcWaterAllocation}
\alias{calcWaterAllocation}
\title{calcWaterAllocation}
\usage{
calcWaterAllocation(
  selectyears = "all",
  output = "consumption",
  version = "LPJmL4",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  harmonize_baseline = FALSE,
  ref_year = "y2015",
  allocationrule = "optimization",
  allocationshare = NULL,
  gainthreshold = 1,
  irrigationsystem = "initialization",
  irrigini = "Jaegermeyr_lpjcell"
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{output}{Water availability output to be returned: withdrawal or consumption}

\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{allocationshare}{Share of water to be allocated to cell (only needs to be selected in case of allocationrule=="equality")}

\item{gainthreshold}{Threshold of yield improvement potential required for water allocation in upstreamfirst algorithm (in tons per ha)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{irrigini}{When "initialization" selected for irrigation system: choose initialization data set for irrigation system initialization ("Jaegermeyr_lpjcell", "LPJmL_lpjcell")}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water availability for MAgPIE retrieved from LPJmL using a river routing and allocation algorithm for distribution of discharge within the river basin
}
\examples{
\dontrun{ calcOutput("WaterAllocation", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
