% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterUseCommittedAg.R
\name{calcWaterUseCommittedAg}
\alias{calcWaterUseCommittedAg}
\title{calcWaterUseCommittedAg}
\usage{
calcWaterUseCommittedAg(climatetype, selectyears, iniyear)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned}

\item{iniyear}{Year of initialization for cropland area}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates committed agricultural water uses that are used in the river routing algorithm for distributing available water across the basin
}
\examples{
\dontrun{ calcOutput("WaterUseCommittedAg", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
