% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterUseCommittedAg.R
\name{calcWaterUseCommittedAg}
\alias{calcWaterUseCommittedAg}
\title{calcWaterUseCommittedAg}
\usage{
calcWaterUseCommittedAg(
  climatetype = "GSWP3-W5E5:historical",
  selectyears = seq(1995, 2095, by = 5),
  time = "spline",
  dof = 4,
  averaging_range = NULL,
  harmonize_baseline = "CRU_4",
  ref_year = "y2015",
  iniyear = 1995
)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{selectyears}{Years to be returned}

\item{time}{Time smoothing: average or spline (default)}

\item{dof}{Only specify if time=="spline": degrees of freedom needed for spline}

\item{averaging_range}{Only specify if time=="average": number of time steps to average}

\item{harmonize_baseline}{FALSE (default): no harmonization of input data to this function, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline of input data to this function (just specify when harmonize_baseline=TRUE)}

\item{iniyear}{Year of initialization for cropland area}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates committed agricultural water uses that are used in the river routing algorithm for distributing available water across the basin
}
\examples{
\dontrun{ calcOutput("WaterUseCommittedAg", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
