% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldgainArea.R
\name{calcYieldgainArea}
\alias{calcYieldgainArea}
\title{calcYieldgainArea}
\usage{
calcYieldgainArea(
  GT_range,
  lpjml,
  selectyears,
  climatetype,
  EFRmethod,
  yieldcalib,
  thresholdtype,
  avlland_scen,
  cropmix,
  multicropping
)
}
\arguments{
\item{GT_range}{Range of gainthreshold for calculation of potentially
irrigated areas}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{Years for which irrigatable area is calculated}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{thresholdtype}{TRUE: monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{avlland_scen}{Land availability scenario (currCropland, currIrrig, potIrrig)
combination of land availability scenario and initialization year separated by ":".
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
reports potentially irrigated area depending on gainthreshold
             and land constraint only
}
\examples{
\dontrun{
calcYieldgainArea(GT_range = seq(0, 10000, by = 100), scenario = "ssp2")
}

}
\author{
Felicitas Beier
}
