% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldgainArea.R
\name{calcYieldgainArea}
\alias{calcYieldgainArea}
\title{calcYieldgainArea}
\usage{
calcYieldgainArea(
  rangeGT,
  lpjml,
  selectyears,
  iniyear,
  climatetype,
  yieldcalib,
  unit,
  irrigationsystem,
  landScen,
  cropmix,
  multicropping
)
}
\arguments{
\item{rangeGT}{Range of gainthreshold for calculation of potentially
irrigated areas (in USD per hectare)}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{Years for which irrigatable area is calculated}

\item{iniyear}{Initialization year for cropland area}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
         Also needs specification of refYields, separated by ":".
         Options: FALSE (for single cropping analyses) or
                  "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{unit}{Unit of yield improvement potential used as threshold,
consisting of unit and price aggregation level separated by ":".
Unit:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return.
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
reports potentially irrigated area depending on gainthreshold
             and land constraint only
}
\examples{
\dontrun{
calcYieldgainArea(rangeGT = seq(0, 10000, by = 100), scenario = "ssp2")
}

}
\author{
Felicitas Beier
}
