% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldgainPotential.R
\name{calcYieldgainPotential}
\alias{calcYieldgainPotential}
\title{calcYieldgainPotential}
\usage{
calcYieldgainPotential(
  scenario,
  selectyears,
  iniyear,
  lpjml,
  climatetype,
  EFRmethod,
  yieldcalib,
  irrigationsystem,
  accessibilityrule,
  rankmethod,
  thresholdtype,
  gainthreshold,
  allocationrule,
  avlland_scen,
  cropmix,
  multicropping,
  unlimited
)
}
\arguments{
\item{scenario}{Non-agricultural water use and EFP scenario, separated
by "." (e.g. "on.ssp2")}

\item{selectyears}{Years for which yield gain potential is calculated}

\item{iniyear}{Initialization year}

\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}

\item{yieldcalib}{FAO (LPJmL yields calibrated with current FAO yield) or
calibrated (LPJmL yield potentials harmonized to baseline and calibrated for proxycrops) or
smoothed (smoothed LPJmL yield potentials, not harmonized, not calibrated) or
smoothed_calib (smoothed LPJmL yield potentials, not harmonized, calibrated for proxycrops)}

\item{irrigationsystem}{Irrigation system used
("surface", "sprinkler", "drip", "initialization")}

\item{accessibilityrule}{Strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year (Qx).
(e.g. Q75: 0.25, Q50: 0.5)}

\item{rankmethod}{Method of calculating the rank:
"meancellrank": mean over cellrank of proxy crops,
"meancroprank": rank over mean of proxy crops (normalized),
"meanpricedcroprank": rank over mean of proxy crops (normalized using price),
"watervalue": rank over value of irrigation water;
and fullpotentail TRUE/FALSE separated by ":"
(TRUE: Full irrigation potential (cell receives full irrigation requirements in total area).
FALSE: reduced potential of cell receives at later stage in allocation algorithm)}

\item{thresholdtype}{Thresholdtype of yield improvement:
TRUE: monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{gainthreshold}{Threshold of yield improvement potential required
(same unit as thresholdtype)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst", "equality")}

\item{avlland_scen}{Land availability scenario (currCropland, currIrrig, potIrrig)
combination of land availability scenario and initialization year separated by ":".
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{unlimited}{TRUE: no water limitation to potentially irrigated area
FALSE: irrigatable area limited by water availability}
}
\value{
magpie object in cellular resolution
}
\description{
reports yield gain potential for irrigatable area under different
             scenarios
}
\examples{
\dontrun{
calcOutput("YieldgainPotential", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
