% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldgainWatUse.R
\name{calcYieldgainWatUse}
\alias{calcYieldgainWatUse}
\title{calcYieldgainWatUse}
\usage{
calcYieldgainWatUse(
  lpjml,
  climatetype,
  selectyears,
  iniyear,
  landScen,
  cropmix,
  yieldcalib,
  multicropping,
  unit,
  irrigationsystem,
  rangeGT
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{Years to be returned}

\item{iniyear}{Year of initialization for cropland area}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_IFL, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
         Also needs specification of refYields, separated by ":".
         Options: FALSE (for single cropping analyses) or
                  "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from LandInG
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{unit}{Unit of yield improvement potential to be returned and
level of price aggregation used, separated by ":".
Unit:
USD_ha (USD per hectare) for relative area return, or
USD_m3 (USD per cubic meter) for relative volumetric return;
USD for absolute return (total profit);
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices}

\item{irrigationsystem}{Irrigation system used: system share as in initialization year,
or drip, surface, sprinkler for full irrigation by selected system}

\item{rangeGT}{Range of gainthreshold for calculation of potentially
irrigated areas (in USD per hectare)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water use on selected land areas
             without considering water constraints
}
\examples{
\dontrun{
calcOutput("YieldgainWatUse", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
