% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYields.R
\name{calcYields}
\alias{calcYields}
\title{calcYields}
\usage{
calcYields(
  version = "LPJmL5",
  climatetype = "CRU_4",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = FALSE,
  ref_year = "y2015",
  calib_proxy = TRUE,
  split_cropcalc = TRUE,
  crops = "magpie",
  selectyears = "all",
  replace_isimip3b = FALSE,
  replace_years = "lpj",
  isimip_subtype = "ISIMIP3b:yields.EPIC-IIASA_ukesm1-0-ll_ssp585_default"
)
}
\arguments{
\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{average, spline or raw (default)}

\item{averaging_range}{just specify for time=="average": number of time steps to average}

\item{dof}{just specify for time=="spline": degrees of freedom}

\item{harmonize_baseline}{FALSE (default) nothing happens, if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}

\item{calib_proxy}{calibrated proxy to FAO values if set to TRUE}

\item{split_cropcalc}{split calculation for different crop types (e.g. for parallelization)}

\item{crops}{magpie (default) crops or lpjml crops}

\item{selectyears}{defaults to all years available}

\item{replace_isimip3b}{replace yields of maize rice wheat soy with outputs from isimip3b crop modesl}

\item{replace_years}{only to subset years for isimip replacement, to reduce memory requirement
@return magpie object in cellular resolution}

\item{isimip_subtype}{crop model, gcm, ssp and co2 fert of isimip outputs"ISIMIP3b:yields.EPIC-IIASA_ukesm1-0-ll_ssp585_default"}
}
\description{
This function extracts yields from LPJmL to MAgPIE
}
\examples{
\dontrun{ calcOutput("Yields", aggregate = FALSE) }

}
\author{
Kristine Karstens, Felicitas Beier
}
