% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYields.R
\name{calcYields}
\alias{calcYields}
\title{calcYields}
\usage{
calcYields(
  version = "LPJmL5",
  climatetype = "CRU_4",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = FALSE,
  ref_year = "y2015",
  calib_proxy = TRUE,
  split_cropcalc = TRUE,
  selectyears = "all"
)
}
\arguments{
\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{average, spline or raw (default)}

\item{averaging_range}{just specify for time=="average": number of time steps to average}

\item{dof}{just specify for time=="spline": degrees of freedom}

\item{harmonize_baseline}{FALSE (default) nothing happens, if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}

\item{calib_proxy}{calibrated proxy to FAO values if set to TRUE}

\item{split_cropcalc}{split calculation for different crop types (e.g. for parallelization)}

\item{selectyears}{defaults to all years available}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts yields from LPJmL to MAgPIE
}
\examples{
\dontrun{ calcOutput("Yields", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
