% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsAdjusted.R
\name{calcYieldsAdjusted}
\alias{calcYieldsAdjusted}
\title{calcYieldsAdjusted}
\usage{
calcYieldsAdjusted(lpjml, climatetype, iniyear, selectyears, FAOyieldcalib)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical" for yields}

\item{iniyear}{year to be used for cropland of yield calibration}

\item{selectyears}{years to be returned by the function}

\item{FAOyieldcalib}{TRUE (LPJmL yields calibrated with current FAO yield) or FALSE (LPJmL yield potentials)}
}
\value{
magpie object in cellular resolution
}
\description{
This function returns irrigated and rainfed yields for magpie crops. It can either return LPJmL potential yields or LPJmL yields calibrated to FAO country yields
####### NOT FULLY IMPLEMENTED YET... Will potentially replace calcYields in calcIrrigYieldImprovementPotential ##############
}
\examples{
\dontrun{ calcOutput("IrrigYieldImprovementPotential", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
