% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsValued.R
\name{calcYieldsValued}
\alias{calcYieldsValued}
\title{calcYieldsValued}
\usage{
calcYieldsValued(
  lpjml,
  climatetype,
  unit,
  iniyear,
  selectyears,
  cropmix,
  yieldcalib,
  multicropping
)
}
\arguments{
\item{lpjml}{LPJmL version used for yields}

\item{climatetype}{Climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{unit}{Unit of yield improvement potential to be returned:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{iniyear}{initialization year for food price and cropmix area}

\item{selectyears}{Years to be returned by the function}

\item{cropmix}{Selected cropmix for which yield improvement potential
is calculated (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)
NULL returns all crops individually}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates yields per crop
             valued at FAO prices (optionally per area unit (ha)
             or water volume unit (m3))
}
\examples{
\dontrun{
calcOutput("YieldsValued", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
