% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctSoilClassification.R
\name{correctSoilClassification}
\alias{correctSoilClassification}
\title{correctSoilClassification}
\usage{
correctSoilClassification(x)
}
\arguments{
\item{x}{Magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct soil classification
}
\examples{

\dontrun{
  readSource("SoilClassification", subtype="HWSD.soil", convert="onlycorrect")
}

}
\seealso{
\code{\link{readSoilClassification}},
}
\author{
Marcos Alves, Kristine Karstens
}
