% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTransportDistance.R
\name{correctTransportDistance}
\alias{correctTransportDistance}
\title{correctTransportDistance}
\usage{
correctTransportDistance(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read transport distance file (no source information available)
}
\examples{

\dontrun{
  readSource("TransportDistance", convert="onlycorrect")
}

}
\seealso{
\code{\link{readDams}}
}
\author{
Kristine Karstens
}
