% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullCellularMAGPIE.R
\name{fullCellularMAGPIE}
\alias{fullCellularMAGPIE}
\title{fullCellularMAGPIE}
\usage{
fullCellularMAGPIE(rev = 0.1)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).
\code{\link{setConfig}} (e.g. for setting the mainfolder if not already set
properly).}
}
\description{
Function that produces the complete cellular data set required for running the
MAgPIE model.
}
\examples{

\dontrun{
fullMAgPIE(revision=12, mainfolder="pathtowhereallfilesarestored")
}
}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}},\code{\link{setConfig}}
}
\author{
Kristine Karstens
}
