% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullCELLULARMAGPIE.R
\name{fullCELLULARMAGPIE}
\alias{fullCELLULARMAGPIE}
\title{fullCELLULARMAGPIE}
\usage{
fullCELLULARMAGPIE(rev = 0.1, ctype = "c200")
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).}

\item{ctype}{aggregation clustering type, which is a combination of a single letter, indicating the cluster methodology, and a number,
indicating the number of resulting clusters. Available methodologies are hierarchical clustering (h), normalized k-means clustering
(n) and combined hierarchical/normalized k-means clustering (c). In the latter hierarchical clustering is used to determine the
cluster distribution among regions whereas normalized k-means is used for the clustering within a region.
\code{\link{setConfig}} (e.g. for setting the mainfolder if not already set
properly).}
}
\description{
Function that produces the complete cellular data set required for running the
MAgPIE model.
}
\examples{

\dontrun{
retrieveData("CELLULARMAGPIE", revision=12, mainfolder="pathtowhereallfilesarestored")
}
}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}},\code{\link{setConfig}}
}
\author{
Kristine Karstens, Jan Philipp Dietrich
}
