% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullIRRIGATIONPOTENTIAL.R
\name{fullIRRIGATIONPOTENTIAL}
\alias{fullIRRIGATIONPOTENTIAL}
\title{fullIRRIGATIONPOTENTIAL}
\usage{
fullIRRIGATIONPOTENTIAL(
  efrMethod = "VMF:fair",
  accessibilityrule = "CV:2",
  transDist = 0,
  fossilGW = FALSE,
  allocationrule = "optimization",
  rankmethod = "USD_ha:GLO:TRUE",
  gainthreshold = 500,
  protectLand = "HalfEarth",
  yieldcalib = "TRUE:FALSE",
  multicropping = FALSE,
  cropmix = "hist_total",
  climatetype = "MRI-ESM2-0:ssp370",
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb")
)
}
\arguments{
\item{efrMethod}{EFR method used including selected strictness of EFRs
(Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Method used: Quantile method (Q) or
Coefficient of Variation (CV)
combined with scalar value defining the strictness
of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year
(Qx, e.g. Q75: 0.25, Q50: 0.5)
or base value for exponential curve separated by : (CV:2)}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}

\item{fossilGW}{If TRUE: non-renewable groundwater can be used.
If FALSE: non-renewable groundwater cannot be used.}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst")}

\item{rankmethod}{Rank and optimization method consisting of
Unit according to which rank is calculated:
USD_ha (USD per hectare) for relative area return, or
USD_m3 (USD per cubic meter) for relative volumetric return;
USD for absolute return (total profit);
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices
and boolean indicating fullpotential (TRUE, i.e. cell
receives full irrigation requirements in total area)
or reduced potential (FALSE, reduced potential of cell
receives at later stage in allocation algorithm);
separated by ":"}

\item{gainthreshold}{Threshold of yield improvement potential required for
water allocation in upstreamfirst algorithm
(in same unit as in rankmethod)}

\item{protectLand}{Land protection scenario (WDPA, 
or one of the scenarios available in calcConservationPriorities,
e.g., 30by20, BH, BH_IFL, PBL_HalfEarth,
or NA for no protection).}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
     Also needs specification of refYields, separated by ":".
     Options: FALSE (for single cropping analyses) or
              "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from LandInG
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}
}
\description{
Function that produces the objects for Technical and Economic
             Irrigation Potentials within land and water boundaries
}
\author{
Felicitas Beier
}
