% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMULTICROPPING.R
\name{fullMULTICROPPING}
\alias{fullMULTICROPPING}
\title{fullMULTICROPPING}
\usage{
fullMULTICROPPING(suitability = "endogenous")
}
\arguments{
\item{suitability}{"endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set}
}
\description{
Function that produces output for irrigation potentials
             under multipl cropping on cellular resolution.
}
\author{
Felicitas Beier
}
