% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMULTICROPPING.R
\name{fullMULTICROPPING}
\alias{fullMULTICROPPING}
\title{fullMULTICROPPING}
\usage{
fullMULTICROPPING(
  multicropping = "TRUE:potential:endogenous",
  cropmix = c("maiz", "rapeseed", "puls_pro"),
  yieldcalib = TRUE
)
}
\arguments{
\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:cropIrrig" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}
}
\description{
Function that produces output for irrigation potentials
             under multipl cropping on cellular resolution.
}
\author{
Felicitas Beier
}
