% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMULTICROPPING.R
\name{fullMULTICROPPING}
\alias{fullMULTICROPPING}
\title{fullMULTICROPPING}
\usage{
fullMULTICROPPING(
  allocationrule = "optimization",
  comAg = TRUE,
  fossilGW = TRUE,
  transDist = 100
)
}
\arguments{
\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst")}

\item{comAg}{if TRUE: the currently already irrigated areas
         in initialization year are reserved for irrigation,
if FALSE: no irrigation areas reserved (irrigation potential)}

\item{fossilGW}{If TRUE: non-renewable groundwater can be used.
If FALSE: non-renewable groundwater cannot be used.}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}
}
\description{
Function that produces output for multiple cropping and
             irrigation potentials on current cropland
             at cellular resolution.
}
\author{
Felicitas Beier
}
