% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMULTICROPPING.R
\name{fullMULTICROPPING}
\alias{fullMULTICROPPING}
\title{fullMULTICROPPING}
\usage{
fullMULTICROPPING(
  cropmix = c("maiz", "rapeseed", "puls_pro"),
  yieldcalib = "TRUE:TRUE:actual:irrig_crop",
  allocationrule = "optimization",
  rankmethod = "USD_ha:GLO:TRUE",
  thresholdtype = "USD_ha:GLO",
  transDist = 100
)
}
\arguments{
\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
        Also needs specification of refYields, separated by ":".
        Options: FALSE (for single cropping analyses) or
                 "TRUE:actual:irrig_crop" (for multiple cropping analyses)
If FALSE: uncalibrated LPJmL yields are used}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst")}

\item{rankmethod}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return;
Unit of yield improvement potential to be returned;
Level of price aggregation used:
"GLO" for global average prices, or
"ISO" for country-level prices;
and boolean indicating fullpotential (TRUE, i.e. cell
receives full irrigation requirements in total area)
or reduced potential (FALSE, reduced potential of cell
receives at later stage in allocation algorithm);
separated by ":"}

\item{thresholdtype}{Unit of yield improvement potential used as threshold,
consisting of two components:
Unit:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return.
Price aggregation:
"GLO" for global average prices, or
"ISO" for country-level prices}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}
}
\description{
Function that produces output for irrigation potentials
             under multiple cropping on cellular resolution.
}
\author{
Felicitas Beier
}
