% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMULTICROPPING.R
\name{fullMULTICROPPING}
\alias{fullMULTICROPPING}
\title{fullMULTICROPPING}
\usage{
fullMULTICROPPING()
}
\description{
Function that produces output for multiple cropping and
             irrigation potentials on current cropland
             at cellular resolution.
}
\author{
Felicitas Beier
}
