% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullWATER4MAGPIE.R
\name{fullWATER4MAGPIE}
\alias{fullWATER4MAGPIE}
\title{fullWATER4MAGPIE}
\usage{
fullWATER4MAGPIE(lpjml = "lpjml5.9.16-m2", climatetype = "MRI-ESM2-0:ssp370")
}
\arguments{
\item{lpjml}{lpjml version}

\item{climatetype}{climate model and rcp}
}
\description{
Function that produces the complete cellular water data set
             for the MAgPIE model.
             Note: This is only needed temporarily until it is included
             in the fullCELLULARMAGPIE
}
\author{
Felicitas Beier
}
