% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrwater-package.R
\docType{package}
\name{mrwater-package}
\alias{mrwater}
\alias{mrwater-package}
\title{mrwater: madrat based MAgPIE water Input Data Library}
\description{
Provides functions for MAgPIE cellular input data generation and stand-alone water calculations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrwater,}
}

}
\author{
\strong{Maintainer}: Felicitas Beier \email{beier@pik-potsdam.de}

Authors:
\itemize{
  \item Jens Heinke \email{heinke@pik-potsdam.de}
  \item Kristine Karstens \email{karstens@pik-potsdam.de}
  \item Benjamin Leon Bodirsky \email{bodirsky@pik-potsdam.de}
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}
}

}
