% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistShrCurrIrrigFulfilled.R
\name{plotHistShrCurrIrrigFulfilled}
\alias{plotHistShrCurrIrrigFulfilled}
\title{plotHistShrCurrIrrigFulfilled}
\usage{
plotHistShrCurrIrrigFulfilled(
  scenario,
  iniyear,
  lpjml,
  selectyears,
  climatetype,
  EFRmethod
)
}
\arguments{
\item{scenario}{EFP and non-agricultural water use scenario separated by "."
(e.g. "on.ssp2")}

\item{iniyear}{Initialization year for committed agricultural uses}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{Year of plot (Note: one single year must be selected)}

\item{climatetype}{Switch between different climate models or historical baseline "GSWP3-W5E5:historical"}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:fair)}
}
\value{
map of magpie cells
}
\description{
plots histogram of share of current irrigation that can be
             fulfilled given water availability of the algorithm
             weighted with irrigated area
}
\examples{
\dontrun{ plotHistShrCurrIrrigFulfilled() }

}
\author{
Felicitas Beier
}
