% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapCellranking.R
\name{plotMapCellranking}
\alias{plotMapCellranking}
\title{plotMapCellranking}
\usage{
plotMapCellranking(
  lpjml,
  climatetype,
  selectyears,
  rankmethod,
  proxycrop,
  iniyear,
  yieldcalib
)
}
\arguments{
\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{year to be displayed}

\item{rankmethod}{method of calculating the rank: "meancellrank" (default): mean over cellrank of proxy crops, "meancroprank": rank over mean of proxy crops (normalized), "meanpricedcroprank": rank over mean of proxy crops (normalized using price), "watervalue": rank over value of irrigation water; and fullpotentail TRUE/FALSE separated by ":" (TRUE: Full irrigation potential (cell receives full irrigation requirements in total area). FALSE: reduced potential of cell receives at later stage in allocation algorithm)}

\item{proxycrop}{proxycrop(s) selected for crop mix specific calculations: average over proxycrop(s) yield gain. NULL returns all crops individually}

\item{iniyear}{Initialization year of irrigation system}

\item{yieldcalib}{FAO (LPJmL yields calibrated with current FAO yield) or calibrated (LPJmL yield potentials harmonized to baseline and calibrated for proxycrops) or none (smoothed LPJmL yield potentials, not harmonized, not calibrated)}
}
\value{
magpie object in cellular resolution
}
\description{
map of cellranking
}
\examples{
\dontrun{ plotMapCellranking() }

}
\author{
Felicitas Beier
}
