% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapIrrigatableArea.R
\name{plotMapIrrigatableArea}
\alias{plotMapIrrigatableArea}
\title{plotMapIrrigatableArea}
\usage{
plotMapIrrigatableArea(
  scenario,
  lpjml,
  climatetype,
  selectyears,
  rankmethod,
  cropmix,
  yieldcalib,
  EFRmethod,
  accessibilityrule,
  allocationrule,
  thresholdtype,
  gainthreshold,
  irrigationsystem,
  avlland_scen,
  potential_wat,
  com_ag,
  multicropping
)
}
\arguments{
\item{scenario}{Non-agricultural water use scenario and EFP "on" or "off" separated by .}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{selectyears}{years for which irrigatable area is calculated}

\item{rankmethod}{Method of calculating the rank: "meancellrank" (default): mean over cellrank of proxy crops, "meancroprank": rank over mean of proxy crops (normalized), "meanpricedcroprank": rank over mean of proxy crops (normalized using price), "watervalue": rank over value of irrigation water; and fullpotentail TRUE/FALSE separated by ":" (TRUE: Full irrigation potential (cell receives full irrigation requirements in total area). FALSE: reduced potential of cell receives at later stage in allocation algorithm)}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{EFRmethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Scalar value defining the strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year (Qx, e.g. Q75: 0.25, Q50: 0.5)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation across cells of river basin ("optimization" (default), "upstreamfirst", "equality")}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm: TRUE (default): monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{gainthreshold}{Threshold of yield improvement potential required for water allocation in upstreamfirst algorithm (in tons per ha)}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{avlland_scen}{Land availability scenario (currCropland, currIrrig, potIrrig)
combination of land availability scenario and initialization year separated by ":".
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{potential_wat}{if TRUE: potential available water and areas used, if FALSE: currently reserved water on current irrigated cropland used}

\item{com_ag}{if TRUE: the currently already irrigated areas in initialization year are reserved for irrigation, if FALSE: no irrigation areas reserved (irrigation potential)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
map of areas that can be potentially irrigated
}
\examples{
\dontrun{
plotMapIrrigatableArea()
}

}
\author{
Felicitas Beier
}
