% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapIrrigYieldImprovementPotential.R
\name{plotMapIrrigYieldImprovementPotential}
\alias{plotMapIrrigYieldImprovementPotential}
\title{plotMapIrrigYieldImprovementPotential}
\usage{
plotMapIrrigYieldImprovementPotential(
  selectyears,
  iniyear,
  lpjml,
  climatetype,
  yieldcalib,
  unit,
  multicropping,
  cropmix,
  landScen
)
}
\arguments{
\item{selectyears}{Years for which yield gain potential is calculated}

\item{iniyear}{Initialization year}

\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{unit}{Unit of yield improvement potential to be returned:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{landScen}{Land availability scenario (currCropland, currIrrig, potCropland)
combination of land availability scenario and initialization year separated by ":".
Initialization year only relevant for curr scenarios.
protection scenario separated by "_" (only relevant when potCropland selected):
WDPA, BH, FF, CPD, LW, HalfEarth}
}
\value{
map of magpie cells
}
\description{
plots map of irrigation yield improvement potential in USD/ha
}
\examples{
\dontrun{
plotMapYieldgainPotential()
}

}
\author{
Felicitas Beier
}
