% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapIrrigYieldImprovementPotential.R
\name{plotMapIrrigYieldImprovementPotential}
\alias{plotMapIrrigYieldImprovementPotential}
\title{plotMapIrrigYieldImprovementPotential}
\usage{
plotMapIrrigYieldImprovementPotential(
  selectyears,
  iniyear,
  lpjml,
  climatetype,
  yieldcalib,
  unit,
  multicropping,
  cropmix,
  avlland_scen
)
}
\arguments{
\item{selectyears}{Years for which yield gain potential is calculated}

\item{iniyear}{Initialization year}

\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{yieldcalib}{Calibrated (LPJmL yield potentials smoothed and harmonized
to baseline and calibrated with global FAO calibration factor
for proxycrops where LPJmL crops mapped multiple times to MAgPIE crops) or
FAO (LPJmL yields calibrated with current FAO yield)}

\item{unit}{Unit of yield improvement potential to be returned:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{avlland_scen}{Land availability scenario (currCropland, currIrrig, potIrrig)
combination of land availability scenario and initialization year separated by ":".
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth}
}
\value{
map of magpie cells
}
\description{
plots map of irrigation yield improvement potential in USD/ha
}
\examples{
\dontrun{
plotMapYieldgainPotential()
}

}
\author{
Felicitas Beier
}
