% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapPotReturnToIrrigation.R
\name{plotMapPotReturnToIrrigation}
\alias{plotMapPotReturnToIrrigation}
\title{plotMapPotReturnToIrrigation}
\usage{
plotMapPotReturnToIrrigation(
  selectyears,
  unit,
  iniyear,
  lpjml,
  climatetype,
  yieldcalib,
  multicropping,
  cropmix,
  landScen
)
}
\arguments{
\item{selectyears}{year for which map is displayed}

\item{unit}{unit of return to irrigation
(perha for monetary yield gain per cell in USD05/ha
or perm3 for irrigation water value in USD05/m^3)}

\item{iniyear}{Initialization year for return to irrigation}

\item{lpjml}{LPJmL version required for return to irrigation}

\item{climatetype}{Climate scenario for return to irrigation}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{multicropping}{multicropping for return to irrigation (TRUE or FALSE)}

\item{cropmix}{cropmix for calculation of return to irrigation
(proxycrop(s) or hist_irrig or hist_total)}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. landScen (currCropland, currIrrig, potCropland)
2. for curr-scenarios: initialization year;
for pot-scenarios: protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_FF, NA).
For case of pot-scenario without land protection select "NA"
or do not specify second part of the argument}
}
\value{
map of magpie cells
}
\description{
plots map of return to irrigation in terms of monetary yield gain in USD/ha
             or water value in USD/m^3
}
\examples{
\dontrun{
plotMapPotReturnToIrrigation()
}

}
\author{
Felicitas Beier
}
