% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapPotReturnToIrrigation.R
\name{plotMapPotReturnToIrrigation}
\alias{plotMapPotReturnToIrrigation}
\title{plotMapPotReturnToIrrigation}
\usage{
plotMapPotReturnToIrrigation(
  selectyears,
  unit,
  iniyear,
  lpjml,
  climatetype,
  yieldcalib,
  multicropping,
  cropmix,
  land
)
}
\arguments{
\item{selectyears}{year for which map is displayed}

\item{unit}{unit of return to irrigation
(perha for monetary yield gain per cell in USD05/ha
or perm3 for irrigation water value in USD05/m^3)}

\item{iniyear}{Initialization year for return to irrigation}

\item{lpjml}{LPJmL version required for return to irrigation}

\item{climatetype}{Climate scenario for return to irrigation}

\item{yieldcalib}{yieldcalib for return to irrigation (FAO, calibrated,
smoothed, smoothed_calib)}

\item{multicropping}{multicropping for return to irrigation (TRUE or FALSE)}

\item{cropmix}{cropmix for calculation of return to irrigation
(proxycrop(s) or hist_irrig or hist_total)}

\item{land}{land area to be displayed
combination of land availability scenario and initialization year separated by ":"
land availability scenario: currIrrig, currCropland, potIrrig
protection scenario separated by "_" (only relevant when potIrrig selected):
WDPA, BH, FF, CPD, LW, HalfEarth
(e.g. "potIrrig_HalfEarth:1995", currCropland:2010)}
}
\value{
map of magpie cells
}
\description{
plots map of return to irrigation in terms of monetary yield gain in USD/ha
             or water value in USD/m^3
}
\examples{
\dontrun{ plotMapPotReturnToIrrigation() }

}
\author{
Felicitas Beier
}
