% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapYieldgainPotential.R
\name{plotMapYieldgainPotential}
\alias{plotMapYieldgainPotential}
\title{plotMapYieldgainPotential}
\usage{
plotMapYieldgainPotential(
  output,
  scenario,
  selectyears,
  iniyear,
  lpjml,
  climatetype,
  efrMethod,
  accessibilityrule,
  yieldcalib,
  rankmethod,
  thresholdtype,
  gainthreshold,
  allocationrule,
  irrigationsystem,
  multicropping,
  cropmix,
  landScen
)
}
\arguments{
\item{output}{Output to be plotted: "unlimited", "limited" or "ratio}

\item{scenario}{Non-agricultural water use and EFP scenario, separated
by "." (e.g. "on.ssp2")}

\item{selectyears}{Years for which yield gain potential is calculated}

\item{iniyear}{Initialization year}

\item{lpjml}{LPJmL version used}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year (Qx).
(e.g. Q75: 0.25, Q50: 0.5)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{rankmethod}{Method of calculating the rank:
"meancellrank": mean over cellrank of proxy crops,
"meancroprank": rank over mean of proxy crops (normalized),
"meanpricedcroprank": rank over mean of proxy crops (normalized using price),
"watervalue": rank over value of irrigation water;
and fullpotentail TRUE/FALSE separated by ":"
(TRUE: Full irrigation potential (cell receives full irrigation requirements in total area).
FALSE: reduced potential of cell receives at later stage in allocation algorithm)}

\item{thresholdtype}{Thresholdtype of yield improvement:
TRUE: monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{gainthreshold}{Threshold of yield improvement potential required
(same unit as thresholdtype)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst", "equality")}

\item{irrigationsystem}{Irrigation system used
("surface", "sprinkler", "drip", "initialization")}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}

\item{cropmix}{Selected cropmix (options:
"hist_irrig" for historical cropmix on currently irrigated area,
"hist_total" for historical cropmix on total cropland,
or selection of proxycrops)}

\item{landScen}{Land availability scenario consisting of two parts separated by ":":
1. available land scenario (currCropland, currIrrig, potCropland)
2. protection scenario (WDPA, BH, FF, CPD, LW, HalfEarth, BH_FF, NA).
For case of no land protection select "NA"
or do not specify second part of the argument}
}
\value{
map of magpie cells
}
\description{
plots map of irrigation yield gain potential in USD/ha
             for water limited or unlimited irrigatable areas or their
             ratio
}
\examples{
\dontrun{
plotMapYieldgainPotential()
}

}
\author{
Felicitas Beier
}
