% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivityAllocationrule.R
\name{plotSensitivityAllocationrule}
\alias{plotSensitivityAllocationrule}
\title{plotSensitivityAllocationrule}
\usage{
plotSensitivityAllocationrule(
  x_axis_range,
  region = "GLO",
  output,
  scenario,
  lpjml,
  selectyears,
  climatetype,
  efrMethod,
  accessibilityrule,
  yieldcalib,
  thresholdtype,
  irrigationsystem,
  landScen,
  cropmix,
  potential_wat = TRUE,
  com_ag,
  multicropping
)
}
\arguments{
\item{x_axis_range}{range of x-axis (gainthreshold) to be depicted on the curve}

\item{region}{regional resolution (can be country iso-code, "GLO" for global,
region name and respective mapping "EUR:H12")}

\item{output}{output to be displayed: irrigated area "IrrigArea" or
available water volume "wat_ag_ww" "wat_ag_wc"}

\item{scenario}{non-agricultural water use scenario to be displayed in plot}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{selectyears}{years for which irrigatable area is calculated}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{efrMethod}{EFR method used including selected strictness of EFRs (e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Scalar value defining the strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year (Qx).
Default: 0.5 (Q50) (e.g. Q75: 0.25, Q50: 0.5)}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{thresholdtype}{Thresholdtype of yield improvement potential required for water allocation in upstreamfirst algorithm:
TRUE: monetary yield gain (USD05/ha), FALSE: yield gain in tDM/ha}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{landScen}{Land availability scenario (currCropland, currIrrig, potCropland)
combination of land availability scenario and initialization year separated by ":".
Initialization year only relevant for curr scenarios.
protection scenario separated by "_" (only relevant when potCropland selected):
WDPA, BH, FF, CPD, LW, HalfEarth}

\item{cropmix}{Cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{potential_wat}{if TRUE: potential available water and areas used, if FALSE: currently reserved water on current irrigated cropland used}

\item{com_ag}{if TRUE: the currently already irrigated areas in initialization year are reserved for irrigation,
FALSE: no irrigation areas reserved (irrigation potential)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
plot of irrigatable area depending on costs paid for irrigation for different water accessibility
}
\examples{
\dontrun{
plotSensitivityAllocationrule(x_axis_range = seq(0, 10000, by = 100), scenario = "ssp2")
}

}
\author{
Felicitas Beier
}
