% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCO2Atmosphere.R
\name{readCO2Atmosphere}
\alias{readCO2Atmosphere}
\title{readCO2Atmosphere}
\usage{
readCO2Atmosphere(subtype = "CO2Atmosphere:Elevated.CO2")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read CO2 global atmospheric concentration
}
\examples{

\dontrun{
readSource("CO2Atmosphere", subtype="Elevated.CO2", convert="onlycorrect")
}

}
\seealso{
\code{\link{readCO2Atmosphere}}
}
\author{
Marcos Alves, Kristine Karstens
}
