% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGrassYieldsEmulator.R
\name{readGrassYldEmu}
\alias{readGrassYldEmu}
\title{readGrassYldEmu}
\usage{
readGrassYldEmu(subtype = "GrassYldEmu:20f33a2280.max_harvest")
}
\arguments{
\item{subtype}{subtype of file to be opened}
}
\value{
Magpie object on a celullar level with maximum gass harvest in gC/m²
}
\description{
Read the results of the grass harvest maximization performed for the grass havest emulators
}
\examples{

\dontrun{
readSource("GrassYldEmu", subtype = "GrassYldEmu:20f33a2280.max_harvest", convert="onlycorrect")
}

}
\author{
Marcos Alves
}
