% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIrrigationSystem.R
\name{readIrrigationSystem}
\alias{readIrrigationSystem}
\title{readIrrigationSystem}
\usage{
readIrrigationSystem(subtype = "Jaegermeyr_magpiecell")
}
\arguments{
\item{subtype}{Data source to be used: Jaegermeyr (irrigation system share based on FAO 2014, ICID 2012 and Rohwer et al. 2007) or LPJmL (dominant irrigation system per country) and number of cells (lpjcell or magpiecell) separated by _}
}
\value{
MAgPIE object of at country-level
}
\description{
Read in irrigation system type for initialization
}
\examples{
\dontrun{ readSource("IrrigationSystem", convert="onlycorrect") }

}
\author{
Felicitas Beier
}
