% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readISIMIP.R
\name{readISIMIP}
\alias{readISIMIP}
\title{readISIMIP}
\usage{
readISIMIP(subtype = "water_abstraction")
}
\arguments{
\item{subtype}{Data source to be read from}
}
\value{
MAgPIE object of non-agricultural water demand at 0.5 cellular level in mio. m^3
}
\description{
Read in non-agricultural water demand data from ISIMIP2b input data
}
\examples{
\dontrun{ readSource("ISIMIP", convert="onlycorrect") }

}
\author{
Felicitas Beier
}
