% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readISIMIPoutputs.R
\name{readISIMIPoutputs}
\alias{readISIMIPoutputs}
\title{readISIMIPoutputs}
\usage{
readISIMIPoutputs(
  subtype = "ISIMIP2b:water.histsoc_airrww_pcr-globwb_gfdl-esm2m"
)
}
\arguments{
\item{subtype}{Data source to be read from and subtype separated by ".";
For water: subtype consisting of variable ("airrww") watermodel ("cwatm","h08","lpjml","matsiro","mpi-hm","pcr-globwb") and GCM ("ipsl-cm5a-lr","gfdl-esm2m","miroc5","hadgem2-es") separated by "_"
For yields: subtype is "ISIMIP3b:yields.cropmodel_gcm_ssp_co2" cropmodels ("LPJmL", "EPIC-IIASA") gcms ("ukesm1-0-ll","gfdl-esm4","ipsl-cm6a-lr") ssp ("ssp126", "ssp585")
Yields subtype example: "ISIMIP3b:yields.EPIC-IIASA_ukesm1-0-ll_ssp585_default"}
}
\value{
MAgPIE object of non-agricultural water demand at 0.5 cellular level in mio. m^3
}
\description{
Reads in agricultural water withdrawals from ISIMIP2b output data, and GGCMI potential global crop yields from ISIMIP3b
}
\examples{
\dontrun{
 readSource("ISIMIPoutputs", convert=TRUE)
}

}
\author{
Felicitas Beier, David CHen
}
