% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLabourProdImpactOrlov.R
\name{readLabourProdImpactOrlov}
\alias{readLabourProdImpactOrlov}
\title{readLabourProdImpactOrlov}
\usage{
readLabourProdImpactOrlov(
  outdoor = TRUE,
  intensity = "high",
  hothaps = TRUE,
  gcm = "IPSL-CM5A-LR",
  rcp = "rcp85"
)
}
\arguments{
\item{outdoor}{outdoor or indoor (shady) work}

\item{intensity}{high or low - 400W is high intensity 300W moderate industrial intensity}

\item{hothaps}{method of translating heat into labour prod. hothaps (epidemiological) or iso (standards and working time based)}

\item{gcm}{option of 3 gcms IPSL-CM5A-LR, HadGEM2-ES, GFDL-ESM2M}

\item{rcp}{option of rcp26,rcp60, rcp85}
}
\value{
magpie object of gridded productivity as share of 1 (full productivity)
}
\description{
read in labour productivity impacts from climate change from Orlov (see Orlov et al. 2019. Economic Losses of Heat-Induced Reductions in Outdoor Worker Productivity: a Case Study of Europe. Economics of Disasters and Climate Change, 3(3), 191-211.)
}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
