% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSoilClassification.R
\name{readSoilClassification}
\alias{readSoilClassification}
\title{readSoilClassification}
\usage{
readSoilClassification(subtype = "SoilClassification:HWSD.soil")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
Magpie object with results on cellular level for soil characteristics.
}
\description{
Read Soil parameter data for LPJmL
}
\examples{

\dontrun{
readSource("SoilClassification", subtype="HWSD.soil", convert="onlycorrect")
}

<<<<<<< HEAD
=======
}
\seealso{
\code{\link{readSoilClassification}}
>>>>>>> 6971f48ef1772a1c023b7e2013df42d80df3edbe
}
\author{
Marcos Alves, Kristine Karstens
}
