% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSoilClassification.R
\name{readSoilClassification}
\alias{readSoilClassification}
\title{readSoilClassification}
\usage{
readSoilClassification(subtype = "SoilClassification:HWSD.soil")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read soil classification
}
\examples{

\dontrun{
readSource("SoilClassification", subtype="HWSD.soil", convert="onlycorrect")
}

}
\seealso{
\code{\link{readSoilClassification}}
}
\author{
Marcos Alves, Kristine Karstens
}
