% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCountryResults.R
\name{reportCountryResults}
\alias{reportCountryResults}
\title{reportCountryResults}
\usage{
reportCountryResults(
  output,
  lpjml,
  climatetype,
  gainthreshold,
  efrMethod,
  accessibilityrule,
  rankmethod,
  yieldcalib,
  allocationrule,
  thresholdtype,
  irrigationsystem,
  cropmix,
  multicropping
)
}
\arguments{
\item{output}{Output to be displayed: irrigated area "IrrigArea" or
available water volume "wat_ag_ww" "wat_ag_wc"}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{gainthreshold}{gainthreshold chosen for reporting}

\item{efrMethod}{EFR method used including selected strictness of EFRs
(e.g. Smakhtin:good, VMF:fair)}

\item{accessibilityrule}{Method used: Quantile method (Q) or Coefficient of Variation (CV)
combined with scalar value defining the strictness of accessibility restriction:
discharge that is exceeded x percent of the time on average throughout a year
(Qx, e.g. Q75: 0.25, Q50: 0.5)
or base value for exponential curve separated by : (CV:2)}

\item{rankmethod}{Rank and optimization method consisting of
Unit according to which rank is calculated:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return;
and boolean indicating fullpotential
(TRUE, i.e. cell receives full irrigation requirements in total area)
or reduced potential
(FALSE, reduced potential of cell receives at later stage in allocation algorithm);
separated by ":"}

\item{yieldcalib}{If TRUE: LPJmL yields calibrated to FAO country yield in iniyear
If FALSE: uncalibrated LPJmL yields are used}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst")}

\item{thresholdtype}{Unit of yield improvement potential used as threshold:
tDM (tons per dry matter),
USD_ha (USD per hectare) for area return, or
USD_m3 (USD per cubic meter) for volumetric return}

\item{irrigationsystem}{Irrigation system to be used for river basin discharge
allocation algorithm ("surface", "sprinkler", "drip", "initialization")}

\item{cropmix}{cropmix for which irrigation yield improvement is calculated
can be selection of proxycrop(s) for calculation of average yield gain
or hist_irrig or hist_total for historical cropmix}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE)}
}
\value{
data frame in country-resolution
}
\description{
reports country-values (irrigated area or irrigation water availability)
}
\examples{
\dontrun{
reportCountryResults()
}

}
\author{
Felicitas Beier
}
