% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDischargeAllocation.R
\name{toolDischargeAllocation}
\alias{toolDischargeAllocation}
\title{toolDischargeAllocation}
\usage{
toolDischargeAllocation(y, rs, inoutLIST, inLIST, allocationrule, glocellrank)
}
\arguments{
\item{y}{Current year of loop}

\item{rs}{River structure (list of river structure details and
cell numbers including ordered cell number)}

\item{inoutLIST}{List of inputs that are also and output of the function
(variables that are updated in the discharge allocation loop):
reservedWAT (minimum water requirement reserved per cell);
discharge (discharge to be allocated);
fullirrigFRAC (fraction of full irrigation requirements that can be fulfilled)}

\item{inLIST}{List of inputs to the discharge allocation algorithm
(variables that are not altered by the discharge allocation algorithm):
yieldgain (irrigation yield value gain);
fullIrrigReqWW (required withdrawal for full irrigation in specific cell);
fullIrrigReqWC (required consumption for full irrigation in specific cell);
gainthreshold (Minimum yield gain in USD/ha (as scalar value));
iwpWW; iwpWC (irrigation water potential (WW: withdrawal, WC: consumption),
i.e. water that is potentially available for irrigation water withdrawal/consumption)}

\item{allocationrule}{Rule to be applied for river basin discharge allocation
across cells of river basin ("optimization", "upstreamfirst")}

\item{glocellrank}{Cell ranking for different years
Note: only applicable when allocationrule "optimization" chosen}
}
\value{
magpie object in cellular resolution
}
\description{
This tool function executes the Allocation Algorithm in
             cell order of selected allocation rule
}
\author{
Felicitas Beier, Jens Heinke, Jan Philipp Dietrich
}
