% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolNeighborUpDownProvision.R
\name{toolNeighborUpDownProvision}
\alias{toolNeighborUpDownProvision}
\title{toolNeighborUpDownProvision}
\usage{
toolNeighborUpDownProvision(rs, transDist, listNeighborIN)
}
\arguments{
\item{rs}{River structure including information 
on neighboring cells}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}

\item{listNeighborIN}{List of arrays required for the algorithm:
yearlyRunoff, lakeEvap
reserved flows from previous water allocation round 
(prevReservedWC, prevReservedWW)
missing water at this stage of water allocation 
(missingWW, missingWC)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water provision by 
             surrounding grid cells for upstream-downstream
             allocation set-up
}
\examples{
\dontrun{
calcOutput("RiverHumanUses", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
