% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRiverUpDownBalance.R
\name{toolRiverUpDownBalance}
\alias{toolRiverUpDownBalance}
\title{toolRiverUpDownBalance}
\usage{
toolRiverUpDownBalance(c, rs, transDist, inLIST, inoutLIST)
}
\arguments{
\item{c}{Current cell for which water shall be allocated}

\item{rs}{River structure with information on upstreamcells,
downstreamcells and neighboring cells and distances}

\item{transDist}{Water transport distance allowed to fulfill locally
unfulfilled water demand by surrounding cell water availability}

\item{inLIST}{List of objects that are inputs to the function:
yearly runoff in current cell; lake evaporation in current cell;
previously reservered withdrawal in all cells;
previously reserved consumption in current cell}

\item{inoutLIST}{List of objects that are inputs to the function and
are updated by the function:
discharge, inflow,
currently requested flows (withdrawal and consumption)}
}
\value{
list of magpie objects in cellular resolution
}
\description{
This function calculates the cell water balance
             under consideration of different reserved human uses
             (non-agricultural, neighbor water requirements, 
              committed-agricultural uses)
}
\author{
Felicitas Beier, Jens Heinke, Jan Philipp Dietrich
}
