% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRiverUpDownBalanceSINGLE.R
\name{toolRiverUpDownBalanceSINGLE}
\alias{toolRiverUpDownBalanceSINGLE}
\title{toolRiverUpDownBalanceSINGLE}
\usage{
toolRiverUpDownBalanceSINGLE(inLIST, inoutLIST)
}
\arguments{
\item{inLIST}{List of objects that are inputs to the function:
yearly runoff in current cell; lake evaporation in current cell;
previously reservered withdrawal in all cells;
previously reserved consumption in current cell}

\item{inoutLIST}{List of objects that are inputs to the function and
are updated by the function:
discharge, inflow,
currently requested flows (withdrawal and consumption)}
}
\value{
list of arrays objects in cellular resolution
}
\description{
This function calculates the cell water balance
             under consideration of different reserved human uses
             (non-agricultural, neighbor water requirements,
              committed-agricultural uses)
}
\author{
Felicitas Beier, Jens Heinke
}
